% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aad.R
\docType{class}
\encoding{UTF-8}
\name{aad-class}
\alias{aad-class}
\title{An S4 Class to Average Absolute Deviation}
\description{
An S4 Class to Average Absolute Deviation
}
\section{Slots}{

\describe{
\item{\code{estimate}}{Estimated value.}
}}

