% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aad.R
\docType{methods}
\encoding{UTF-8}
\name{aad}
\alias{aad}
\alias{aad,ANY-method}
\alias{aad,numeric,logical,aad-method}
\title{Average (Mean) Absolute Deviation}
\usage{
aad(x, na.rm = TRUE)

\S4method{aad}{ANY}(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector containing the observations.}

\item{na.rm}{A logical value for \code{na.rm}, default is \code{na.rm=TRUE}.}
}
\description{
Calculates the average (mean) absolute deviation from the sample mean.
}
\details{
The statistical literature has not yet adopted a standard notation, as both the "Mean absolute deviation around the mean" and the "Median absolute deviation around the median" have been denoted as "MAD", which may lead to confusion once they may produce different values. Thus, the \code{aad} version calculates the "Mean Absolute Deviation"--or  "Average Deviation (AD)" ad suggested by Garrett, who defines it as "the mean of the deviation of all the separate scores in the series taken from their mean (occasionally from the median or mode)", (1971, p. 481).
The R \code{\link[stats]{mad}} by default computes the "Median Absolute Deviation" with an adjusting factor for asymptotically normal consistency. To obtain the "Mean Absolute Deviation" one should use \code{stats::mad(x, constant = 1)}.
}
\examples{
x <- c(15, 10, 6, 8, 11)
# With the 'aad' function:
aad(x)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Garrett, Henry (1982) \emph{Statistics in Psychology and Education}. 6th, Paragon.
}
\seealso{
\code{\link[stats]{mad}}
}
\keyword{Exploratory}

