% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{abortion}
\alias{abortion}
\title{Haberman's Abortion, Education, and Religion Data}
\format{A 3x3x3 (contingency) table}
\usage{
data(abortion)
}
\description{
A multi-way contingency table with the results of a survey concerning attitudes concerning abortion.
}
\author{
1972 National Opinion Research Center
}
\references{
Haberman, S. (1978). \emph{Analysis of Qualitative Data} 1, 2. Academic Press, Orlando, FL.
}
\keyword{datasets}

