% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.footnote}
\alias{add.footnote}
\title{Footnote to ggplot Graphs}
\usage{
add.footnote(text = paste(Sys.info()["user"], format(Sys.time(),
  "\%d \%b \%Y"), sep = " "), size = 0.7, color = grey(0.75))
}
\arguments{
\item{text}{Any text or empty to use default}

\item{size}{The size for text.}

\item{color}{Color for the text}
}
\description{
Add footnotes to \pkg{ggplot2} objects.
}
\details{
Only works with a ggplot object.
}
\examples{
# setup data
x <- seq(0, 50, 1)
supply <- x * -2 + 100
demand <- x * 2
df <- data.frame( x = x, supply=supply, demand=demand)

library(ggplot2)
ggplot(df, aes(x)) +
geom_line(aes(y = supply, colour = supply))+
geom_line(aes(y = demand, colour = demand))
add.footnote()
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Graphs}

