% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/VISUALIZATION.R
\encoding{UTF-8}
\name{addErrorBar}
\alias{addErrorBar}
\title{Add Error Bars on the Graph}
\usage{
addErrorBar(x, y, ebl, ebu = ebl, length = 0.08, ...)
}
\arguments{
\item{x}{coordinates for the error bars, or simply the name of the graph.}

\item{y}{coordinates for the center of the error bars, which is the group means.}

\item{ebl}{length of the error bars, which should be 1 se in each direction.}

\item{ebu}{the error bars.}

\item{length}{the length.}

\item{\dots}{typically additional noninteresting arguments to pass.}
}
\description{
Draw error bars on the graph.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

