% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VISUALIZATION.R
\name{addLegend}
\alias{addLegend}
\title{Add legends to a plot}
\usage{
addLegend(...)
}
\arguments{
\item{\dots}{Legend parameters.}
}
\description{
To add special styles of legends.
}
\examples{
par(mar = c(5, 4, 1.4, 0.2))
plot(rnorm(50), rnorm(50), col=c("steelblue", "indianred"), pch=20)
addLegend("topright", legend=c("Foo", "Bar"), pch=20,
col=c("steelblue", "indianred"),
horiz=TRUE, bty='n', cex=0.8)
}

