% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{adjResiduals}
\alias{adjResiduals}
\title{Adjusted Residuals}
\usage{
adjResiduals(object, ...)
}
\arguments{
\item{object}{a model object of type \code{glm} or \code{lm}.}

\item{\dots}{further arguments passed to or used by other methods.}
}
\description{
Adjusted Residuals
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

