% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Computes the age}
\usage{
age(from, to)
}
\arguments{
\item{from}{The first date (birthdate)}

\item{to}{The last date, usually the event date.}
}
\description{
The `age` function computes the age given a bithdate (from) and elapsed time (to) variables.
}
\examples{
birth = '1977-07-12'; today = '2015-07-12'
age(birth, today)
}

