% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{allPossiblePermutations}
\alias{allPossiblePermutations}
\title{perms}
\usage{
allPossiblePermutations(x)
}
\arguments{
\item{x}{a integer vector}
}
\value{
a matrix with n! rows and n columns
}
\description{
This function generates all the possible permutations of the set
{1, 2, ..., n} (for some given n). A function removes missing values.
}
\seealso{
Other Exploratory: \code{\link{findMode}}
}

