% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTS.R
\encoding{UTF-8}
\name{anderson.darling}
\alias{anderson.darling}
\title{Anderson-Darling test for normality}
\usage{
anderson.darling(x)
}
\arguments{
\item{x}{A numeric vector of data values, the number of observations must be greater than 7. Missing values are allowed.
@details The Anderson-Darling test is the recommended EDF test by Stephens (1986) followed by the Cramer-von Mises test. Compared to the later, the Anderson-Darling gives more weight to the tails of the distribution.}
}
\description{
Performs the Anderson-Darling test for the composite hypothesis of normality. See details.
}
\examples{
set.seed(1234)
x = rnorm(1000)
anderson.darling(x)

}
\keyword{Tests}

