% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{around}
\alias{around}
\title{Find the Values Around a Particular Value}
\usage{
around(x, value)
}
\arguments{
\item{x}{A vector.}

\item{value}{Specified value}
}
\value{
lo The maximum value of x that is less than or equal to the value parameter.

hi The minimum value of x that is greater than or equal to the value parameter.
}
\description{
Find the location of values around a specified value
}
\examples{
set.seed(123)
x = rnorm(25, 5, 10)
value = 9
around(x, value)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

