% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.percent.R
\encoding{UTF-8}
\name{as.percent}
\alias{as.percent}
\title{Converts to percentiles}
\usage{
as.percent(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Converts a numeric vector to percentiles.
}
\examples{
vec <- seq(1:5)
as.percent(vec)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

