% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as.png}
\alias{as.png}
\title{Save Plot to a PNG}
\usage{
as.png(file, width, height, FUN, png = FALSE)
}
\arguments{
\item{file}{file location}

\item{width}{Width for the plot}

\item{height}{Height for the plot}

\item{FUN}{Function to be plotted}

\item{png}{a logical argument}
}
\description{
Save plots as PNG to working directory
}

