% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.roman.R
\name{as.roman}
\alias{as.roman}
\title{Converts integer into its Roman Numeral equivalent}
\usage{
as.roman(x)
}
\arguments{
\item{x}{The integer to be translated.}
}
\description{
Converts integer into its Roman Numeral equivalent
}
\examples{
as.roman(1977)
}

