% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{as.tsdf}
\alias{as.tsdf}
\title{A data.frame with time series parameters}
\usage{
as.tsdf(timevar, format, x = NULL, tz = "GMT")
}
\arguments{
\item{timevar}{A time variable}

\item{format}{A time format}

\item{x}{Data variable to convert to a time series}

\item{tz}{Time zone code, default is \dQuote{GMT}}
}
\description{
Return a \code{data.frame} with time parameters and time variable
}
\details{
A data frame containing time variable parameters and measuring variable
}
\examples{
data(ssex)
myts <- as.tsdf(timevar = ssex[,1], format = "\%Y", x= ssex[,3])
peek(myts)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Ripley, B. D. and Hornik, K. (2001) Date-time classes. \emph{R News}, 1/2, 8–11 \url{http://www.r-project.org/doc/Rnews/Rnews_2001-2.pdf}
}
\seealso{
\link{by_day}, \link{by_month}, \link{by_year}
}
\keyword{Data-Manipulation}

