% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{AsNumeric}
\alias{AsNumeric}
\title{Clean up a character vector to make it numeric}
\usage{
AsNumeric(x)
}
\arguments{
\item{x}{character vector to process}
}
\value{
numeric vector
}
\description{
Remove commas and whitespace, primarily
}
\keyword{Clean-up}

