% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{AsShare}
\alias{AsShare}
\title{Clean up a character vector to make it a percent}
\usage{
AsShare(x)
}
\arguments{
\item{x}{character vector to process}
}
\value{
numeric vector
}
\description{
Remove "%" primarily, convert to numeric and divide by 100
}
\keyword{Clean-up}

