% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/electoralTool.R
\docType{methods}
\name{atkinson}
\alias{atkinson}
\alias{atkinson,ANY-method}
\alias{atkinson,numeric,numeric,numeric,logical,ANY-method}
\title{Atkinson Index of Inequality}
\usage{
atkinson(x, n = rep(1, length(x)), parameter = 0.5, na.rm = FALSE)

\S4method{atkinson}{ANY}(x, n = rep(1, length(x)), parameter = 0.5,
  na.rm = FALSE)
}
\arguments{
\item{x}{A vector of data values of non-negative elements.}

\item{n}{A vector of frequencies of the same length as \code{x}.}

\item{parameter}{A parameter of the inequality measure (if set to \code{NULL} the default parameter of the respective measure is used).}

\item{na.rm}{A logical. Should missing values be removed? The Default is set to \code{na.rm=FALSE}.}
}
\description{
Calculates the Atkinson Index. This inequality measure is espcially good at determining which end of the distribution is contributing most to the observed inequality.
}
\examples{
# generate a vector (of incomes)
x <- c(778, 815, 857, 888, 925, 930, 965, 990, 1012)

# compute Atkinson coefficient with parameter=0.5
atkinson(x, parameter=0.5)
}
\references{
Cowell, F. A. (2000) Measurement of Inequality in Atkinson, A. B. / Bourguignon, F. (Eds): \emph{Handbook of Income Distribution}. Amsterdam.

Cowell, F. A. (1995) \emph{Measuring Inequality} Harvester Wheatshef: Prentice Hall.
}
\seealso{
\code{\link{herfindahl}}, \code{\link{rosenbluth}},  \code{\link{gini}}.
}

