% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AverageAbsoluteDeviation.R
\encoding{UTF-8}
\name{AverageAbsoluteDeviation}
\alias{AverageAbsoluteDeviation}
\alias{AverageAbsoluteDeviation.default}
\title{Average Absolute Deviation}
\usage{
AverageAbsoluteDeviation(x, na.rm = TRUE, ...)

\method{AverageAbsoluteDeviation}{default}(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector containing the observations.}

\item{na.rm}{A logical value for \code{na.rm}, default is \code{na.rm=TRUE}.}

\item{\dots}{Additional arguements (currently ignored)}
}
\description{
Calculates the average (mean) absolute deviation from the sample mean.
}
\details{
Proposed by Garrett, the "Mean Absolute Deviation"--or "Average Deviation (AD)" is "the mean of the deviation of all the separate scores in the series taken from their mean (occasionally from the median or mode)", (1971, p. 481). The statistical literature has not yet adopted a standard notation for the "Mean Absolute Deviation" and the "Median Absolute Deviation". As a result, both statistics have been denoted as "MAD", which may lead to confusion once they may produce different values.
The R \code{\link[stats]{mad}} by default computes the "Median Absolute Deviation"; Thus, to obtain the "Mean Absolute Deviation", one has to use \code{mad(x, constant = 1)}.
The function \code{\link[SciencesPo]{AverageAbsoluteDeviation}} will calculate the "Mean Absolute Deviation"--or "Average Deviation (AD)" right way.
}
\examples{
x <- c(15, 10, 6, 8, 11)
AverageAbsoluteDeviation(x)
mad(x) # median

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Garrett, Henry (1982) \emph{Statistics in Psychology and Education}. 6th, Paragon.
}
\seealso{
\code{\link[stats]{mad}}
}
\keyword{Exploratory}

