% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{baseball}
\alias{baseball}
\title{Bradley Efron and Carl Morris ("Stein's Paradox in Statistics)}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::baseball)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::baseball)} observations.}
\usage{
data(baseball)
}
\description{
A dataset used in Efron and Morris's 1977 paper, "Stein's Paradox in Statistics".

\itemize{
\item {name}{First and last name of selected player}.
\item {atBats}{number of times at bats}.
\item {hits}{number of hits after 45 at bats}.
\item {avg45}{is the average after 45 at bats}.
\item {remainingAtBats}{remaining number at bats}.
\item {remainingAvg}{is the remaining average to the end of the season}.
\item {seasonAtBats}{is the number of times at bats in the end of the season}.
\item {seasonHits}{is the number of hits in the end of the season}.
\item {avgSeason}{is the end of the season average}.
}
}
\keyword{datasets}

