% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{bhodrick93}
\alias{bhodrick93}
\title{Bekaert's and Hodrick's (1993) Data}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::bhodrick93)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::bhodrick93)} observations.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(bhodrick93)
}
\description{
Data set used by Bekaert and Hodrick (1993) on biases in the measurement of foreign exchange risk premiums. This dataset contains the following columns:

\itemize{
 \item date A character vector for date.
 \item jyspot Price of USD in JY, spot.
 \item jyfwd Price of USD in JY, 30-day forward.
 \item jys30 Price of USD in JY, spot market at 30-day forward deliver/date.
 \item dmspot Price of USD in DM, spot.
 \item dmfwd Price of USD in DM, 30-day forward
 \item dms30 Price of USD in DM, spot market at 30-day forward deliver/date.
 \item bpspot Price of USD in BP, spot.
 \item bpfwd Price of USD in BP, 30-day forward.
 \item bps30 Price of USD in BP, spot market at 30-day forward deliver/date.
\item quote A numeric vector.
}
}
\references{
Bekaert, G., and Hodrick, R. J. (1993) On biases in the measurement of foreign exchange risk premiums. \emph{Journal of International Money and Finance,} \bold{12(2),} 115-138.
}
\keyword{datasets}

