% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/between.R
\name{between}
\alias{between}
\title{Classify values into groups}
\usage{
between(x, chunks)
}
\arguments{
\item{x}{Numeric vector to classify}

\item{chunks}{Vector listing what values the grouping should be done on. Should include the max and the min in this list as well.}
}
\value{
Vector of length(x) indicating which group each element is in (for between). Or vector of length(x) indicating the lower bound of the group that it is in.
}
\description{
Classify values into groups based on which numbers they're between.
}
\examples{
age <-sample(1:100, 1000, rep=TRUE)
between(age,c(0, 15, 30, 50, 100))
}
\seealso{
\code{\link{dummy}}.
}

