% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{bollen}
\alias{bollen}
\title{Bollen Data on Industrialization and Political Democracy}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::bollen)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::bollen)} observations.}
\usage{
data(bollen)
}
\description{
The data were reported in Bollen
(1989, p. 428, Table 9.4) This set includes data from
75 developing countries each assessed on four measures
of democracy measured twice (1960 and 1965), and three
measures of industrialization measured once (1960).
\itemize{
\item{y1}{Freedom of the press, 1960}
\item{y2}{Freedom of political opposition, 1960}
\item{y3}{Fairness of elections, 1960}
\item{y4}{Effectiveness of elected legislature, 1960}
\item{y5}{Freedom of the press, 1965}
\item{y6}{Freedom of political opposition, 1965}
\item{y7}{Fairness of elections, 1965}
\item{y8}{Effectiveness of elected legislature, 1965}
\item{x1}{GNP per capita, 1960}
\item{x2}{Energy consumption per capita, 1960}
\item{x3}{Percentage of labor force in industry, 1960}
}
 @references
 Bollen, K. A. (1979). Political democracy and the timing of
 development. \emph{American Sociological Review}, 44, 572-587.

 Bollen, K. A. (1980). Issues in the comparative measurement of
political democracy. \emph{American Sociological Review}, 45, 370-390.
}
\keyword{datasets}

