% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTS.R
\encoding{UTF-8}
\name{bonett.seier}
\alias{bonett.seier}
\title{Bonett-Seier test of Geary's kurtosis}
\usage{
bonett.seier(x, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{A numeric vector of data values.}

\item{alternative}{A character string specifying the alternative hypothesis,
 must be one of '"two.sided"' (default), '"greater"' or '"less"'.
  You can specify just the initial letter

  @details  Under the hypothesis of normality, data should have Geary's
   kurtosis equal to \code{sqrt(2/pi)} (0.7979). This test has such null
    hypothesis and is useful to detect a significant difference of Geary's
     kurtosis in normally distributed data.}
}
\description{
Performs the Bonett-Seier test of Geary's measure of kurtosis for normally distributed data.
}
\examples{
set.seed(1234)
x = rnorm(1000)
geary(x)
bonett.seier(x)
}
\keyword{(2002)}
\keyword{40,}
\keyword{435-445.}
\keyword{@references}
\keyword{A}
\keyword{Analysis,}
\keyword{Bonett,}
\keyword{Computational}
\keyword{D.G.,}
\keyword{Data}
\keyword{E.}
\keyword{Seier,}
\keyword{Statistics}
\keyword{Tests}
\keyword{and}
\keyword{high}
\keyword{normality}
\keyword{of}
\keyword{power.}
\keyword{test}
\keyword{uniform}
\keyword{with}

