% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boot.default}
\alias{boot.default}
\title{Bootstrap Parameters of Statistical Model

This method is the default for bootstrapping statistical models}
\usage{
\method{boot}{default}(mod, ...)
}
\arguments{
\item{mod}{A fitted model object, typically of class \dQuote{lm} or \dQuote{glm}}

\item{...}{A list of optional parameters}
}
\value{
a list with the \dQuote{alpha} and \dQuote{beta} slots set
}
\description{
Bootstrap Parameters of Statistical Model

This method is the default for bootstrapping statistical models
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

