% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boot}
\alias{boot}
\title{Generic Method for Bootstrap

This method is intended to be overried by statistical models that would like
to support statistical bootstrapping.}
\usage{
boot(mod, ...)
}
\arguments{
\item{mod}{a fitted model object that will be used to produce boot-strapped
parameters. This object usually inherits the class \dQuote{glm} or \dQuote{lm} object}

\item{...}{unspecified parameters}
}
\value{
a list with the \dQuote{alpha} and \dQuote{beta} slots set. Note that \dQuote{alpha}
corresponds to ancillary parameters and \dQuote{beta} corresponds to systematic
components of the model
}
\description{
Generic Method for Bootstrap

This method is intended to be overried by statistical models that would like
to support statistical bootstrapping.
}
\note{
This method has private memory storage and can reference the objects:
\dQuote{.fitted}, \dQuote{.data}, \dQuote{.call}, \dQuote{.env}, despite having no declaration in
the argument list.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

