% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bootstrap.model}
\alias{bootstrap.model}
\title{Bootstrap Parameters of a Statistical Model

This method is used for bootstrapping statistical models typically of class \dQuote{lm} or \dQuote{glm}.}
\usage{
\method{bootstrap}{model}(x, ...)
}
\arguments{
\item{x}{A fitted model object, typically of class \dQuote{lm} or \dQuote{glm}}

\item{...}{A list of optional parameters}
}
\value{
a list with the \dQuote{alpha} and \dQuote{beta} slots set
}
\description{
Bootstrap Parameters of a Statistical Model

This method is used for bootstrapping statistical models typically of class \dQuote{lm} or \dQuote{glm}.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

