% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/box.plot.R
\name{box.plot}
\alias{box.plot}
\title{Boxplot}
\usage{
box.plot(x = NULL, data = .data, n.cat = getOption("n.cat"),
  col.fill = getOption("col.fill.bar"),
  col.stroke = getOption("col.stroke.bar"), col.bg = getOption("col.bg"),
  col.grid = getOption("col.grid"), cex.axis = 0.85, col.axis = "gray30",
  xlab = NULL, main = NULL, digits.d = NULL, horiz = TRUE,
  add.points = FALSE, quiet = getOption("quiet"), pdf.file = NULL,
  pdf.width = 5, pdf.height = 5, ...)
}
\arguments{
\item{x}{Variable(s) to analyze. Can be a single variable, either
within a data frame or as a vector in the user's workspace.}

\item{data}{The data frame that contains the variables of interest, default is \code{.data}.}

\item{n.cat}{When analyzing all the variables in a data frame, specifies the largest number of unique values of variable of a numeric data type for which the variable will be analyzed as a categorical. Set to 0 to turn off.}

\item{col.fill}{Color of the box.}

\item{col.stroke}{Color of any points that designate outliers. By default this is the same color as the box.}

\item{col.bg}{Color of the plot background.}

\item{col.grid}{Color of the grid lines.}

\item{cex.axis}{Scale magnification factor, which by defaults displays the axis values to be smaller than the axis labels. Provides the functionality of, and can be replaced by, the standard R \code{cex.axis.}}

\item{col.axis}{Color of the font used to label the axis values.}

\item{xlab}{Label for the value axis, which defaults to the variable's name.}

\item{main}{Title of graph.}

\item{digits.d}{Number of decimal digits displayed in the listing of the summary statistics.}

\item{horiz}{Orientation of the boxplot. Set \code{FALSE} for vertical.}

\item{add.points}{If \code{TRUE}, then place a dot plot (i.e., stripchart) over the box plot.}

\item{quiet}{If set to \code{TRUE}, no text output. Can change system default with \code{\link{set}} function.}

\item{pdf.file}{Name of the pdf file to which graphics are redirected. If there is no filetype of \code{.pdf}, the filetype is added to the name.}

\item{pdf.width}{Width of the pdf file in inches.}

\item{pdf.height}{Height of the pdf file in inches.}

\item{\dots}{Other parameter values for graphics as defined processed by \code{\link{boxplot}} such as \code{whiskcol} for the whisker color, etc. and \code{\link{par}}, including \code{ylim} to set the limits of the value axis, \code{lwd} for the line width, \code{cex.lab} for the size of the label, \code{col.main} for the title, etc., and \code{col.ticks} to specify the color of the tick marks.}
}
\description{
Uses the standard R boxplot function, \code{\link{boxplot}} to display a boxplot in color. Also display the relevant statistics such as the hinges, median and IQR.
}
\examples{
y <- rnorm(100,50,10)
y[1] <- 90

box.plot(y)
}

