% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{brpopage}
\alias{brpopage}
\title{Age Distribution of Population in Brazil}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::brpopage)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::brpopage)} observations.}
\usage{
data(brpopage)
}
\description{
Age Distribution of Population in Brazil.
\itemize{
\item {Year}{year}
\item {AgeGroup}{the age groups}
\item {Thousands}{the counts}
}
}
\keyword{datasets}

