% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{Bush}
\alias{Bush}
\title{Approval Ratings for President George W. Bush}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::Bush)} variables and \Sexpr{nrow(SciencesPo::Bush)} observations.
\itemize{
\item start.date. Start date of the survey.
\item end.date. End date of the survey.
\item approve. Percent which approve of the president.
\item disapprove. Percent which disapprove of the president.
\item undecided. Percent undecided about the president.
}}
\usage{
data(Bush)
}
\description{
Approval ratings for George W. Bush.
}
\keyword{datasets}

