% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/butterfly.R
\encoding{UTF-8}
\name{butterfly}
\alias{butterfly}
\title{The Butterfly Curve}
\usage{
butterfly(n = 100, nb = 500, title = element_blank())
}
\arguments{
\item{n}{An integer for background points.}

\item{nb}{An integer for the butterfly's points.}

\item{title}{A character vector for plot title.}
}
\description{
The butterfly curve is a parametric equation discovered by Temple Fay where two functions in a plane produces a butterfly-like curves.
}
\examples{
if (interactive()) {
butterfly(10, 100, title="10 x 100");
butterfly(10, 200, title="10 x 200");
butterfly(10, 500, title="100 x 500");
butterfly(100, 1000, title="100 x 1000");
}
}
\references{
Fay, Temple H. (May 1989). The Butterfly Curve. \emph{Amer. Math. Monthly} 96 (5): 442-443. doi:10.2307/2325155.
}

