% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{by_day}
\alias{by_day}
\title{A data.frame with time series parameters}
\usage{
by_day(data, FUN, days = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A time-series object}

\item{FUN}{A function to aggregate data}

\item{days}{A numeric value for the number of days to perform aggregation}

\item{na.rm}{A logical value for na.rm, default is \code{FALSE}}
}
\description{
Return a \code{data.frame} with time parameters and time variable
}
\details{
A data.frame of type long with dates formatted as as.Date and aggregated data
}
\note{
To weekly aggregation, set \code{days = 7}
}
\examples{
data(us2012)
\dontrun{Obama.ts <- as.tsdf(us2012[,4], '\%Y-\%m-\%d  \%H:\%M:\%S', us2012[,8])
# Daily aggregate means for data:
daily <- by_day(Obama.ts, mean)
# Weekly aggregated means for data:
weekly <-by_day(Obama.ts, mean, 7) }
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Ripley, B. D. and Hornik, K. (2001) Date-time classes. \emph{R News}, 1/2, 8–11 \url{http://www.r-project.org/doc/Rnews/Rnews_2001-2.pdf}
}
\seealso{
\link{by_month}, \link{by_year}
}
\keyword{Data-Manipulation}

