% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{by_month}
\alias{by_month}
\title{Monthly aggregation}
\usage{
by_month(data, FUN, months = NULL, na.rm = FALSE)
}
\arguments{
\item{data}{A time-series object}

\item{FUN}{A function to aggregate data}

\item{months}{A numeric value for the number of months to perform aggregation}

\item{na.rm}{A logical value for na.rm, default is \code{FALSE}}
}
\description{
Yield for monthly aggregation. It does handle discontinuous time series by months. The user can also set its own method by passing an ad-hoc function for the aggregation method.
}
\details{
A data.frame of type long with dates formatted as as.Date and aggregated data
}
\examples{
data(us2012)
\dontrun{Obama.ts <- as.tsdf(us2012[,4], '\%Y-\%m-\%d  \%H:\%M:\%S', us2012[,8])
# monthly aggregated means for data:
 monthly <-by_month(Obama.ts, mean)
# bimonthly or semimonthly aggregated means for data:
bimonthly <-by_month(Obama.ts, mean, 2) }
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Ripley, B. D. and Hornik, K. (2001) Date-time classes. \emph{R News}, 1/2, 8–11 \url{http://www.r-project.org/doc/Rnews/Rnews_2001-2.pdf}
}
\seealso{
\link{by_day}, \link{by_year}
}
\keyword{Data-Manipulation}

