% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{by_year}
\alias{by_year}
\title{Yearly aggregation}
\usage{
by_year(data, FUN, years = NULL, na.rm = FALSE, civil = FALSE)
}
\arguments{
\item{data}{A time-series object}

\item{FUN}{A function to aggregate data}

\item{years}{A numeric value for the number of years to perform aggregation}

\item{na.rm}{A logical value for na.rm, default is \code{FALSE}}

\item{civil}{Wether the years start from the civil date or from the first observation in data. The default is \code{civil = FALSE}}
}
\description{
Yield for yearly aggregation. It does handle discontinuous time series by years. The user can also set its own method by passing an ad-hoc function for the aggregation method.
}
\details{
A data.frame of type long with dates formatted as as.Date and aggregated data
}
\examples{
data(us2012)
\dontrun{Obama.ts <- as.tsdf(us2012[,4], '\%Y-\%m-\%d  \%H:\%M:\%S', us2012[,8])
# yearly aggregated means for data:
 yearly <-by_year(Obama.ts, mean)
# civil year arrangement:
yearly <-by_year(Obama.ts, mean, civil=TRUE)
# Means aggregated for data:
yearly <-by_year(Obama.ts, mean, 2) }
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Ripley, B. D. and Hornik, K. (2001) Date-time classes. \emph{R News}, 1/2, 8–11 \url{http://www.r-project.org/doc/Rnews/Rnews_2001-2.pdf}
}
\seealso{
\link{by_day}, \link{by_month}
}
\keyword{Data-Manipulation}

