% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{bysort}
\alias{bysort}
\title{Sort data set and related vector}
\usage{
bysort(..., data = .data, inclusive = TRUE)
}
\arguments{
\item{data}{The data frame where all variables of the same length are sorted.}

\item{inclusive}{Whether vectors outside the default data frame should also be sorted.}

\item{\dots}{Index variable(s) used for sorting.}
}
\description{
Sort data set and related vector
}
\examples{
df=data.frame(group=sample(letters,10, TRUE),y=sample(10) )
use(df)
y2 <- y^2
bysort(y, inclusive = FALSE)
y2 # unsorted
bysort(y, inclusive = TRUE)
y2 # sorted
bysort(y, decreasing=TRUE)
.data
}

