% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibratePlot.R
\name{Calibrateplot}
\alias{Calibrateplot}
\title{Calibrate Plot}
\usage{
Calibrateplot(.data = NULL, x, y, ci = 0.95,
  xlab = "Forecast Probability", ylab = "Observed Frequency")
}
\arguments{
\item{.data}{the data frame.}

\item{x, y}{are the forecast and observed data vectors.}

\item{ci}{is the size of the confidence interval}

\item{xlab}{a character name for horizontal axis.}

\item{ylab}{a character name for vertical axis.}
}
\description{
Plots a calibration plot for overlaping observed versus actual values.
}
\examples{
data <- data.frame(x=c(10,20,25,30,35,40,45,50,55,60,65,70,75,80,85,90,95,100),
y=c(0,05,10,20,25,30,40,50,60,70,75,80,95,100,100,100,100,100))

Calibrateplot(data, x, y)

}
\author{
Daniel Marcelino
}

