% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{cathedral}
\alias{cathedral}
\title{Cathedrals}
\format{This data.frame uses names as row label and contains the following columns:}
\usage{
data(cathedral)
}
\description{
Heights and lengths of Gothic and Romanesque cathedrals.

\itemize{
\item Type Romanesque or Gothic.
\item Height Total height, feet.
\item Length Total length, feet.
 }
}
\references{
Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}

