% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{cathedrals}
\alias{cathedrals}
\title{Cathedrals}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::cathedrals)} variables and \Sexpr{nrow(SciencesPo::cathedrals)} observations.}
\usage{
data(cathedrals)
}
\description{
Heights and lengths of Gothic and Romanesque cathedrals. This dataset contains the following columns:

\itemize{
\item Type Romanesque or Gothic.
\item Height Total height, feet.
\item Length Total length, feet.
 }

 @references
Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}

