% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{cgreene76}
\alias{cgreene76}
\title{Christensen's and Greene's (1976) Data}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::cgreene76)} variables and \Sexpr{nrow(SciencesPo::cgreene76)} observations.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(cgreene76)
}
\description{
Data set used by Christensen and Greene (1976) on economies of scale in US electric power generation. This dataset contains the following columns:

\itemize{
\item firmid Observation id.
\item costs Costs in 1970, MM USD.
\item output Output, million KwH.
\item plabor Price of labor.
\item pkap Price of capital.
\item pfuel Price of fuel.
\item labshr Labor's cost share.
\item kapshr Capital's cost share.
}
}
\references{
Christensen, L. R., and Greene, W. H. (1976) Economies of scale in US electric power generation. \emph{The Journal of Political Economy,} 655--676.
}
\keyword{datasets}

