% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{chismoke}
\alias{chismoke}
\title{Smoking and Lung Cancer in China}
\format{A \code{data.frame} object with \Sexpr[eval=TRUE,echo=FALSE,results=text]{ncol(SciencesPo::chismoke)} variables and \Sexpr[eval=TRUE,echo=FALSE,results=text]{nrow(SciencesPo::chismoke)} observations.}
\usage{
data(chismoke)
}
\description{
The Chinese smoking and lung cancer data (Agresti,  Table 5.12, p. 60).
\itemize{
\item {City}{city name}.
\item {Smoker}{if smoker or not}.
\item {Cancer}{if detected cancer or not}.
\item {Count}{the frequencies}.
}
 @references
Agresti, A. \emph{An introduction to categorical Data Analysis}, (2nd ed.). Wiley.
}
\keyword{datasets}

