% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\docType{class}
\encoding{UTF-8}
\name{ci-class}
\alias{ci}
\alias{ci-class}
\title{An S4 Class to Confidence Intervals}
\section{Slots}{

\describe{
\item{\code{lower}}{Lower bound of interval.}

\item{\code{mean}}{Estimated mean.}

\item{\code{upper}}{Upper bound of interval.}

\item{\code{stderr}}{Standard Error of the mean.}
}}

