% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{circularReplicatedSampling}
\alias{circularReplicatedSampling}
\title{Replicated Circular-Systematic Sampling}
\usage{
circularReplicatedSampling(N = 500, n = 30, g = 6)
}
\arguments{
\item{N}{The population size.}

\item{n}{The sample size.}

\item{g}{Number of independent sub-samples, each containing m = n/g units. Notice that m has to be a multiple of n and g.}
}
\description{
Replicated circular systematic sampling.
}
\examples{
circularReplicatedSampling(500, 30, 6)
}

