% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{circularSampling}
\alias{circularSampling}
\title{Circular Systematic Sampling}
\usage{
circularSampling(N = 500, n = 30)
}
\arguments{
\item{N}{The population size.}

\item{n}{The sample size.}
}
\description{
Circular systematic sampling.
}
\examples{
circularSampling(500, 30)
}
\keyword{Sampling}

