% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{colMedians}
\alias{colMedians}
\title{Calculate the Median per Column}
\usage{
colMedians(x, na.rm = FALSE)
}
\arguments{
\item{x}{a data.frame or array.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
 removed before the computation proceeds. The default is \code{FALSE}.}
}
\value{
A vector or array of the medians of each column with dimension \code{dim( x )[-1]}.
}
\description{
Compute the sample medians of the columns (non-rows) of a data.frame or array.
}
\examples{
data(ltaylor96)
colMedians(ltaylor96)
}
\seealso{
\code{\link{rowMedians}},\code{\link{median}},\code{\link{colMeans}}.
}

