% Generated by roxygen2 (4.0.2): do not edit by hand
\name{colNames}
\alias{colNames}
\title{Column names as (always) a character vector}
\usage{
colNames(data, cols)
}
\arguments{
\item{data}{the input \code{data.frame}.}

\item{cols}{the \code{names} or numeric position you want.}
}
\value{
A character vector of the desired names.
}
\description{
A convenience function using either character vectors or numeric vectors to specify a subset of a \code{data.frame}.
}
\examples{
\dontrun{colNames(iris, 1:3)}
}

