% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{compare}
\alias{compare}
\title{Mehtod of comparing things}
\usage{
compare(object, ...)
}
\arguments{
\item{object}{a data object.}

\item{\dots}{Ignored parameters.
other parameters}
}
\description{
Generic function for comparing things.
}

