% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{compareMeans}
\alias{compareMeans}
\title{Compare means between 2 groups}
\usage{
compareMeans(formula, data = parent.frame(), ...)
}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame in which \code{formula} is evaluated.
Note that the default is \code{data=parent.frame()}. This makes it convenient to
use this function interactively by treating the working envionment as if it were
a data frame.}

\item{\dots}{other arguments}
}
\value{
the difference in means between the second and first group
}
\description{
A function to calculate the difference between the means of a continuous
variable for two groups.
}
\examples{
# Some data:
ID=1:10
Age=round(rnorm(10,50,1))
diag=c("Depression","Bipolar");
Diagnosis=sample(diag,10,replace=T);
data=data.frame(ID,Age,Diagnosis);

mean(Age ~ Diagnosis, data=data)
compareMeans(Age ~ Diagnosis, data=data);
}
\keyword{stats}

