% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Condorcet}
\alias{Condorcet}
\title{Condorcet Voting}
\usage{
Condorcet()
}
\description{
Condorcet Voting.
}
\examples{

voters <- structure(
list(Candidates = structure(c(5L, 1L, 2L, 3L, 6L, 4L),
.Label = c("Albert", "Bruce", "Charles", "Edward", "Mary",
"Rose"), class = "factor"),
Vote_A = c(1, 2, 3, 4, 5, 6),
Vote_B = c(2, 4, 1, 5, 6, 3),
Vote_C = c(6, 5, 3, 2, 4, 1),
Vote_D = c(3, 6, 4, 1, 5, 2),
Vote_E = c(2, 5, 1, 4, 3, 6),
Vote_F = c(1, 4, 2, 6, 5, 3),
Vote_G = c(3, 1, 6, 4, 2, 5),
Vote_H = c(4, 2, 5, 1, 6, 3)),
.Names = c("Candidates", "Vote_A", "Vote_B", "Vote_C",
"Vote_D", "Vote_E", "Vote_F", "Vote_G", "Vote_H"),
row.names = c(NA, -6L), class = "data.frame")


}

