% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{ci}
\alias{ci}
\title{Confidence Interval}
\usage{
ci(x, conf.level = 0.95, alpha = 1 - conf.level, ...)
}
\arguments{
\item{x}{a vector of data.}

\item{conf.level}{confidence level. Default is 0.95.}

\item{alpha}{confidence level. Default is 1-conf.level.}

\item{\dots}{Extra parameters.}
}
\value{
\item{CI lower}{Lower bound of interval.}
\item{Est. Mean}{Mean of data.}
\item{CI upper}{Upper bound of interval.}
\item{Std. Error}{Standard Error of the mean.}
}
\description{
Calculates the confidence interval of a vector of data.
}
\examples{
x = rnorm(1000)
ci(x, conf.level=.95)
}
\keyword{univariate}

