% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TESTS.R
\name{contingency}
\alias{contingency}
\alias{contingency.default}
\title{Pearson's Contingency Coefficient for Tables}
\usage{
contingency(x, y = NULL)

\method{contingency}{default}(x, y = NULL)
}
\arguments{
\item{x}{A vector or a matrix.}

\item{y}{A vector that is ignored if x is a matrix and required if x is a vector.}
}
\description{
Computes the Pearson's C coefficient of contingency for tables.
}
\examples{
if (interactive()) {
# some data:
male <- c(33, 76, 6)
female <- c(47, 153, 25)
mat <- cbind( male, female )
colnames(mat) <- c( 'good', 'satisfactory', 'bad')

contingency(mat)
}
}
\references{
Agresti, Alan (1996) \emph{Introduction to categorical data
 analysis}. NY: John Wiley and Sons

Blaikie, N. 2003. \emph{Analyzing Quantative Data}. London: SAGE.
}

