% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{CountMissing}
\alias{CountMissing}
\title{Count missing data.}
\usage{
CountMissing(.data, vars = NULL, share = FALSE, exclude.complete = TRUE)
}
\arguments{
\item{.data}{The \code{data.frame}.}

\item{vars}{a vector with name of variables.}

\item{share}{logical. If \code{TRUE}, it will show the share of missing data by variable.}

\item{exclude.complete}{logical. Exclude complete variables from the output.}
}
\description{
Function to count missing data.
}
\examples{
library(data.table)

a <- c(NA,2,3,4, NA)
b <- c(2,3,4,NA, 3)
c <- c(1,2,3,1, 1)
data <- data.table(a,b,c)
CountMissing(data, share = TRUE)

}

