% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countNAs.R
\name{countNAs}
\alias{countNAs}
\title{Count NA values}
\usage{
countNAs(.data)
}
\arguments{
\item{.data}{The data.frame.}
}
\description{
The function counts NAs for each column of a \code{data.frame}.
}
\examples{
countNAs(ssex)
}

