% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countString.R
\name{countString}
\alias{countString}
\title{Count how often a pattern occurs in a character vector.}
\usage{
countString(x, str, ...)
}
\arguments{
\item{x}{Character vector.}

\item{str}{Regular expression pattern passed to \code{gregexpr}.}

\item{...}{Additional arguments passed to \code{\link{gregexpr}}.}
}
\description{
Count how often a pattern occurs in a character vector.
}
\examples{
countString(c("doo", "dar", "daz"), "a");
countString(c("doo", "dar", "daz"), "^d")
}
\keyword{character}

