% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELECTORALTOOL.R
\name{cox.shugart.inv}
\alias{cox.shugart.inv}
\title{Inverse Cox-Shugart Measure of Proportionality}
\usage{
cox.shugart.inv(v, s, ...)

cox.shugart.inv(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculate the inverse Cox and Shugart measure of
 proportionality based on votes and seats,
 the electoral outcome.
}
\examples{

# 2012 Queensland state elecion:
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

cox.shugart.inv(pvotes, pseats)

# 2012 Quebec provincial election:
pvotes = c(PQ=31.95, Lib=31.20, CAQ=27.05, QS=6.03, Option=1.89, Other=1.88)
pseats = c(PQ=54, Lib=50, CAQ=19, QS=2, Option=0, Other=0)

cox.shugart.inv(pvotes, pseats)

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\seealso{
\code{\link{cox.shugart}}, \code{\link{farina}}, \code{\link{politicalDiversity}}, \code{\link{grofman}}, \code{\link{gallagher}},  \code{\link{lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}.
}

