% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{CoxShugart}
\alias{CoxShugart}
\title{Cox-Shugart Measure of Proportionality}
\usage{
CoxShugart(v, s, ...)

CoxShugart(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculate the Cox and Shugart measure of
 proportionalitybased on a vector of votes and a vector for
 the electoral outcome. This measure is also referred to as the regression index.
}
\examples{
if (interactive()) {
# 2012 Queensland state elecion:
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

CoxShugart(pvotes, pseats)
}

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\seealso{
\code{\link{Farina}}, \code{\link{PoliticalDiversity}}, \code{\link{Grofman}}, \code{\link{Gallagher}},  \code{\link{Lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}.
}

