% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTS.R
\name{cronbach.alpha}
\alias{cronbach.alpha}
\title{Cronbach's Alpha}
\usage{
cronbach.alpha(.data)
}
\arguments{
\item{.data}{\code{data.frame} or matrix with more than 2 columns.}
}
\value{
The Cronbach's alpha value for \code{.data}.
}
\description{
Computes the Cronbach's alpha--a measure of internal consistency ("reliability") for a data frame or matrix.
}
\details{
Rows with \code{NA} values will be ommitted.
}
\note{
A reliability coefficient of .70 or higher is considered "acceptable" in most social science research situations.
}
\examples{
data(alpha)
cronbach.alpha(alpha)

# The alpha coefficient for the four items is .839,
# suggesting that the items have relatively high internal
# consistency.

}
\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of tests. Psychometrika. 16, 297-334.

Tavakol, M., & Dennick, R. (2011). Making sense of Cronbach's alpha. International Journal of Medical Education, 2, 53-55.
}

