% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TESTS.R
\name{cronbach}
\alias{cronbach}
\title{Cronbach's Alpha for a matrix or data frame}
\usage{
cronbach(df)
}
\arguments{
\item{df}{\code{data.frame} or matrix with more than 2 columns.}
}
\value{
The Cronbach's alpha value for \code{df}.
}
\description{
This function calculates the Cronbach's alpha
 value of a data frame or matrix.

#
}
\seealso{
\code{\link{reliability.test}}
}

