% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\name{crosstable}
\alias{crosstable}
\title{Cross-tabulation}
\usage{
crosstable(..., deparse.level = 2)
}
\arguments{
\item{\dots}{The data paremeters.}

\item{digits}{An integer for decimal places.}

\item{latex}{A logical for latex output.}

\item{If}{TRUE, association statistics is shown.}
}
\value{
Well-formatted cross tabulation. Also can genarate latex syntax of cross tabulation.
}
\description{
\code{crosstab} produces all possible two-way tabulations of the variables specified.
}
\examples{
with(titanic, crosstable( SEX, AGE))
with(titanic, crosstable( SEX, AGE, SURVIVED))
}

