% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossTable.R
\encoding{UTF-8}
\name{crossTable}
\alias{crossTable}
\alias{crossTable.default}
\title{Cross-tabulation}
\usage{
crossTable(.data, rowVar, colVar, freqVar = NULL, observed = TRUE,
  expected = FALSE, sresid = FALSE, total.pct = FALSE, row.pct = FALSE,
  col.pct = FALSE, chisq = FALSE, phi = FALSE, cramerV = FALSE,
  contingency = FALSE)

\method{crossTable}{default}(.data, rowVar, colVar, freqVar = NULL,
  observed = TRUE, expected = FALSE, sresid = FALSE, total.pct = FALSE,
  row.pct = FALSE, col.pct = FALSE, chisq = FALSE, phi = FALSE,
  cramerV = FALSE, contingency = FALSE)
}
\arguments{
\item{.data}{The data.frame.}

\item{rowVar}{The row variable.}

\item{colVar}{The column variable.}

\item{freqVar}{The frequency variable.}

\item{observed}{If FALSE, observed values are excluded from the table.}

\item{expected}{If TRUE, expected values are included.}

\item{sresid}{If TRUE, standardized residuals are included.}

\item{total.pct}{If TRUE, total proportions are included.}

\item{row.pct}{If TRUE, row proportions are included.}

\item{col.pct}{If TRUE, column proportions are included.}

\item{chisq}{If TRUE, the result of chi-square test of independence is included.}

\item{phi}{If TRUE, the Phi coefficient is included.}

\item{cramerV}{If TRUE, Cramer's V coefficient is included.}

\item{contingency}{If TRUE, the Contingency coefficient (C) is included.}
}
\description{
Makes a contingency table and computes chi-square test of independence, Phi coefficient, Cramer's V and Contingency coefficient.
}
\examples{
titanic \%>\% crossTable("SEX", "AGE")

crossTable(titanic, "SEX", "AGE",
row.pct=TRUE,
 expected=TRUE,
 sresid=TRUE)

# Two variables
titanic \%>\%
crossTable(c("SEX","AGE"), "SURVIVED")

#' # Agresti (2002), table 3.11, p. 106
# 1992 General Social Survey- Sex and Party affiliation
gss <- data.frame(
   expand.grid(sex=c("female", "male"),
   party=c("dem", "indep", "rep")),
   count=c(279,165,73,47,225,191))

crossTable(gss, "sex", "party", "count")
}

