% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CrossValidation}
\alias{CrossValidation}
\alias{CrossValidation.Forest}
\alias{CrossValidation.Tree}
\title{Cross Validation of SciencesPo Trees and Forests}
\usage{
CrossValidation(object, ...)

\method{CrossValidation}{Tree}(object, FUN, ...)

\method{CrossValidation}{Forest}(object, FUN, ...)
}
\arguments{
\item{object}{a fitted \code{Tree} or \code{Forest} object in the formula.}

\item{...}{additional arguments.}

\item{FUN}{either \code{tree} or \code{forest}}
}
\value{
An object of class \code{CV} with elements: \describe{
  \item{\code{call}}{the function call}
}
}
\description{
Cross Validation of SciencesPo Trees and Forests.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}

