% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMENTS.R
\name{css}
\alias{css}
\alias{css.data.frame}
\alias{css.default}
\title{Corrected Sum of Squares}
\usage{
css(x, na.rm = TRUE)

\method{css}{default}(x, na.rm = TRUE)

\method{css}{data.frame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether \code{NA}
values should be stripped before the computation proceeds.}
}
\description{
Computes the corrected sum of squares.
}

