% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{cumlogit}
\alias{cumlogit}
\title{Cumulative Logit}
\usage{
cumlogit(y, adj = 0.5)
}
\arguments{
\item{y}{the dependent variable}

\item{adj}{adjustment constant}
}
\description{
Cumulative Logit
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

