% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cumulativeLogit.R
\encoding{UTF-8}
\name{cumulativeLogit}
\alias{cumulativeLogit}
\title{Cumulative Logit}
\usage{
cumulativeLogit(y, adj = 0.5)
}
\arguments{
\item{y}{The dependent variable.}

\item{adj}{The adjustment constant.}
}
\description{
Cumulative Logit
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}.
}

