% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hamilton.R
\name{Danish}
\alias{Danish}
\title{# dHondt <- Jefferson <- function(...) quotient(..., f=function(s) s + 1)}
\usage{
Danish(...)
}
\description{
# dHondt <- Jefferson <- function(...) quotient(..., f=function(s) s + 1)
}

