% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{data.split}
\alias{data.split}
\title{Split Data into Test and Train Sets}
\usage{
data.split(x, ratio = 2/3, group = NULL)
}
\arguments{
\item{x}{Vector of data labels.}

\item{ratio}{The spliting ratio. Note that: if (0<=ratio<1), then \code{ratio} fraction of points from \code{x} will be set to \code{TRUE}. if (ratio==1) then one random point from \code{x} will be set to \code{TRUE}. if (ratio>1) then \code{ratio} number of points from \code{x} will be set to \code{TRUE}.}

\item{group}{Optional vector/list used when multiple copies of each sample are present.}
}
\description{
Split data given a vector \code{x} into two sets using a predefined ratio while preserving relative ratios of different labels in \code{x}. It is often used to split data for classification models into train and test subsets.
}
\examples{
data(titanic)
y = titanic[,4] # extract labels from the data
set.seed(88)
survived = data.split(y, ratio=3/5)
table(y,survived)

t=sum( survived)  # n of elements in one class
f=sum( !survived) # n of elements in the other class
stopifnot( round((t+f)*3/5) == t) # test ratios

# use results of data.split to subset data into train and test sets
train = subset(titanic, survived == TRUE)
test  = subset(titanic, survived == FALSE)

}
\keyword{manipulation}

