% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{ddirichlet}
\alias{ddirichlet}
\title{Dirichlet distribution}
\usage{
ddirichlet(x, alpha, log = FALSE, sum.up = FALSE)
}
\arguments{
\item{x}{a matrix containing observations.}

\item{alpha}{the Dirichlet distribution's parameters. Can be a vector (one set of parameters for all observations) or a matrix (a different set of parameters for each observation), see \dQuote{Details}.}

\item{log}{if \code{TRUE}, logarithmic densities are returned.}

\item{sum.up}{if \code{TRUE}, the (log-)likelihood is returned.}
}
\description{
Density function and random number generation for the Dirichlet distribution
}
\examples{
a.mat <- cbind(1:10, 5, 10:1);
a.mat;
X2 <- rdirichlet(10, a.mat);
ddirichlet(X2, a.mat);
}

