% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detach.all.R
\name{detach.all}
\alias{detach.all}
\title{Detach all data frame from the search path}
\usage{
detach.all()
}
\description{
Detach all data frame from the search path, but keeping it on the memory.
}
\examples{
detach.all()
}

