% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot.plot.R
\encoding{UTF-8}
\name{dot.plot}
\alias{dot.plot}
\title{Dot Plot}
\usage{
dot.plot(x, pch = 16, bins = 50, spacing = 1, xlab, ...)
}
\arguments{
\item{x}{The data vector}

\item{pch}{The plotting "character" or symbol, default is dots.}

\item{bins}{The bins width.}

\item{spacing}{A value for vertically spacing between dots.}

\item{xlab}{The axis label.}

\item{\dots}{Other parameters passed on to `plot`.}
}
\description{
Makes a dot plot.
}
\examples{
with(iris, dot.plot(Sepal.Length,
xlab="Length", col = as.numeric(Species)))

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

