% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{dotfy}
\alias{dotfy}
\title{Replace commas by dots}
\usage{
dotfy(x)
}
\arguments{
\item{x}{A vector whose elements contain commas or commas and dots.}
}
\description{
Replace commas by dots in that order.
}
\details{
This function works for numeric vectors, typically currency variables stored in non-english format.
}
\examples{
x <- c('500,00', '0,001', '25.000', '10,100.10', 'him, you, and I.')

dotfy(x)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Data}
\keyword{Manipulation}

