% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keep.R
\name{drop}
\alias{drop}
\title{Keep all columns, except those specified}
\usage{
drop(.data, names, info = TRUE)
}
\arguments{
\item{.data}{dataframe to be modified.}

\item{names}{character vector containing the current name of each variable to be dropped/removed.}

\item{info}{boolean value indicating whether to print details of the renaming.  Defaults to TRUE.}
}
\description{
Stata-like drop function. Keep all columns from the data.frame, except those specified.
}
\examples{
data(women)
drop(women, "weight")
}

