% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\encoding{UTF-8}
\name{Dummify}
\alias{Dummify}
\title{Generate Dummy Variables in a Data Frame}
\usage{
Dummify(x, .data = NULL, drop = TRUE)
}
\arguments{
\item{x}{the column position to generate dummies}

\item{.data}{the a data.frame.}

\item{drop}{A logical value. If \code{TRUE}, unused levels will be omitted.}
}
\description{
Quickly create dummies in a data frame.
}
\details{
A matrix object
}
\examples{
df <- data.frame(y = rnorm(10), x = runif(10,0,1), sex = sample(1:2, 10, rep=TRUE))

Dummify(df$sex) ;

cbind(df, Dummify(df$sex));
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Models}

