% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\encoding{UTF-8}
\name{elapsed}
\alias{elapsed}
\title{Return Elapsed Time in Years}
\usage{
elapsed(from, to, format)
}
\arguments{
\item{from}{the date of origin, typically birthdate}

\item{to}{the date up to compute the age}

\item{format}{the date format see \code{as.Date}}
}
\description{
Return the elapsed time in years.
}
\examples{
elapsed(from="1988-12-19", to="2014-12-31", format="\%Y-\%m-\%d")

elapsed("1jan1960", "2jan1990", "\%d\%b\%Y")
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Data}
\keyword{Descriptive}
\keyword{Manipulation}
\keyword{Stats}

