% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{election2008}
\alias{election2008}
\title{2008 U.S. presidential election}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::election2008)} variables and \Sexpr{nrow(SciencesPo::election2008)} observations.
\itemize{
\item State Name of the state.
\item Abr Abbreviation for the state.
\item Income Per capita income in the state as of 2007 (in dollars).
\item HS Percentage of adults with at least a high school education.
\item BA Percentage of adults with at least a college education.
\item Dem.Rep Difference in \%Democrat-\%Republican (according to 2008 Gallup survey).
\item ObamaWin \code{1}= Obama (Democrat) wins state in 2008 or \code{0}=McCain (Republican wins).
}
 @source
State income data from: Census Bureau Table 659. Personal Income Per Capita (in 2007).
High school data from: U.S. Census Bureau, 1990 Census of Population,
\url{http://www.nces.ed.gov/programs/digest/d08/tables/dt08_011.asp}.
College data from: Census Bureau Table 225. Educational Attainment by State (in 2007)
 Democrat and Republican data from:
\url{http://www.gallup.com/poll/114016/state-states-political-party-affiliation.aspx}}
\usage{
data(election2008)
}
\description{
State-by-state information from the 2008 U.S. presidential election.
}
\keyword{datasets}

