% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\encoding{UTF-8}
\name{eliminateNA}
\alias{eliminateNA}
\title{Eliminate NA observations}
\usage{
eliminateNA(data)
}
\arguments{
\item{data}{the \code{data.frame}}
}
\value{
two matrices.
}
\description{
Generates two matrices: One with complete observations and the other with all observations.
}
\examples{
id <- 1:10; var1 <- rnorm(10); var2 <- rgamma(10, 2, 1);
df <- data.frame(cbind(id, var1, var2))
df[c(5, 9), 3] <- NA
eliminateNA(df)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

