% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ellipses.R
\name{ellipses}
\alias{ellipses}
\title{Draw ellipses}
\usage{
ellipses(x, y, r, ...)
}
\arguments{
\item{x}{The longitude coordinate.}

\item{y}{The latitude coordinate.}

\item{r}{The radio.}

\item{\dots}{Extra parameters to plot, such as col, lty etc.}
}
\description{
Draw ellipses given the coordinates and radio.
}
\examples{
plot(-60:-30, -30:0)
ellipses(-49,-25,7)
}

