% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/env.R
\name{env}
\alias{env}
\title{Display Information about All Loaded Environments}
\usage{
env(unit = c("MB", "KB", "bytes"), digits = 0)
}
\arguments{
\item{unit}{Required unit for displaying environment size: "bytes", "KB", "MB", or first letter.}

\item{digits}{Number of decimals to display when rounding environment size.}
}
\description{
Display name, number of objects, and size of all loaded environments.
}

