% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VISUALIZATION.R
\encoding{UTF-8}
\name{errorbars}
\alias{errorbars}
\title{Add Error Bars on the Graph}
\usage{
errorbars(x, y, ebl, ebu = ebl, length = 0.08, ...)
}
\arguments{
\item{x}{coordinates for the error bars, or simply the name of the graph.}

\item{y}{coordinates for the center of the error bars, which is the group means.}

\item{ebl}{length of the error bars, which should be 1 se in each direction.}

\item{ebu}{the error bars.}

\item{length}{the length.}

\item{\dots}{typically additional noninteresting arguments to pass.}
}
\description{
Draw error bars on the graph.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
x <- c(1:5)
y <- c(1.1,1.5,2.9,3.8,5.2)
sd <- (0.2,0.3,0.2,0.0,0.4)
}

