% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/piping.R
\name{eval_lambda}
\alias{eval_lambda}
\title{evaluate lambda expression
x : object
symbol : symbol part
expr : expression part
envir : environment for evaluation}
\usage{
eval_lambda(x, symbol, expr, envir)
}
\description{
evaluate lambda expression
x : object
symbol : symbol part
expr : expression part
envir : environment for evaluation
}

