% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{experiment}
\alias{experiment}
\title{Mehtod for experiment}
\usage{
experiment(object, ...)
}
\arguments{
\item{object}{a data object.}

\item{\dots}{Ignored parameters.}
}
\description{
Mehtod for experiment
}

