% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPlot.R
\name{exportPlot}
\alias{exportPlot}
\title{Export a plot as PDF, EPS and PNG at once}
\usage{
exportPlot(gplot, filename, width = 2, height = 1.5, dpi = 600)
}
\arguments{
\item{gplot}{the graphic object.}

\item{filename}{the file name.}

\item{width, }{height dimension parameters.}

\item{dpi}{the resolution in dpi used when falling back the output.}
}
\description{
Export a plot as PDF, EPS and PNG at once.
}

