% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{factorial}
\alias{factorial}
\title{Conditional replacement}
\usage{
factorial(n)
}
\arguments{
\item{n}{The number to be factored out.}

\item{.data}{The data object.}
}
\description{
Conditional replacement
}
\examples{
factorial(5)
}

