% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{Farina}
\alias{Farina}
\title{Farina Index}
\usage{
Farina(v, s, ...)

Farina(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates the Farina index also referred to as the cosine proportionality score based on a vector of votes and a vector for the electoral outcome.
}
\examples{

# 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

Farina(pvotes, pseats)

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Koppel, M., and A. Diskin. (2009) Measuring disproportionality, volatility and malapportionment: axiomatization and solutions. Social Choice and Welfare 33, no. 2: 281-286.
}
\seealso{
\code{\link{CoxShugart}}, \code{\link{PoliticalDiversity}}, \code{\link{Grofman}}, \code{\link{Gallagher}},  \code{\link{Lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}.
}

