% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fetchGoogle}
\alias{fetchGoogle}
\title{Fetch data from a web service}
\usage{
fetchGoogle(URL, key = NULL)
}
\arguments{
\item{URL}{the URL to retrieve a CSV file from the service}

\item{key}{for convenience, just the "key" part of the Google link}
}
\description{
Read a data set generated from a web service such as Google Docs.
}
\details{
Generating the URL from the web service will, of course, depend on
how that service is set up.  For Google Spreadsheets, you, the owner of a
spreadsheet, can
(1) open the spreadsheet in a browser
(2) select the File/Publish to the Web menu item
(3) in the resulting dialog box, press "Start publishing"
(4) under "Get a link to the published data", select CSV format
(5) copy the \code{https://docs.google.com/spreadsheet/pub?...} link and post
it where your users can get to it.
}
\note{
The \code{key=} argument is provided as a convenience so that a shorter
character string can be used to refer to a Google document.  Use \code{URL} rather than
\code{key} if you are using a non-Google service or if the Google interface
changes.
\code{fetchData()} expects the spreadsheet to be in a straightforward
rectangular spreadsheet format.
}
\examples{
\dontrun{s = fetchGoogle(key="0Am13enSalO74dEVzMGJSMU5TbTc2eWlWakppQlpjcGc")}
}
\author{
Daniel Kaplan (\email{kaplan@macalester.edu})
}

