% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{Freq}
\alias{Freq}
\alias{Freq.default}
\alias{oneway}
\title{Frequency Table}
\usage{
Freq(.data, x, verbose = TRUE, ...)

\method{Freq}{default}(.data, x, verbose = TRUE, ...)
}
\arguments{
\item{.data}{The data.frame.}

\item{x}{A column for which a frequency of values is desired.}

\item{verbose}{A logical value, if \code{TRUE}, extra statistics are also provided.}

\item{\dots}{Additional arguements (currently ignored)}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
data(cathedrals)

Freq(cathedrals, Type)

cathedrals \%>\% Freq(Height)
}
\seealso{
\code{\link{freq}}, \code{\link{CrossTabs}}.
}

