% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{file.browser}
\alias{file.browser}
\title{Text-based File Interface}
\source{
\url{http://stackoverflow.com/questions/9122600/r-command-line-file-dialog-similar-to-file-choose}
}
\usage{
file.browser(root = getwd(), multiple = FALSE)
}
\arguments{
\item{root}{the root directory to look in}

\item{multiple}{logical. enable multiple files?}
}
\description{
Text-based file interface.
}
\examples{
if (interactive()) {
file.browser()
	}

}

