% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{scale_fill_party}
\alias{scale_fill_party}
\title{Political Parties Color Palette (Discrete) and Scales}
\usage{
scale_fill_party(palette = "BRA", ...)
}
\arguments{
\item{palette}{the palette name, a character string.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
An N-color discrete palette for political parties.
}
\seealso{
\code{\link{party_color_pal}} for details and references.
}
\keyword{ggplot2}

