% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CONVENIENCE.R
\name{fillNAByPreviousData}
\alias{fillNAByPreviousData}
\title{Fill NA by Previous Value}
\usage{
fillNAByPreviousData(column)
}
\arguments{
\item{column}{the column that contains NAs}
}
\description{
Fill NA by Previous Value
}
\examples{
data(ssex)
peek(ssex)

fillNAByPreviousData(ssex$Favor)
}

