% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fillup.R
\encoding{UTF-8}
\name{fillup}
\alias{fillup}
\title{Identify columns with at least one NA value and fill with the mean value}
\usage{
fillup(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\value{
A message indicating whether any column in \code{x} has missing data and filling missings with the mean of the valid cases.
}
\description{
Identify columns with at least one NA value and fill with the mean value
}
\examples{
view(ssex)
    fillup(ssex)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\keyword{Missings}

