% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findMode}
\alias{findMode}
\title{Find the most frequently occurring value (mode)}
\usage{
findMode(x)
}
\description{
This function determines the most frequently occurring value in an integer
vector (mode). If the mode is ambiguous, a function returns any mode.
}
\details{
ARGUMENTS:
x - a integer vector

RETURN VALUE:
a integer value
}
\seealso{
Other Exploratory: \code{\link{allPossiblePermutations}}
}

