% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mi.R
\encoding{UTF-8}
\name{fix.missing}
\alias{fix.missing}
\title{Identify columns with at least one NA value}
\usage{
fix.missing(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\value{
A message indicating whether any column in \code{x} has missing data and filling missings with the mean of the valid cases.
}
\description{
Identify columns with at least one NA value
}
\examples{
data(ssex)
    ssex_fixed <-fix.missing(ssex)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

