% Generated by roxygen2 (4.0.2): do not edit by hand
\name{followup}
\alias{followup}
\title{Mark Followup Interviews by id and time}
\usage{
followup(id, time)
}
\arguments{
\item{id}{the identification of the subject.}

\item{time}{the date followup contact.}
}
\description{
Mark Followup Interviews by id and time
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

