% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fontSelector.R
\name{fontSelector}
\alias{fontSelector}
\title{Font selector, internal helper function}
\usage{
fontSelector(...)
}
\arguments{
\item{...}{list of fonts requested}
}
\description{
Will return the first available font
}

