% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/footnote.R
\name{footnote}
\alias{footnote}
\title{Generate a footnote for graphs

Generate a footnote for graphs}
\usage{
footnote(study = NULL, adj = 0, line = 3, ...)
}
\arguments{
\item{study}{The study id.}

\item{adj}{adj=0 left align the footnote while adj=0.5 or 1 center or right align the text.}

\item{line}{The line to plot footnote.}

\item{\dots}{Additional parameters.}
}
\description{
Generate a footnote for graphs

Generate a footnote for graphs
}
\examples{
plot(1:10,10:1)
footnote()
footnote(study="vote:16")
footnote(line=-1, adj=.5, cex=.5)
}

