% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{Forge}
\alias{Forge}
\title{Veritical, left-aligned layout for plots}
\usage{
Forge(...)
}
\arguments{
\item{...}{one or more waffle plots}
}
\description{
Left-align the waffle plots by x-axis. Use the \code{pad} parameter in
\code{waffle} to pad each plot to the max width (num of squares), otherwise
the plots will be scaled.
}
\examples{
parts <- c(80, 30, 20, 10)

w1 <- Waffleplot(parts, rows=8)
w2 <- Waffleplot(parts, rows=8)
w3 <- Waffleplot(parts, rows=8)
chart <- Forge(w1, w2, w3)
print(chart)

}

