% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatNumbers.R
\encoding{UTF-8}
\name{formatNumbers}
\alias{formatNumbers}
\title{Format numeric digits}
\usage{
formatNumbers(x, digits = 2)
}
\arguments{
\item{x}{the object whose values to format}

\item{digits}{An integer for the number of decimal places.}
}
\description{
Format numeric digits
}

