% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format_numeric.R
\encoding{UTF-8}
\name{formatNumeric}
\alias{formatNumeric}
\title{Format numeric digits}
\usage{
formatNumeric(x, digits = 2)
}
\arguments{
\item{x}{the object whose values to format}

\item{digits}{an integer for the number of decimal places.}
}
\description{
Format numeric digits
}

