% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{formatR}
\alias{formatR}
\title{Format numeric digits}
\usage{
formatR(x, digits = 3)
}
\arguments{
\item{x}{the object whose values to format}

\item{digits}{an integer for the number of decimal places.}
}
\description{
Format numeric digits
}

