% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/frac.R
\name{frac}
\alias{frac}
\title{Fractional part of a numeric vector.}
\usage{
frac(x, digits = NA)
}
\arguments{
\item{x}{A vector of numerics.}

\item{digits}{An integer. If not missing, the fractional part will be rounded.}
}
\description{
Split off fractional part of a numeric vector, compute and evaluate continuous fractions.
}
\examples{
frac(c(0, pi, 2*pi))
}
\keyword{Describe}

