% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq.plot.R
\name{freq.plot}
\alias{freq.plot}
\title{Frequency Plot of One or Two Variables}
\usage{
freq.plot(x = NULL, by = NULL, data = .data, n.cat = getOption("n.cat"),
  col.fill = NULL, col.stroke = getOption("col.stroke.bar"),
  col.bg = getOption("col.bg"), col.grid = getOption("col.grid"),
  random.col = FALSE, colors = c("rainbow", "terrain", "heat"),
  horiz = FALSE, over.grid = FALSE, addtop = 1, gap = NULL,
  prop = FALSE, xlab = NULL, ylab = NULL, main = NULL,
  cex.axis = 0.85, col.axis = "gray30", beside = FALSE, col.low = NULL,
  col.hi = NULL, count.levels = NULL, legend.title = NULL,
  legend.loc = "right.margin", legend.labels = NULL, legend.horiz = FALSE,
  quiet = getOption("quiet"), pdf.file = NULL, pdf.width = 5,
  pdf.height = 5, ...)
}
\arguments{
\item{x}{Variable(s) to analyze. Can be a single variable, either
within a data frame or as a vector in the user's workspace.}

\item{by}{For each level of the first variable, x, display the frequencies at each level of this second variable, y.}

\item{data}{The data frame that contains the variables of interest, default is \code{.data}.}

\item{n.cat}{When analyzing all the variables in a data frame, specifies the largest number of unique values of variable of a numeric data type for which the variable will be analyzed as a categorical. Set to 0 to turn off.}

\item{col.fill}{Specified bar colors.}

\item{col.stroke}{Color of the border of the bars. Black by default unless the background is dark. Specify NA to remove the border.}

\item{col.bg}{Color of the plot background.}

\item{col.grid}{Color of the grid lines.}

\item{random.col}{For two variable plots, when \code{TRUE}, randomizes the order of the colors within the chosen color palette, when the second variable is not ordered, otherwise is ignored. Each analysis generally yields different colors of the bars.}

\item{colors}{Optional palettes that provide more options, which include values of \code{"heat"}, \code{"rainbow"} and \code{"terrain"}.}

\item{horiz}{By default bars are vertical, but can set this option to \code{TRUE}.}

\item{over.grid}{If \code{TRUE}, plot the grid lines over the histogram.}

\item{addtop}{When \code{horiz=FALSE}, in the same scale as the vertical axis, puts more space between the bars and the top of the plot area, usually to accommodate the legend when plotting two variables.}

\item{gap}{Gap between bars. Provides the value of the \code{space} option from the standard R \code{\link{barplot}} function with a default of 0.2 unless two variables are plotted and beside=\code{TRUE}, in which case the default is c(.1,1).}

\item{prop}{Display proportions instead raw frequencies.}

\item{xlab}{Label for x-axis, more generally the label for the values which could be on the vertical axis for a two variable plot if \code{horiz=TRUE}. Defaults to variable name.}

\item{ylab}{Label for y-axis, more generally the frequency axis. Defaults to Frequency.}

\item{main}{Title of graph.}

\item{cex.axis}{Scale magnification factor, which by default displays the axis values to be smaller than the axis labels.}

\item{col.axis}{Color of the font used to label the axis values.}

\item{beside}{For a two variable plot, set to \code{TRUE} for the levels of the first variable to be plotted as adjacent bars instead of stacked on each other.}

\item{col.low}{Only when the variable is an ordered factor, sets the color for the lowest level of the factor in the resulting ordered progression of colors.}

\item{col.hi}{Only when the variable is an ordered factor, sets the color for the highest level of the factor in the resulting ordered progression of colors.}

\item{count.levels}{If the name of a variable, this signals that the primary variable \code{x} has values that are counts, already tabulated, and that the specified variable here contains the names of the levels of x.}

\item{legend.title}{Title of the legend, which is usually set by default except when raw counts are entered as a matrix. Then a title must be specified to generate a legend.}

\item{legend.loc}{When plotting two variables, location of the legend, with the default in the right margin. Additional options from standard R are "topleft", "top", "topright" and others as shown in the help for the \code{\link{legend}} function.}

\item{legend.labels}{When plotting two variables, labels for the legend, which by default are the levels for the second or \code{by} variable.}

\item{legend.horiz}{By default the legend is vertical, but can be changed to horizontal.}

\item{quiet}{If set to \code{TRUE}, no text output. Can change system default with \code{\link{set}} function.}

\item{pdf.file}{Name of the pdf file to which graphics are redirected.}

\item{pdf.width}{Width of the pdf file in inches.}

\item{pdf.height}{Height}{of the pdf file in inches.}

\item{\dots}{Other parameter values for graphics as defined by \code{\link{barplot}}, \code{\link{legend}}, and  \code{\link{par}} including \code{space} for one variable only, \code{las=2} for vertical axis labels, and \code{cex.lab}, \code{col.main}, etc., and \code{col.ticks} to specify the color of the tick marks.}
}
\description{
Plots a bar chart of one or two variables with default colors,
including background color and grid lines, from a variety of
different types of data. For two variables a legend is also provided.
}
\examples{
travel <- sample(c("Bike", "Bus", "Car", "Motorcycle"), size=25, replace=TRUE)
freq.plot(travel, horiz=TRUE)
}

