% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency table}
\usage{
freq(x, ...)
}
\arguments{
\item{x}{The data.frame}

\item{\dots}{Further arguments.}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
freq2(titanic$CLASS)
}

