% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frequency.R
\encoding{UTF-8}
\name{Frequency}
\alias{Freq}
\alias{Frequency}
\alias{Frequency.default}
\title{Frequency Table}
\usage{
Frequency(.data, x = NULL, verbose = TRUE, ...)

\method{Frequency}{default}(.data, x = NULL, verbose = TRUE, ...)

Freq(.data, x = NULL, verbose = TRUE, ...)
}
\arguments{
\item{.data}{The data.frame.}

\item{x}{A column for which a frequency of values is desired.}

\item{verbose}{A logical value, if \code{TRUE}, extra statistics are also provided.}

\item{\dots}{Additional arguements (currently ignored)}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
data(cathedrals)

Frequency(cathedrals, Type)

# may work with operators like \%>\%
cathedrals \%>\% Frequency(Height)

}
\seealso{
\code{\link{freq}}, \code{\link{Crosstable}}.
}

