% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{fte_color_pal}
\alias{fte_color_pal}
\title{Extended fivethirtyeight.com color palette}
\usage{
fte_color_pal()
}
\description{
The standard fivethirtyeight.com palette for line plots is blue, red, green.
 I add an orange ton.
}
\examples{
scales::show_col(fte_color_pal()(4))
}
\seealso{
Other colour fte: \code{\link{scale_color_fte}},
  \code{\link{scale_colour_fte}},
  \code{\link{scale_fill_fte}}
}

