\name{ge2010}
\alias{ge2010}
\docType{data}
\title{
2010 British General Election Data
}
\description{
This is the 2010 British general election data as made available by the electoral commission.
}
\usage{data(ge2010)}
\format{
  A data frame with 651 observations on the following 144 variables.
  \describe{
    \item{\code{Ref}}{a numeric vector}
    \item{\code{Constituency.Name}}{a character vector}
    \item{\code{Region}}{a character vector}
    \item{\code{Election.Year}}{a numeric vector}
    \item{\code{Electorate}}{a numeric vector}
    \item{\code{Votes}}{a numeric vector}
    \item{\code{AC}}{a numeric vector}
    \item{\code{AD}}{a numeric vector}
    \item{\code{AGS}}{a numeric vector}
    \item{\code{APNI}}{a numeric vector}
    \item{\code{APP}}{a numeric vector}
    \item{\code{AWL}}{a numeric vector}
    \item{\code{AWP}}{a numeric vector}
    \item{\code{BB}}{a numeric vector}
    \item{\code{BCP}}{a numeric vector}
    \item{\code{Bean}}{a numeric vector}
    \item{\code{Best}}{a numeric vector}
    \item{\code{BGPV}}{a numeric vector}
    \item{\code{BIB}}{a numeric vector}
    \item{\code{BIC}}{a numeric vector}
    \item{\code{Blue}}{a numeric vector}
    \item{\code{BNP}}{a numeric vector}
    \item{\code{BP.Elvis}}{a numeric vector}
    \item{\code{C28}}{a numeric vector}
    \item{\code{Cam.Soc}}{a numeric vector}
    \item{\code{CG}}{a numeric vector}
    \item{\code{Ch.M}}{a numeric vector}
    \item{\code{Ch.P}}{a numeric vector}
    \item{\code{CIP}}{a numeric vector}
    \item{\code{CITY}}{a numeric vector}
    \item{\code{CNPG}}{a numeric vector}
    \item{\code{Comm}}{a numeric vector}
    \item{\code{Comm.L}}{a numeric vector}
    \item{\code{Con}}{a numeric vector}
    \item{\code{Cor.D}}{a numeric vector}
    \item{\code{CPA}}{a numeric vector}
    \item{\code{CSP}}{a numeric vector}
    \item{\code{CTDP}}{a numeric vector}
    \item{\code{CURE}}{a numeric vector}
    \item{\code{D.Lab}}{a numeric vector}
    \item{\code{D.Nat}}{a numeric vector}
    \item{\code{DDP}}{a numeric vector}
    \item{\code{DUP}}{a numeric vector}
    \item{\code{ED}}{a numeric vector}
    \item{\code{EIP}}{a numeric vector}
    \item{\code{EPA}}{a numeric vector}
    \item{\code{FAWG}}{a numeric vector}
    \item{\code{FDP}}{a numeric vector}
    \item{\code{FFR}}{a numeric vector}
    \item{\code{Grn}}{a numeric vector}
    \item{\code{GSOT}}{a numeric vector}
    \item{\code{Hum}}{a numeric vector}
    \item{\code{ICHC}}{a numeric vector}
    \item{\code{IEAC}}{a numeric vector}
    \item{\code{IFED}}{a numeric vector}
    \item{\code{ILEU}}{a numeric vector}
    \item{\code{Impact}}{a numeric vector}
    \item{\code{Ind1}}{a numeric vector}
    \item{\code{Ind2}}{a numeric vector}
    \item{\code{Ind3}}{a numeric vector}
    \item{\code{Ind4}}{a numeric vector}
    \item{\code{Ind5}}{a numeric vector}
    \item{\code{IPT}}{a numeric vector}
    \item{\code{ISGB}}{a numeric vector}
    \item{\code{ISQM}}{a numeric vector}
    \item{\code{IUK}}{a numeric vector}
    \item{\code{IVH}}{a numeric vector}
    \item{\code{IZB}}{a numeric vector}
    \item{\code{JAC}}{a numeric vector}
    \item{\code{Joy}}{a numeric vector}
    \item{\code{JP}}{a numeric vector}
    \item{\code{Lab}}{a numeric vector}
    \item{\code{Land}}{a numeric vector}
    \item{\code{LD}}{a numeric vector}
    \item{\code{Lib}}{a numeric vector}
    \item{\code{Libert}}{a numeric vector}
    \item{\code{LIND}}{a numeric vector}
    \item{\code{LLPB}}{a numeric vector}
    \item{\code{LTT}}{a numeric vector}
    \item{\code{MACI}}{a numeric vector}
    \item{\code{MCP}}{a numeric vector}
    \item{\code{MEDI}}{a numeric vector}
    \item{\code{MEP}}{a numeric vector}
    \item{\code{MIF}}{a numeric vector}
    \item{\code{MK}}{a numeric vector}
    \item{\code{MPEA}}{a numeric vector}
    \item{\code{MRLP}}{a numeric vector}
    \item{\code{MRP}}{a numeric vector}
    \item{\code{Nat.Lib}}{a numeric vector}
    \item{\code{NCDV}}{a numeric vector}
    \item{\code{ND}}{a numeric vector}
    \item{\code{New}}{a numeric vector}
    \item{\code{NF}}{a numeric vector}
    \item{\code{NFP}}{a numeric vector}
    \item{\code{NICF}}{a numeric vector}
    \item{\code{Nobody}}{a numeric vector}
    \item{\code{NSPS}}{a numeric vector}
    \item{\code{PBP}}{a numeric vector}
    \item{\code{PC}}{a numeric vector}
    \item{\code{Pirate}}{a numeric vector}
    \item{\code{PNDP}}{a numeric vector}
    \item{\code{Poet}}{a numeric vector}
    \item{\code{PPBF}}{a numeric vector}
    \item{\code{PPE}}{a numeric vector}
    \item{\code{PPNV}}{a numeric vector}
    \item{\code{Reform}}{a numeric vector}
    \item{\code{Respect}}{a numeric vector}
    \item{\code{Rest}}{a numeric vector}
    \item{\code{RRG}}{a numeric vector}
    \item{\code{RTBP}}{a numeric vector}
    \item{\code{SACL}}{a numeric vector}
    \item{\code{Sci}}{a numeric vector}
    \item{\code{SDLP}}{a numeric vector}
    \item{\code{SEP}}{a numeric vector}
    \item{\code{SF}}{a numeric vector}
    \item{\code{SIG}}{a numeric vector}
    \item{\code{SJP}}{a numeric vector}
    \item{\code{SKGP}}{a numeric vector}
    \item{\code{SMA}}{a numeric vector}
    \item{\code{SMRA}}{a numeric vector}
    \item{\code{SNP}}{a numeric vector}
    \item{\code{Soc}}{a numeric vector}
    \item{\code{Soc.Alt}}{a numeric vector}
    \item{\code{Soc.Dem}}{a numeric vector}
    \item{\code{Soc.Lab}}{a numeric vector}
    \item{\code{South}}{a numeric vector}
    \item{\code{Speaker}}{a numeric vector}
    \item{\code{SSP}}{a numeric vector}
    \item{\code{TF}}{a numeric vector}
    \item{\code{TOC}}{a numeric vector}
    \item{\code{Trust}}{a numeric vector}
    \item{\code{TUSC}}{a numeric vector}
    \item{\code{TUV}}{a numeric vector}
    \item{\code{UCUNF}}{a numeric vector}
    \item{\code{UKIP}}{a numeric vector}
    \item{\code{UPS}}{a numeric vector}
    \item{\code{UV}}{a numeric vector}
    \item{\code{VCCA}}{a numeric vector}
    \item{\code{Vote}}{a numeric vector}
    \item{\code{Wessex.Reg}}{a numeric vector}
    \item{\code{WRP}}{a numeric vector}
    \item{\code{You}}{a numeric vector}
    \item{\code{Youth}}{a numeric vector}
    \item{\code{YRDPL}}{a numeric vector}
  }
}
\source{\url{http://www.electoralcommission.org.uk/elections/}
}
\references{
Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}
\examples{
data(ge2010)
}
\keyword{datasets}
