% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geary.R
\encoding{UTF-8}
\name{geary}
\alias{geary}
\title{Geary's test for normality}
\usage{
geary(x)
}
\arguments{
\item{x}{the numeric vector.}
}
\value{
statistic The Geary's test statistic G

p.value The significant probability of the null-hypothesis testing.
}
\description{
Geary's test for normality. Null hypothesis is that the data obeys to normal distribution.
}
\examples{
set.seed(1234)
x = rnorm(1000)
skewness(x)
geary(x)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

