% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{geary.test}
\alias{geary.test}
\title{Geary's test for normality}
\usage{
geary.test(x)
}
\arguments{
\item{x}{the numeric vector.}
}
\value{
statistic The Geary's test statistic G

p.value The significant probability of the null-hypothesis testing.
}
\description{
Geary's test for normality. Null hypothesis is that the data obeys to normal distribution.
}
\examples{
s <-sample(100, 20)
geary.test(s)
geary.test(rnorm(100))
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

