% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{geometricMean}
\alias{geometricMean}
\title{Calculate a geometric mean}
\usage{
geometricMean(x, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{further arguments passed to \code{\link{prod}}}
}
\description{
Calculate a geometric mean
}
\seealso{
\code{\link{harmonicMean}} and \code{\link{mean}}
}

