\encoding{UTF-8}
\name{get.valid}
\alias{get.valid}
\title{Get valid observations in a data frame
}
\description{Indicate which observations are valid.
}
\usage{
get.valid(x)
}
\arguments{
  \item{x}{ A data frame or vector
}
}

\value{ Numeric values
}
\author{Daniel Marcelino <dmarcelino@live.com> 
}
\examples{
data(montreal)
get.valid(montreal)
}
\keyword{ descriptive }
