% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getAttrNames}
\alias{getAttrNames}
\title{Helper function for determining the vector of attribute names
of a given object.}
\usage{
getAttrNames(obj)
}
\arguments{
\item{obj}{[\code{mixed}]\cr
Source object.}
}
\value{
[\code{character}]
  Vector of attribute names for the source object.
}
\description{
Helper function for determining the vector of attribute names
of a given object.
}

