% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Load or import a data file}
\usage{
getdata(file = NULL, format = c("csv", "SPSS", "R", "Excel", "SAS"),
  labels = NULL, widths = NULL, missing = "", n.mcut = 1,
  miss.show = 30, miss.zero = FALSE, miss.matrix = FALSE,
  max.lines = 30, sheet = 1, brief = getOption("brief"),
  quiet = getOption("quiet"), fun.call = NULL, ...)
}
\arguments{
\item{file}{The file reference, either omitted to browse for the data file, or (except for .Rda files) a full path name or web URL, included in quotes.  A URL begins with \code{http://}.}

\item{format}{Format of the data in the file, which by default is a \code{csv} file, which also will recognize tab-delimited text. As an option can be an Excel \code{.xls} or \code{.xlsx} file or an SPSS \code{.sav} file, which also reads the variable labels if present, or a native R data file with a file type of \code{.rda}, or a (native R) data file.}

\item{labels}{File name for the file of variable labels. Either a full path name, or just the file name if in the same directory as the data file, or no reference between the quotes, which allows the user to browse for the labels file. Or, if \code{row2}, then the labels are in the second line of the data file.}

\item{widths}{Specifies the width of the successive columns for fixed width formatted data.}

\item{missing}{Missing value code, which by default is literally a missing data value in the data table.}

\item{n.mcut}{For the missing value analysis, list the row name and number of missing values if the number of missing exceeds or equals this cutoff.}

\item{miss.show}{For the missing value analysis, the number of rows, one row per observation, that has as many or missing values as \code{n.mcut}.}

\item{miss.zero}{For the missing value analysis, list the variable name or the row name even for values of 0. By default only variables and rows with missing data are listed.}

\item{miss.matrix}{For the missing value analysis, if there is any missing data, list a version of the complete data table with a 0 for a non-missing value and a 1 for a missing value.}

\item{max.lines}{Maximum number of lines to list of the data and labels.}

\item{sheet}{For Excel files, specifies the work sheet to read. The default is the first work sheet.}

\item{brief}{If \code{TRUE}, display only variable names table plus any variable labels.}

\item{quiet}{If set to \code{TRUE}, no text output. Can change the corresponding system default with \code{\link{set}} function.}

\item{fun.call}{Function call. Used with \code{knitr} to pass the function call when obtained from the abbreviated function call \code{rd}.}

\item{\dots}{Other parameter values define with the R read functions, such as the \code{read.table} function for text files, with row.names and header such as \code{sep} and \code{dec}.}
}
\description{
Load or import a data file from "csv", "SPSS", "R", "Excel", "SAS" formats with optional parameters.
}
\examples{
\dontrun{
 getdata("~/daniel/MyData.xlsx", "Excel");
 getdata("~/daniel/MyData.sav", "SPSS");
 getdata("~/daniel/MyData.RData", "R");
 getdata("~/daniel/MyData.sas7bdat", "SAS");
}
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

