% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dirichlet.R
\encoding{UTF-8}
\name{getDirichletDensity}
\alias{getDirichletDensity}
\title{The Dirichlet-Multinomial Distribution}
\usage{
getDirichletDensity(x, alpha, log = FALSE, sum = FALSE)
}
\arguments{
\item{x}{a matrix containing observations.}

\item{alpha}{the Dirichlet distribution's parameters. Can be a vector (one set of parameters for all observations) or a matrix (a different set of parameters for each observation), see \dQuote{Details}.}

\item{log}{if \code{TRUE}, logarithmic densities are returned.}

\item{sum}{if \code{TRUE}, the (log-)likelihood is returned.}
}
\value{
the \code{getDirichletSamples} returns a vector of densities (if \code{sum = FALSE}) or the (log-)likelihood (if \code{sum = TRUE}) for the given data and alphas.
}
\description{
Density function and random number generation for the Dirichlet distribution
}
\examples{
mat <- cbind(1:10, 5, 10:1);
mat;
draws <- getDirichletSamples(10, mat);

getDirichletDensity(draws, mat);

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Simulation}

