% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getGoogleMapsAddress}
\alias{getGoogleMapsAddress}
\title{Gets Google Maps address}
\usage{
getGoogleMapsAddress(street = "Banacha 2", city = "Warszawa",
  country = "Poland", positionOnly = TRUE, delay = 1)
}
\description{
Gets Google Maps address
}

