% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{getGoogleMapsSignleAddress}
\alias{getGoogleMapsSignleAddress}
\title{Gets Google Maps single address}
\usage{
getGoogleMapsSignleAddress(apiHttp, positionOnly = TRUE)
}
\description{
Gets Google Maps single address
}

