% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggMultiplot.R
\name{ggplotArrange}
\alias{ggplotArrange}
\title{Arrange multiple ggplot objects}
\usage{
ggplotArrange(...)
}
\arguments{
\item{\dots}{a list of ggplot objects}
}
\description{
Arrange multiple ggplot objects
}

