% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.simpson.R
\docType{methods}
\encoding{UTF-8}
\name{gini}
\alias{gini}
\alias{gini,ANY-method}
\alias{gini,numeric,numeric,ANY-method}
\title{Weighted Gini Index}
\usage{
gini(x, weights = rep(1, length = length(x)))

\S4method{gini}{ANY}(x, weights = rep(1, length = length(x)))
}
\arguments{
\item{x}{A data.frame, a matrix-like, or a vector.}

\item{weights}{A vector containing weights for \code{x}.}
}
\description{
Computes the unweighted and weighted Gini coefficient.
}
\examples{
# generate a vector (of incomes)
x <- c(778, 815, 857, 888, 925, 930, 965, 990, 1012)
# compute Gini index
gini(x)
# generate some weights:
wgt <- runif(n=length(x))
# compute the weighted Gini index
gini(x, wgt)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{gini.simpson}}.
}
\keyword{Concentration}
\keyword{Diversity,}

