% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/piping.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe an object forward}
\usage{
x \%>\% expr
}
\arguments{
\item{x}{object}

\item{expr}{expression}
}
\description{
The \code{\%>\%} operator pipes the object on the left-hand side to the right-hand side according to the syntax.
}
\details{
Pipe operator \code{\%>\%} determines the piping mechanism by the syntax of the expression on the right-hand side.
}
\examples{
# Pipe to . in an expression enclosed by parentheses representing the piped object
rnorm(100) \%>\% (plot(.,col="red",main=length(.)))

# Pipe to an expression enclosed by parentheses with lambda expression in the form of x ~ f(x).
rnorm(100) \%>\% (x ~ plot(x,col="red",main=length(x)))

# Pipe to an expression for side effect and return the input value
rnorm(100) \%>\%
  (~ cat("Number of points:",length(.))) \%>\%
  summary
# Pipe for element extraction
mtcars \%>\% (mpg)

# Pipe for element evaluation
 mtcars \%>\% (mpg) \%>\% ci()

# Pipe for questioning and printing
rnorm(100) \%>\%
  (? summary(.)) \%>\%
  plot(col="red")

mtcars \%>\%
  "My Regression Results" \%>\%
  lm(formula = mpg ~ wt + cyl) \%>\%
  summary \%>\%
  coef

mtcars \%>\%
  ("Sample data" ? head(., 3)) \%>\%
  lm(formula = mpg ~ wt + cyl) \%>\%
  summary \%>\%
  coef
}

