% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenation operator}
\usage{
... \%+\% NA
}
\arguments{
\item{\dots}{Paramenters}
}
\description{
Intuitive and handy function to concatanete things.
}
\examples{
"var" \%+\% 1:3
strings <- c("abc")
strings \%+\% "dfg"
"Y~" \%+\% paste0("z",1:3, "*x",1:3,collapse="+")
}

