\name{griliches76}
\alias{griliches76}
\docType{data}
\title{Griliches's (1976) Data

}
\description{Data set used by Griliches (1976) on wages of very young men. 
}
\usage{data(griliches76)}
\format{
  A data frame with 758 observations on the following 20 variables.
  \describe{
    \item{\code{rns}}{residency in South.}
    \item{\code{rns80}}{a numeric vector.}
    \item{\code{mrt}}{ marital status = 1 if married.}
    \item{\code{mrt80}}{a numeric vector.}
    \item{\code{smsa}}{reside metro area = 1 if urban.}
    \item{\code{smsa80}}{a numeric vector.}
    \item{\code{med}}{mother's education, years.}
    \item{\code{iq}}{iq score.}
    \item{\code{kww}}{score on knowledge in world of work test.}
    \item{\code{year}}{Year.}
    \item{\code{age}}{a numeric vector.}
    \item{\code{age80}}{a numeric vector.}
    \item{\code{s}}{completed years of schooling.}
    \item{\code{s80}}{a numeric vector.}
    \item{\code{expr}}{experience, years.}
    \item{\code{expr80}}{a numeric vector.}
    \item{\code{tenure}}{tenure, years.}
    \item{\code{tenure80}}{a numeric vector.}
    \item{\code{lw}}{log wage.}
    \item{\code{lw80}}{a numeric vector.}
  }
}

\source{Hayashi, F. (2000) \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}

}
\references{
Griliches, Z. (1976) Wages of very young men. \emph{The Journal of Political Economy,} \bold{84(4),} 69--85.
}
\examples{
data(griliches76)
}
\keyword{datasets}
