% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{haberman}
\alias{haberman}
\title{Haberman's Positive Margin no Three-way Interaction MLE}
\format{A 2x2x2 (contingency) table}
\usage{
data(haberman)
}
\description{
Haberman's example of a log-linear model (the no-three way
interaction model) that lacks an MLE even though it has positive
margins.
}
\author{
Haberman, S.
}
\references{
Haberman, S. (1974). \emph{The Analysis of Frequency
  Data} 1, 2. University of Chicago Press, Chicago, IL.
}
\keyword{datasets}

