% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{harmonicMean}
\alias{harmonicMean}
\title{Calculate a harmonic mean}
\usage{
harmonicMean(x, ...)
}
\arguments{
\item{x}{An invertable R object.}

\item{...}{further arguments passed to \code{\link{mean}}}
}
\description{
Calculate a harmonic mean
}
\seealso{
\code{\link{geometricMean}} and \code{\link{mean}}
}

