% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/domain.R
\name{has.domain}
\alias{has.domain}
\title{Find WWW Domains}
\usage{
has.domain(x)
}
\arguments{
\item{x}{A vector from which the domain information is desired.}
}
\description{
Hand function to return the \code{www} domain.
}
\examples{
x1 <- "http://stackoverflow.com/questions/19020749/function-to-extract-domain-name-from-url-in-r"
x2 <- "http://www.talkstats.com/"
x3 <- "www.google.com"

has.domain(x3)

sapply(list(x1, x2, x3), domain)
}
\keyword{Manipulation}

