% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\name{hasAttributes}
\alias{hasAttributes}
\title{Check if given object has certain attributes.}
\usage{
hasAttributes(obj, attributeNames)
}
\arguments{
\item{obj}{[\code{mixed}]\cr
Arbitrary R object.}

\item{attributeNames}{[\code{character}]\cr
Vector of strings, i.e., attribute names.}
}
\value{
[\code{logical(1)}]
  \code{TRUE} if object \code{x} contains all attributes from \code{attributeNames}.
  and otherwise \code{FALSE}.
}
\description{
Check if given object has certain attributes.
}

