% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{hash}
\alias{hash}
\title{Create useless hashes}
\usage{
hash(j = 50, i = 4)
}
\arguments{
\item{j}{How many columns of hashes}

\item{i}{How many rows of hashes}
}
\description{
A function to create useless hashes (####) for ease of commenting
}
\examples{
hash(j=100, i=11)
}
\author{
Dustin Fife
}

