% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{head}
\alias{head}
\alias{head.SQLiteConnection}
\title{Return the first n elements of a SQLiteConnection object}
\usage{
head(x, ...)

\method{head}{SQLiteConnection}(x, table = NULL, n = 10L, temp = FALSE,
  ...)
}
\arguments{
\item{x}{A database connection object or a table name.}

\item{table}{character specifying a table}

\item{n}{integer: Number of rows to output}

\item{temp}{logical should the function list the temp tables}

\item{\dots}{Additional arguments}
}
\description{
If a database connection is selected, returns a dataframe of table names.
If a table name is also supplied, the first n rows from the table are returned.
}

