% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dHondt.R
\docType{methods}
\encoding{latin1}
\name{highestAverages}
\alias{highestAverages}
\alias{highestAverages,ANY-method}
\title{The Highest Averages Method of Allocating Seats Proportionally}
\usage{
highestAverages(parties, votes, seats, ...)

\S4method{highestAverages}{ANY}(parties, votes, seats, method)
}
\arguments{
\item{parties}{A character vector for parties labels or candidates accordingly to the \code{votes} order.}

\item{votes}{A numeric vector for the number of formal votes received by each party or candidate.}

\item{seats}{The number of seats to be filled (scalar or vector).}

\item{method}{A character name for the method to be used. See details.}
}
\value{
A \code{data.frame}.
}
\description{
Computes the highest averages method for a variety of formulas to allocate seats proportionally for voting systems with representative assemblies.
}
\details{
The following methods are available:
\itemize{
\item {"dt"}{d'Hondt method}
\item {"sl"}{Sainte-Laguë method}
\item {"msl"}{Modified Sainte-Laguë method}
\item {"danish"}{Danish method}
\item {"imperiali"}{Imperiali (not to be confused with the Imperiali quota which is a Largest remainder method)}
\item {"hh"}{Huntington-Hill method}
}
}
\examples{
# Results for the state legislative house of Ceara (2014):

votes <- c(187906, 326841, 132531, 981096, 2043217,15061,103679,109830, 213988, 67145, 278267)

parties <- c("PCdoB", "PDT","PEN", "PMDB", "PRB","PSB","PSC", "PSTU", "PTdoB", "PTC", "PTN")

highestAverages(parties, votes, seats = 42, method="dt")

# Let's create a typical election result data.frame with the following parties and votes to return 10 seats:

my_election <- data.frame(
party=c("Yellow", "White", "Red", "Green", "Blue", "Pink"),
votes=c(47000, 16000,	15900,	12000,	6000,	3100))

highestAverages(my_election$party,
my_election$votes,
seats = 10,
method="dt")

# How this compares to the Sainte-Laguë Method

highestAverages(my_election$party,
my_election$votes,
seats = 10,
method="sl")
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{dHondt}}, \code{\link{hamilton}}, \code{\link{politicalDiversity}}.
}
\keyword{Electoral}

