% Generated by roxygen2 (4.0.2): do not edit by hand
\name{info}
\alias{info}
\title{Get Information on Data Objects}
\usage{
info(x = .data, show, ignore)
}
\arguments{
\item{x}{the data frame to be detailed.}

\item{show}{the selection of columns from \code{x}, if not all.}

\item{ignore}{columns from \code{x} to prevent of showing.}
}
\description{
Get Information on Data Objects
}
\examples{
data(titanic)
use(titanic)
## Wildcard for variables
info("C*") # Show all variables starting with 'C'
## Subset of variables
info(show = CLASS:SEX) # Same results
info(show = 1:3)
## Exclusion using wildcard.
info(ignore = "C*")
}

