% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{inv.CoxShugart}
\alias{inv.CoxShugart}
\title{Inverse Cox-Shugart Measure of Proportionality}
\usage{
inv.CoxShugart(v, s, ...)

inv.CoxShugart(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculate the inverse Cox and Shugart measure of
 proportionality based on votes and seats,
 the electoral outcome.
}
\examples{

# 2012 Queensland state elecion:
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

inv.CoxShugart(pvotes, pseats)

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\seealso{
\code{\link{CoxShugart}}, \code{\link{Farina}}, \code{\link{PoliticalDiversity}}, \code{\link{Grofman}}, \code{\link{Gallagher}},  \code{\link{Lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}.
}

