% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{inv.gallagher}
\alias{inv.gallagher}
\title{Inverse of Gallagher Index}
\usage{
inv.gallagher(v, s, ...)

inv.gallagher(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates the inverted Gallagher index or 1 - the Gallagher index.
}
\details{
The score is calculated as: 1 - sqrt(sum((v_{i}-s{i})^2)/2).
}
\examples{

# 2012 Queensland state elecion:
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

inv.gallagher(pvotes, pseats)

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\seealso{
\code{\link{cox.shugart}}, \code{\link{farina}}, \code{\link{politicalDiversity}}, \code{\link{grofman}}, \code{\link{gallagher}},  \code{\link{lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}.
}

