% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DISTRIBUTIONS.R
\encoding{UTF-8}
\name{invnormal}
\alias{invnormal}
\title{Inverse Cumulative Standard Normal Distribution}
\usage{
invnormal(area, mu = 0, sigma = 1)
}
\arguments{
\item{area}{the area or a vector of probabilities.}

\item{mu}{the mean \eqn{\mu}.}

\item{sigma}{the standard deviation of the distribution \eqn{\sigma}.}
}
\description{
Computes the inverse cumulative distribution of \code{x} associated with an \emph{area} under the normal distribution curve given by \eqn{\mu} and standard deviation  \eqn{\sigma}.
}
\examples{
invnormal(area=0.35,mu=0,sigma=1)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\code{\link{plotNorm}}, \code{\link{normalpdf}}, \code{\link{normalcdf}}
}
\keyword{Distribution}

