% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\name{is.subset}
\alias{is.subset}
\title{Check subset relation on two vectors.}
\usage{
is.subset(x, y, strict = FALSE)
}
\arguments{
\item{x}{[\code{vector}]\cr
Source vector.}

\item{y}{[\code{vector}]\cr
Vector of the same mode as \code{x}.}

\item{strict}{[\code{logical(1)}]\cr
Checks for strict/proper subset relation.}
}
\value{
[\code{logical(1)}]
  \code{TRUE} if each element of \code{x} is also contained in \code{y}, i. e.,
  if \code{x} is a subset of \code{y} and \code{FALSE} otherwise.
}
\description{
Check subset relation on two vectors.
}

