% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isid}
\alias{isid}
\title{Finds ID combination}
\usage{
isid(columns, data, verbose = TRUE)
}
\arguments{
\item{columns}{columns to combine}

\item{data}{the data object}

\item{verbose}{if \code{TRUE} messages maybe be displayed.}
}
\description{
Finds unique id combination
}
\details{
The original function was published at \url{https://stackoverflow.com/questions/16045161/unique-identifier-in-data-table}. This was berely touched.
}

