% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jackknife.R
\encoding{UTF-8}
\name{Jackknife}
\alias{Jackknife}
\alias{Jackknife.default}
\title{Resamples Data Using The Jackknife Method}
\usage{
Jackknife(x, FUN = mean, ...)

\method{Jackknife}{default}(x, FUN = mean, ...)
}
\arguments{
\item{x}{A vector}

\item{FUN}{a function name to estimate, i.e., 'mean', 'sd', 'var', 'cov', etc.}

\item{\dots}{further arguments passed to or used by other methods.}
}
\value{
est orignial estimation of parameter

jkest jackknife estimation of parameter

jkvar jackknife estimation of variance

jkbias jackknife estimate of biasness of parameter

jkbiascorr bias corrected parameter estimate
}
\description{
This function is mainly used to estimate statistics (standard errors), when the distribution is not know.
}
\examples{
x = runif(10, 0, 1)
mean(x)
Jackknife(x, mean)

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

