% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/james.stein.R
\name{james.stein}
\alias{james.stein}
\title{James-Stein shrunken estimates}
\usage{
james.stein(y, group)
}
\arguments{
\item{y}{The response variable.}

\item{group}{The grouping factor.}
}
\description{
Computes James-Stein shrunken estimates of cell means given
a response variable (which may be binary) and a grouping indicator.
}

