% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTS.R
\encoding{UTF-8}
\name{JamesStein}
\alias{JamesStein}
\title{James-Stein Shrunken Estimates}
\usage{
JamesStein(x, tau2 = 1)

JamesStein(x, tau2 = 1)
}
\arguments{
\item{x}{a vector, matrix, or array of numerics; the data.}

\item{tau2}{a positive numeric. The variance, assumed known and defaults to 1.
# @param k the grouping factor.}

\item{\dots}{extra parameters typically ignored.}
}
\description{
Computes James-Stein shrunken estimates of cell means given
a response variable (which may be binary) and a grouping indicator.
}
\details{
Missing values are by default removed.
}
\references{
Efron, Bradley and Morris, Carl (1977) ``Stein's Paradox in Statistics.'' \emph{Scientific American} Vol. 236 (5): 119-127.

James, W., & Stein, C. (1961, June).
\href{http://projecteuclid.org/euclid.bsmsp/1200512173}{Estimation with
quadratic loss}. \emph{In Proceedings of the fourth Berkeley symposium on
mathematical statistics and probability} (Vol. 1, No. 1961, pp. 361-379).
}
\keyword{Tests}

