% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TESTS.R
\encoding{UTF-8}
\name{johnson.neyman}
\alias{johnson.neyman}
\title{Johnson-Neyman Regression}
\usage{
johnson.neyman(y, x, z)
}
\arguments{
\item{y}{the dependent variable.}

\item{x}{the independent variable.}

\item{z}{the moderator variable.}
}
\description{
Probing Regression Interactions
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Tests}

