% Generated by roxygen2 (4.0.2): do not edit by hand
\name{joinLists}
\alias{joinLists}
\title{Join a list of data frames}
\usage{
joinLists(x, ...)
}
\arguments{
\item{x}{A list of data frames}

\item{\dots}{Arguments passed onto merge}
}
\description{
Recursively join data frames
}
\examples{
mtcars$cars <- row.names(mtcars)
df1 <- mtcars[,c(1:2,12)]
df2 <- mtcars[,c(3:4,12)]
df3 <- mtcars[,c(5:6,12)]
joinLists(x=list(df1, df2, df3), by="cars")
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

