% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{kalman.filter}
\alias{kalman.filter}
\title{Mehtod of Kalman Filtering}
\usage{
kalman.filter(object, ...)
}
\arguments{
\item{object}{a data object.}

\item{\dots}{Ignored parameters.}
}
\description{
Mehtod of Kalman Filtering
}

