% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kalmanPredict}
\alias{kalmanPredict}
\title{Kalman Filter Predict}
\usage{
kalmanPredict(mean1, var1, mean2, var2)
}
\arguments{
\item{mean1}{Prior mean}

\item{var1}{Prior variance}

\item{mean2}{New mean}

\item{var2}{New variance}
}
\description{
Estimates the Kalman Filtering
}
\examples{
# Equal variances but different means
kalmanPredict(10, 4, 12, 4)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Models}

