% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label.R
\encoding{UTF-8}
\name{Label<-}
\alias{Label<-}
\title{Set Variable Label}
\usage{
Label(x) <- value
}
\arguments{
\item{x}{a valid variable i.e. a non-NULL atomic vector that has no dimension attribute (see dim for details).}

\item{value}{a character value that is to be set as variable label}
}
\description{
This function sets a label to a variable, by storing a character string to its \code{label} attribute.
}
\examples{
\dontrun{
Label(religiosity$Religiosity) <- "Religiosity index"

vec <- rnorm(100)
Label(vec) <- "random numbers"
}
}
\seealso{
\code{\link{label}}
}

