% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\alias{label.data.frame}
\alias{label.default}
\alias{label<-}
\alias{label<-.data.frame}
\alias{label<-.default}
\title{Add variable label}
\usage{
label(x, default = NULL, ...)

\method{label}{default}(x, default = NULL, units = plot, plot = FALSE,
  grid = FALSE, ...)

\method{label}{data.frame}(x, default = NULL, self = FALSE, ...)

label(x, ...) <- value

\method{label}{default}(x, ...) <- value

\method{label}{data.frame}(x, self = TRUE, ...) <- value
}
\arguments{
\item{x}{A character string.}

\item{default}{if \code{x} does not have a \code{'label'} attribute and \code{default} (a character string) is specified, the label will be taken as \code{default}.  For \code{labelLatex} the \code{default} is the name of the first argument if it is a variable and not a label.}

\item{units}{Set to \code{TRUE} to append the \code{'units'} attribute (if present) to the returned label.  The \code{'units'} are surrounded by brackets.}

\item{plot}{set to \code{TRUE} to return a label suitable for \R's \code{plotmath} facility (returns an expression instead of a character string) if R is in effect.  If \code{units} is also \code{TRUE}, and if both \code{'label'} and \code{'units'} attributes are present, the \code{'units'} will appear after the label but in smaller type andwill not be surrounded by brackets.}

\item{grid}{Currently \R's \code{lattice} and \code{grid} functions do not support \code{plotmath} expressions for \code{xlab} and \code{ylab} arguments.  When using \code{lattice} functions in \R, set the argument \code{grid} to \code{TRUE} so that \code{labelPlotmath} can return an ordinary character string instead of an expression.}

\item{self}{Logical, where to interact with the object or its components.}

\item{value}{the label of the object, or "".}

\item{\dots}{a list of variables or expressions to be formed into a \code{list}. Ignored for \code{print.labelled}.}
}
\description{
\code{label(x)} retrieves the \code{label} attribute of \code{x}, while \code{label(x) <- "a label"} stores the label attribute, and also puts the class \code{labelled} as the first class of \code{x}. The \code{label} function can optionally append a \code{"units"}, usually a string or expression suitable for plotting.
}
\details{
The \code{"units"} attributes are lost upon subsetting.
}
\examples{
year <- c(2010,2012,2014,2016)
support   <- sample(60, 4)
label(year) <- "Election Years"
units(year) <- '\%'
plot(year, support, xlab=label(year))

data <- data.frame(year=year, support=support)
label(data)

label(data, self=TRUE) <- "This is a data.frame"
label(data, self=TRUE)
}

