% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\encoding{UTF-8}
\name{labelvar}
\alias{labelvar}
\title{Labels variables}
\usage{
labelvar(variable, label, data, replace = TRUE)
}
\arguments{
\item{variable}{the variable to be labeled}

\item{label}{the label, a short description text.}

\item{data}{the \code{data.frame} where \code{var} is.}

\item{replace}{is logical. If \code{TRUE}, replaces the original column with the new one with label.}
}
\description{
Labels variables
}
\examples{
data(titanic)

info(titanic)

labelvar(CLASS, "4 categories for CLASS", data = titanic)

info(titanic)
}

