% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{last.entry}
\alias{last.entry}
\title{Return only one row per ID}
\usage{
last.entry(ID, sort.var = NULL, decreasing = TRUE, data, FUN = NULL,
  fun.var = NULL, ...)
}
\arguments{
\item{ID}{The name of the unique identifier, expressed as a string (e.g., "Match.Group")}

\item{sort.var}{The variable to be sorted on in order to take the first (or last) sample, expressed as a string.}

\item{decreasing}{How should the sort.var variable be sorted? Defaults to TRUE.}

\item{data}{The dataset with multiple rows per individual}

\item{FUN}{What should be done with the multiple samples? This function can be used to extract the last
(or first) sample using the decreasing/sort.var options, or a function can be performed (such as the mean)
on one or more columns. See examples.}

\item{If}{FUN is not \code{NULL}, the variable (or a vector of variables), expressed as strings to have the function
applied to.}
}
\value{
a new dataframe containing one row per ID
}
\description{
If an individual has multiple observations in the dataset, \code{last.entry} will
loop through the entire dataset and return only one observation per individual, giving the
first (or last) draw for a person.
}
\examples{
# take only group 2 values
last.entry(ID="ID", sort.var="group", data=sleep)

# take only group 1 values
last.entry(ID="ID", sort.var="group", decreasing=FALSE,data=sleep)

# average group 1 and 2 values
last.entry(ID="ID", data=sleep, FUN=mean, fun.var="extra")

# take the maximum extra value
last.entry(ID="ID", data=sleep, FUN=max, fun.var="extra")

# take the mean of two columns extra value
sleep$group = as.numeric(as.character(sleep$group))
last.entry(ID="ID", data=sleep, FUN=mean, fun.var=c("group","extra"))

}
\author{
Daniel Marcelino
}

