% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/levy.flight.R
\name{levy.flight}
\alias{levy.flight}
\title{Simulates a Levy Walk}
\usage{
levy.flight(n = 500, alpha = 3, min.lenght = 1, max.lenght = 5,
  plot = TRUE)
}
\arguments{
\item{n}{The lenght of walk.}

\item{alpha}{The exponent of the Levy distribution.}

\item{min.lenght}{The minimum length of a step.}

\item{max.lenght}{The maximum length of a step.}

\item{plot}{A logical, TRUE will make a plot.}
}
\description{
This function simulates a Levy walk
}
\examples{
levy.flight(n=100, alpha=2)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
#'
}

