% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{lijphart}
\alias{lijphart}
\title{Lijphart index of proportionality}
\usage{
lijphart(v, s)
}
\arguments{
\item{v}{A vector containing the votes for each political party.}

\item{s}{A vector containing the election outcome as seats.}
}
\value{
A single score given the vector of votes and the vector for seats.
}
\description{
Calculates the Lijphart index of proportionality based on a vector of votes and a vector for the electoral outcome.
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}

# 2012 Quebec provincial election:
pvotes = c(PQ=31.95, Lib=31.20, CAQ=27.05,QS=6.03,Option=1.89, Other=1.88)
pseats = c(PQ=54, Lib=50, CAQ=19, QS=2, Option=0, Other=0)

lijphart(pvotes, pseats)
}

