% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TESTS.R
\name{likelihood.ratio}
\alias{likelihood.ratio}
\title{Likelihood Ratio Test (G test) for Tables}
\usage{
likelihood.ratio(x, y = NULL)
}
\arguments{
\item{x}{A vector or a matrix.}

\item{y}{A vector that is ignored if x is a matrix and required if x is a vector.}
}
\description{
Computes the likelihood ratio test (G test) for contingency tables. Currently does not do Williams' and Yates' correction.
}
\examples{
if (interactive()) {
# 1992 General Social Survey- Sex and Party affiliation
gss <- data.frame(
   expand.grid(sex=c("female", "male"),
   party=c("dem", "indep", "rep")),
   count=c(279,165,73,47,225,191))
# expand it:
 GSS <- gss[rep(1:nrow(gss), gss[["count"]]), ]

likelihood.ratio(GSS$party, GSS$sex)
}
}
\references{
Agresti, Alan (1996) \emph{Introduction to categorical data
 analysis}. NY: John Wiley and Sons

Smithson, M.J. (2003) \emph{Confidence Intervals, Quantitative
 Applications in the Social Sciences Series, No. 140}.
 Thousand Oaks, CA: Sage. pp. 39-41.
}

