% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/line.plot.R
\name{line.plot}
\alias{line.plot}
\title{Plot a lineplot}
\usage{
line.plot(data, xvar, yvar, group.var = NULL, add.point = FALSE,
  point.size = 1.5, point.shape = 19, point.fill = NULL,
  point.color = "black", arrow = NULL, x.type = c("categorical",
  "continuous"), group.color = NULL, palette = NULL, ...)
}
\arguments{
\item{data}{The data frame.}

\item{xvar}{The name of column containing x variable. Default value is \code{NULL}.}

\item{yvar}{The name of column containing y variable.}

\item{group.var}{The name of column containing group variable. This variable is used to color plot according to the group.}

\item{add.point}{if \code{TRUE}, points are added to the plot. Default value is \code{FALSE}.}

\item{point.size}{To modify the point size.}

\item{point.shape}{To modify the point shape.}

\item{point.fill}{To modify the point fill option.}

\item{point.color}{To modify the color of the points.}

\item{arrow}{To add arrow to a line, usage: arrow=arrow(). See \code{?grid::arrow} for more details. Default value is \code{NULL}.}

\item{x.type}{Indicate the type of x-axis, either \code{"categorical"} or  \code{"continuous"}. Default value is \code{"categorical"}. When the variable on the x-axis is numeric, it is sometimes useful to treat it as continuous, and sometimes useful to treat it as categorical.}

\item{group.color}{Color of groups. \code{group.color} should have the same length as groups.}

\item{palette}{this can be also used to indicate group colors. In this case the parameter group.color should be \code{NULL}.}

\item{\dots}{Other parameters passed on to ggplot2.customize function.}
}
\description{
Handy function for making a line plot with R package ggplot.
}

