% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{list_parties}
\alias{list_parties}
\title{List Political Parties}
\usage{
list_parties()
}
\description{
Returns a list of available political parties associated with a color.
}
\note{
Warning: the list can be huge!\cr
}

