% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DISTRIBUTIONS.R
\encoding{UTF-8}
\name{logikDistribution}
\alias{logikDistribution}
\title{Calculate the Log Likelihood of a Normal Distribution}
\usage{
logikDistribution(x = data, mu, var)
}
\arguments{
\item{x}{data.}

\item{mu}{estimated mean.}

\item{var}{estimated variance.}
}
\value{
ll logliklihood of the distribution
}
\description{
Find the log likelihood of a normal distribution.
}
\examples{
x = rnorm(100, 3, 7)
logikDistribution(x,3,7)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

