% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lookup}
\alias{lookup}
\title{Lookup}
\usage{
lookup(x, lookup.array)
}
\arguments{
\item{x}{the variable}

\item{lookup.array}{a n-by-2 array used for looking up.}
}
\description{
Recodes values of a vector from a lookup array.
}
\examples{
a <- c( 1, 1, 1, NA, NA, NA)
 b <-cbind(c(1, NA), c(NA, 99))
 lookup( a,b)
}

