% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{LookVariable}
\alias{LookVariable}
\title{Look for Variables in a Data Frame Using Regular Expressions}
\usage{
LookVariable(.data, varnames)
}
\arguments{
\item{.data}{a data.frame object.}

\item{varnames}{a character or vector with regular expressions of variables names.}
}
\value{
Returns the variables names of the variables found.
}
\description{
The function look for variable names in a \code{data.frame}.
}
\examples{
LookVariable(titanic, "clas")

}

