% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{LoosemoreHanby}
\alias{LoosemoreHanby}
\title{Loosemore-Hanby Index of Disproportionality}
\usage{
LoosemoreHanby(v, s, ...)

LoosemoreHanby(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates the Loosemore-Hanby index of disproportionality.
}
\details{
The score is calculated as the sum of the absolute differences \code{|v-s|} divided by two. This index is also known as the Pedersen index, with common usage in studies of volatility.
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Duncan, O. and Duncan, B. (1955) A methodological analysis of segregation indexes. \emph{American Sociological Review} 20:210-7.

Loosemore, J. and wHanby, V. (1971) The theoretical limits of maximum distortion: Som analytical expressions for electoral systems. \emph{British Journal of Political Science} 1:467-77.

 @examples
# 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

LoosemoreHanby(pvotes, pseats)
}
\seealso{
\code{\link{Rose}}, \code{\link{Rae}}, \code{\link{CoxShugart}}, \code{\link{PoliticalDiversity}}, \code{\link{Grofman}}, \code{\link{Farina}},  \code{\link{Lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}
}

