\name{ltaylor96}
\alias{ltaylor96}
\docType{data}
\title{Lothian's and Taylor's (1996) Data Set
}
\description{Data used by Lothian and Taylor (1996) on the real exchange rate behaviour.
}
\usage{data(ltaylor96)}
\format{
  A data frame with 200 observations on the following 4 variables.
  \describe{
    \item{\code{year}}{Year}
    \item{\code{spot}}{dollar/sterling exchange rate}
    \item{\code{USwpi}}{U.S. wholesale price index, 1914==100}
    \item{\code{UKwpi}}{U.K. wholesale price index, 1914==100}
 }
}
\source{Hayashi, F. (2000) \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}

}
\references{Lothian, J. R., and Taylor, M. P. (1996) Real exchange rate behavior: the recent float from the perspective of the past two centuries. \emph{Journal of Political Economy,} 488--509.

}
\examples{
data(ltaylor96)
}
\keyword{datasets}
