% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Create a new Chart}
\usage{
make(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{\code{data.frame}-like or \code{xts} object}

\item{width,height}{Must be a valid unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number.}
}
\description{
Performs basic data setup for use with the other \code{ggplot} functions.
}

