% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{marriage}
\alias{marriage}
\title{Same Sex Marriage Public Opinion Data}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::marriage)} variables and \Sexpr{nrow(SciencesPo::marriage)} observations.}
\usage{
data(marriage)
}
\description{
Data set fielded by the PEW Research Center on same sex marriage support in US. It covers public opinion on the issue starting from 1996 up to date. This dataset contains the following columns:

\itemize{
  \item Date The year of the measurement
  \item Oppose Percent opposing same-sex marriage
  \item Favor Percent favoring same-sex marriage
  \item DK Percent of Don't Know
}
}
\references{
PEW Research Center. \emph{Support for same-sex marriage}. \url{http://www.pewresearch.org}
}
\keyword{datasets}

