% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matching.R
\encoding{UTF-8}
\name{matching}
\alias{matching}
\title{Matching}
\usage{
matching(z, score, replace = FALSE)
}
\arguments{
\item{z}{is the vector of indicators for treatment or control}

\item{score}{is the vector of the propensity scores in the same order as z}

\item{replace}{a logical value for \code{replace}, default is \code{replace=FALSE}.}
}
\value{
the function returns a vector of indices that the corresponding  unit is matched to. 0 means matched to nothing. now also returns a number for each pair making it easier to later recover those pairs.
}
\description{
Matching
}

