% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/maxGaussian.R
\name{maxGaussian}
\alias{maxGaussian}
\title{Maximize Gaussian}
\usage{
maxGaussian(mu, sigma2, x)
}
\arguments{
\item{mu}{The mean component}

\item{sigma2}{The variance component}

\item{x}{Is the state we know nothing about.}
}
\value{
The value that maximize Gaussian
}
\description{
This function Maximizes a Gaussian Distribution
}
\details{
We can set x = mu, so that we get the peak of the Gaussian  as the peak always occurs at the mean (looking at a Gaussian distribution). This is why we set x= when we want to maximize.
}
\examples{
maxGaussian(10, 4, 10)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Udacity course on Kalman Filter
}

