% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.table.R
\name{mcmc.table}
\alias{mcmc.table}
\title{Summarize MCMC output in a table-like}
\usage{
mcmc.table(obj, ci = 0.95)
}
\arguments{
\item{obj}{The MCMC object.}

\item{ci}{The level for the confidence interval.}
}
\description{
The function produces a table with means, SDs, and 95% credible intervals
from MCMC output from the following packages: \code{R2jags}, \code{rjags}, \code{R2WinBUGS}, \code{R2OpenBUGS}, and \code{MCMCpack}.
}

