% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means.plot.R
\name{means.plot}
\alias{means.plot}
\title{Mean plot}
\usage{
means.plot(.data, xvar, yvar, group.var = NULL, plot.type = c("lineplot",
  "barplot"), error.bars = c("se", "sd", "conf.int", "none"), level = 0.95,
  bar.width = 0.2, bar.color = NULL, verbose = TRUE,
  paired.sample = FALSE, id = NULL, position = position_dodge(0.9), ...)
}
\arguments{
\item{.data}{The data frame.}

\item{xvar}{The name of column containing x variable. Default value is NULL.}

\item{yvar}{The name of column containing y variable.}

\item{group.var}{The name of column containing group variable. This variable is used to color plot according to the group.}

\item{plot.type}{The graph type (line.plot or bar.plot).}

\item{error.bars}{Type of the error bars (standard error, standard deviation, confidence interval)}

\item{level}{The confidence level (error.bars = conf.int can be applied).}

\item{bar.width}{The width of the bars.}

\item{bar.color}{The colour of the bar.}

\item{verbose}{If \code{TRUE}, a table with means is shown.}

\item{paired.sample}{A logical in case the sample groups are paried, default is \code{FALSE}.}

\item{id}{The column containing the unique identification of the observations/subjects.Must be provided when \code{paired.sample=TRUE}.}

\item{position}{To adjust the error bar position in the barplot.}
}
\description{
Mean plot with ggplot2
}

