% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanSquaredError.R
\name{meanSquaredError}
\alias{meanSquaredError}
\title{Quadratic Error}
\usage{
meanSquaredError(observed, estimate)
}
\arguments{
\item{observed}{is the vector of observed values.}

\item{estimate}{is the vector of n predictions.}
}
\description{
Handy function to estimate the mean squared error (MSE) of an estimator. It measures the average of the squares of the ``errors'', that is, the difference between the estimator and what is estimated.
}

