% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{mishkin92}
\alias{mishkin92}
\title{Mishkin's (1992) Data}
\format{A data frame with 491 observations on the following 8 variables.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(mishkin92)
}
\description{
Data from the Frederic S. Mishkin (1992) paper \dQuote{Is the Fisher Effect for real?}.

\itemize{
 \item year Year
\item mon a numeric vector
\item inf1mo a numeric vector
\item inf3mo a numeric vector
\item tbill1mo a numeric vector
\item tbill3mo a numeric vector
\item cpiu a numeric vector
\item quote a numeric vector
}
}
\references{
Mishkin, F. S. (1992) Is the Fisher effect for real?: A reexamination of the relationship between inflation and interest rates. \emph{Journal of Monetary Economics,} \bold{30(2),} 195--215.
}
\keyword{datasets}

