% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{modalValue}
\alias{modalValue}
\title{Compute the Modal Value (Mode)}
\usage{
modalValue(x, na.rm = FALSE)
}
\arguments{
\item{x}{the object or variable.}

\item{na.rm}{a logical. Should NA values be removed?}
}
\description{
Compute the statistical mode.
}
\examples{

modalValue(mtcars$mpg)

}

