% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mode}
\alias{mode}
\title{Calculate the Mode}
\usage{
mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A data vector}

\item{na.rm}{A logical value, default is \code{FALSE}}
}
\description{
Estimates the mode for a vector
}
\examples{
myvar <-c(1,1,2,2,3,3,4,4,5, NA)
mode(myvar)

mode(myvar, FALSE)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

