% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modify}
\alias{modify}
\title{Modify data elements}
\usage{
modify(x, position, value)
}
\arguments{
\item{x}{A data object}

\item{position}{The position of the element to be replaced}

\item{value}{The value to modify}
}
\description{
Modify an element in a vector, taking its position as reference.
}
\examples{
x <- seq(1:10)

modify(x, 1, 10)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Data}
\keyword{Manipulation}

