% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{move}
\alias{move}
\title{Move variable order in a data.frame}
\usage{
move(data, var, where)
}
\arguments{
\item{data}{The data object}

\item{var}{The variable to be moved.}

\item{where}{The location where the variable (var) is to be relocated.}
}
\description{
Move variable order in a data.frame
}
\examples{
move(ssex, "DK", "Oppose")
}

