% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{myTab}
\alias{myTab}
\title{\pkg{Rcpp} table function is faster than \code{\link[base:table]{base::table}} R function.}
\usage{
myTab(x)
}
\arguments{
\item{x}{}
}
\description{
Rcpp table function is faster than base R (3x speedup).
}

