% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/na.count.R
\name{na.count}
\alias{na.count}
\title{Count NA values}
\usage{
na.count(.data)
}
\arguments{
\item{.data}{The data.frame.}
}
\description{
The function counts NAs for each column of a \code{data.frame}.
}
\examples{
na.count(ssex)
}

