% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq2.R
\name{NAasMissing}
\alias{NAasMissing}
\title{NA to Missing}
\usage{
NAasMissing(x, convert.to.factor = FALSE, missing.value = "MISSING")
}
\arguments{
\item{x}{is a data.frame of factors}

\item{convert.to.factor}{if TRUE all non-factor columns in x are converted to factors}

\item{missing.value}{is the new value given to the NA values}
}
\description{
Recode all NA values into missing in a data.frame of factors
}
\examples{
x <- as.data.frame(matrix(rep(c("a", "b", "c", NA), 3), ncol=3))
x
NAasMissing(x)
}

