% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/name.slicer.R
\encoding{UTF-8}
\name{name.slicer}
\alias{name.slicer}
\title{Splits name fields}
\usage{
name.slicer(name, data = .data)
}
\arguments{
\item{name}{the name field column.}

\item{data}{the data.frame name.}
}
\value{
two columns or a list.
}
\description{
Splits a name field variable allocating the first and last names into two new columns or a list.
}
\details{
The way one may split names is region dependent, so this function may apply to very few contexts. See for instance \url{http://www.w3.org/International/questions/qa-personal-names}
}
\examples{
df <- data.frame( name = c("Martin Luther King",
 "Nelson Mandela", "Simon Bolivar") )
 name.slicer(df$name)
 df$n<- name.slicer(df$name)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\link{unnest}.
}

