% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{nameSplit}
\alias{nameSplit}
\title{Splits name field variable}
\usage{
nameSplit(name, data = .data)
}
\arguments{
\item{name}{the name field column.}

\item{data}{the data.frame name.}
}
\value{
two columns or a list.
}
\description{
Splits a name field variable allocating the first and last names into two new columns or a list.
}
\details{
The way one may split a name is region dependent, so this function may only apply to very few contexts. See for instance \url{http://www.w3.org/International/questions/qa-personal-names}
}
\examples{
df <- data.frame( name = c("Martin Luther King", "Nelson Mandela", "Simon Bolivar") )
nameSplit(df$name)
df$n<- nameSplit(df$name)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\link{unnest}.
}

