% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\encoding{UTF-8}
\name{NAto0}
\alias{NAto0}
\title{Changes NAs in a vector into a given value}
\usage{
NAto0(x, value = 0)
}
\arguments{
\item{x}{the vector}

\item{value}{the value to be given to the missing value \code{NA}.}
}
\description{
Changes NAs in a vector into a given value
}
\examples{
v <- sample(c(round(runif(5, 1, 3)), rep(NA, 2)))
NAto0(v)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\link{rand.imput}.
}

