% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nearest.loc}
\alias{nearest.loc}
\title{Find Location of Nearest Value}
\usage{
nearest.loc(x, value)
}
\arguments{
\item{x}{A vector.}

\item{value}{The value that you want to find.}
}
\description{
Find the location of the nearest value to a number that you specify.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

