% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{nestedTable}
\alias{nestedTable}
\title{Cross-tabulation}
\usage{
nestedTable(x, y, z, chisq = FALSE, ...)
}
\arguments{
\item{x}{The row variable.}

\item{y}{The column variable.}

\item{z}{The stratum variable.}

\item{chisq}{A logical. If \code{TRUE} the results of a chi-square are shown.}

\item{\dots}{Additional arguements (currently ignored)
#}
}
\description{
Produces a nested crosstable.
}
\examples{
# with(titanic, nestedTable(SEX, SURVIVED, CLASS))
}

