% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Find Matching (or Non-Matching) Elements}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{A vector (numeric, character, factor).}

\item{y}{A vector (numeric, character, factor), matching the mode of \code{x}.
\code{\link{match}}, \code{\link{\%in\%}}.}
}
\description{
\code{\%nin\%} is a binary operator, which returns a logical vector indicating if there is a match or not for its left operand. A true vector element indicates no match in left operand, false indicates a match.
}
\examples{
c('a','b','c') \%nin\% c('a','b')
}
\keyword{Manipulation}

