% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{no_gridlines}
\alias{no_gridlines}
\title{Remove gridlines}
\usage{
no_gridlines()
}
\description{
Remove all major and minor
gridlines from a ggplot2 plot or theme.
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}
\keyword{ggplot2}

