% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{no_minor_gridlines}
\alias{no_minor_gridlines}
\title{Remove minor gridlines}
\usage{
no_minor_gridlines()
}
\description{
Remove all minor
gridlines from a ggplot2 plot or theme.
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}

