% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{no_x_gridlines}
\alias{no_x_gridlines}
\title{Remove vertical gridlines.}
\usage{
no_x_gridlines()
}
\description{
Remove all major and minor vertical
gridlines from a ggplot2 plot or theme.
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}
\keyword{ggplot2}

