% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{nobs}
\alias{nobs}
\alias{nobs,data.frame,ANY-method}
\alias{nobs,data.frame-method}
\alias{nobs,integer,ANY-method}
\alias{nobs,integer-method}
\alias{nobs,matrix,ANY-method}
\alias{nobs,matrix-method}
\alias{nobs,numeric,ANY-method}
\alias{nobs,numeric-method}
\title{Method for the number of observations}
\usage{
nobs(object, ...)

\S4method{nobs}{numeric}(object, ...)

\S4method{nobs}{integer}(object, ...)

\S4method{nobs}{matrix}(object, ...)

\S4method{nobs}{data.frame}(object, ...)
}
\arguments{
\item{object}{the data object}

\item{\dots}{some extra parameters.}
}
\value{
The number of observations.
}
\description{
Method for the number of observations
}
\examples{
nobs(1:50)
nobs(10)
}

