% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{nobs}
\alias{nobs}
\title{nobs method}
\usage{
nobs(object, ...)
}
\arguments{
\item{object}{the input object.}

\item{\dots}{Some extra arguments.}
}
\description{
nobs method
}
\note{
nobs generic is defined only in R 2.13.0: define here for older R.
}

