% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normVarsNames.R
\name{normalizeVariableNames}
\alias{normalizeVariableNames}
\title{Normalize Variable Names}
\usage{
normalizeVariableNames(x, sep = "_")
}
\arguments{
\item{x}{the data frame.}

\item{sep}{the separator character.}
}
\description{
(1) Replace any all capitals words to lowercase letter pattern.
(2) Remove any resulting initial or trailing underscore or multiples.
(3) Replace any number sequences not preceded by an
}
\examples{
names(iris) \%<>\% normalizeVariableNames() \%>\% print()

}

