% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalTransform.R
\name{normalTransform}
\alias{normalTransform}
\title{Normal transformation}
\usage{
normalTransform(dat)
}
\arguments{
\item{dat}{vector of values to be normal transformed}
}
\value{
vector of transformed values, with attributes required for the backtransform
}
\description{
Normal transformation
}

