% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{nperm}
\alias{nperm}
\title{Calculate number of permutations taking repeated elements into consideration}
\usage{
nperm(vec)
}
\arguments{
\item{vec}{The vector which number of permutations will be calculated}
}
\description{
Calculate number of permutations taking repeated elements into consideration
}

