% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutate.R
\encoding{UTF-8}
\name{nPermutate}
\alias{nPermutate}
\title{Calculate number of permutations taking repeated elements into consideration}
\usage{
nPermutate(vec)
}
\arguments{
\item{vec}{The vector which number of permutations will be calculated}
}
\description{
Calculate number of permutations taking repeated elements into consideration
}
\examples{
myvar <- 1:10
nPermutate(myvar)
}

