% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/odds.ratio.R
\name{odds.ratio}
\alias{odds.ratio}
\title{Odds Ratio}
\usage{
odds.ratio(x, ...)
}
\arguments{
\item{x}{object from whom odds ratio will be computed}

\item{...}{further arguments passed to or from other methods}
}
\description{
S3 method for odds ratio
}

