\encoding{UTF-8}
\name{data.clean}
\alias{data.clean}

\title{Clean marks
}
\description{Strips marks from data  
}
\usage{
data.clean(data)
}
\arguments{
  \item{data}{A data object
  }
}
\value{A data object as the input
}
\author{
Daniel Marcelino
}
\examples{
\dontrun{data.clean("dirt_*.*_text+/\+-")}
}
\keyword{ data manipulation }
