% Generated by roxygen2 (4.0.2): do not edit by hand
\name{omegaFactorial}
\alias{omegaFactorial}
\title{Calculates the effect size}
\usage{
omegaFactorial(n, a, b, SSa, SSb, SSab, SSr)
}
\arguments{
\item{n}{}

\item{a}{}

\item{b}{}

\item{SSa}{}

\item{SSb}{}

\item{SSab}{}

\item{SSr}{}
}
\description{
Computes the effect size
}
\examples{
omegaFactorial(8,2,3,169,3332,1978,3488)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

