\name{on2011}
\alias{on2011}
\docType{data}
\title{2011 Ontario Election Results
}
\description{Official results of the 2011 Ontario election as available on Elections Ontario
}
\usage{data(on2011)}
\format{
 A data frame with 655 observations on the following 4 variables.
  \describe{
    \item{\code{partynames}}{Party names}
    \item{\code{riding}}{Name of the riding}
    \item{\code{party}}{Acronym version of the party names}
    \item{\code{votes}}{Quantity of votes}
  }
}

\source{\url{http://www.elections.on.ca/en-ca}
}
\references{
Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}
\examples{
data(on2011)
}
\keyword{datasets}
