% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\encoding{UTF-8}
\name{Operations}
\alias{Max}
\alias{Mean}
\alias{Min}
\alias{Mode}
\alias{Operations}
\alias{SD}
\alias{SE}
\alias{Sum}
\alias{Sums}
\title{Arithmetic and Statistiscal Functions but Dealing with Missing Data}
\usage{
Mean(x)

SD(x)

SE(x)

Sum(x)

Sums(x)

Max(x)

Min(x)

Mode(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Arithmetic and statistiscal functions but dealing with missing data by default.
}
\examples{
x <- c(NA,1,2,3,2,1,NA)

Mean(x)

Sum(x)

Mode(x)

sd(x, na.rm=TRUE)/sqrt(length(x))

SE(x)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

