% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\encoding{UTF-8}
\name{orcalc}
\alias{orcalc}
\title{Odds Ratio Calculation}
\usage{
orcalc(p1, p2)
}
\arguments{
\item{p1}{is a probability for group 1.}

\item{p2}{is another probability, for group 2.}
}
\description{
The orcalc can be used to obtain odds ratios. Simply provide the two probabilities to be used (the probability of success for group 1 is given first, then the probability of success for group 2)
}
\examples{
orcalc(.3, .4)

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

