% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{orderedAxis}
\alias{orderedAxis}
\title{Interactively order columns in ggplot}
\usage{
orderedAxis(data, axis, column)
}
\arguments{
\item{data}{data frame}

\item{axis}{categorical axis}

\item{column}{numerical variable by which to order}
}
\description{
Order a categorical variable by the values in a numerical variable in ggplot
}
\examples{
ggplot(orderedAxis(mtcars, gear, carb),
       aes(x = gear.o, y = carb))

}

