% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{other.names}
\alias{other.names}
\title{Extracts names from a dataset other than the ones indicates}
\usage{
other.names(data, check)
}
\arguments{
\item{data}{the input \code{data.frame}.}

\item{check}{The \code{names} you want to check.}
}
\value{
A character vector of the remaining names.
}
\description{
Extracts names from a dataset other than the ones indicates
}
\examples{
\dontrun{ other.names(iris, "Species")}
}

