% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{pack}
\alias{pack}
\title{Pack all related variables into the existing .data.}
\usage{
pack(data = .data)
}
\arguments{
\item{data}{The data object.}
}
\description{
Tries to  wrap up all related variables (variable with the same length) into the existing `.data` object.
}
\examples{
df=data.frame(group=sample(letters,10, TRUE),y=sample(10) )
use(df)
y2 <- y^2
pack(df)
df;
}

