% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\name{package}
\alias{package}
\title{Load packages}
\usage{
package(x)
}
\arguments{
\item{x}{The package name, or a list of packages.}
}
\description{
Check if package exists and install it otherwise.
}
\examples{
\dontrun{ package(qvalue) # from bioconductor
 package(SoundexBR) # from CRAN }
}

