% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{paired}
\alias{paired}
\title{US General Social Survey 1993}
\format{A data frame containing 1500 observations on 96 variables.}
\usage{
data(gss93)

data(paired)
}
\description{
A data frame containing 1500 observations on 96 variables.

\itemize{
\item ID The respondentid.
\item WRKSTAT Working status.
 \item MARITAL Marital Status.
 \item PARTYID Respondent's party id.
 \item VOTE92 Whether the respondent have voted in the preceeding year.
 \item SEX The sex of the respondent.
 }

Artificial experimental paired data

\itemize{
\item patient the patient id.
\item before_X before treatment.
\item after_Y after treatment.
 }
}
\keyword{datasets}

