% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{pal}
\alias{pal}
\title{Convenience function to check color labels}
\usage{
pal(col, border = "light gray")
}
\arguments{
\item{col}{the color name.}
}
\description{
Convenience function to check color labels
}

