% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette_info}
\alias{palette_info}
\title{Show details of specific palette}
\usage{
palette_info(palette.name)
}
\arguments{
\item{palette.name}{the name of the palette}
}
\description{
Show details of specific palette
}
\examples{
palette_info('colorblind')
}

