% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palletes.R
\docType{data}
\name{Palletes}
\alias{Palletes}
\title{Palette data for the themes used by package}
\format{A \code{list}.}
\usage{
Palletes
}
\description{
Data used by the palettes in the package.
}
\keyword{datasets}

