% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{parties_color_pal}
\alias{parties_color_pal}
\title{Color Palettes for Political Organizations (discrete)}
\usage{
parties_color_pal(palette = "BRA")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes for political organizations.
}
\examples{
scales::show_col(parties_color_pal()(10))
}
\seealso{
Other colour parties: \code{\link{scale_color_parties}},
  \code{\link{scale_colour_parties}},
  \code{\link{scale_fill_parties}}
}

