% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{party_color_pal}
\alias{party_color_pal}
\title{Political Parties Color Palette (Discrete) and Scales}
\usage{
party_color_pal(palette = "BRA", hex = FALSE)
}
\arguments{
\item{palette}{the palette name, a character string.}

\item{hex}{logical, if \code{FALSE}, the associated color name (label) is returned.}
}
\description{
An N-color discrete palette for political parties.
}
\examples{
party_color_pal("BRA")

# Argentine
party_color_pal("ARG")

# US
party_color_pal("USA")

# Canada
party_color_pal("CAN")

}
\seealso{
Other color party: \code{\link{scale_color_party}}
}

