% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pause.R
\encoding{UTF-8}
\name{pause}
\alias{pause}
\title{Pause}
\usage{
pause(x = 0)
}
\arguments{
\item{x}{is optional. If x>0 a call is made to \code{\link{Sys.sleep}}. Else, execution pauses until a key is entered.}
}
\description{
A replication of MatLab pause function.
}

