% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{peek}
\alias{peek}
\title{Show Observations Randomly Drawn from Data Objects}
\usage{
peek(x = .data, n = 10)
}
\arguments{
\item{x}{A matrix or data.frame object}

\item{n}{The number of rows to be shown}
}
\description{
Provide a sly view of the data by randomly draw observations, instead of showing only the first \code{head()} or the last \code{tail()} rows of an object.
}
\examples{
use(titanic)
peek(titanic)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Tables}

