% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/percentify.R
\encoding{UTF-8}
\name{percentify}
\alias{percentify}
\title{Converts to percentiles}
\usage{
percentify(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Converts a numeric vector to percentiles.
}
\examples{
vec <- seq(1:5)
percentify(vec)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

