% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutate.R
\encoding{UTF-8}
\name{Permutate}
\alias{Permutate}
\title{Create k random permutations of a vector}
\usage{
Permutate(x, k)
}
\arguments{
\item{x}{a vector to be permutated.}

\item{k}{number of permutations to be conducted.}
}
\description{
Creates a k random permutation of a vector.
}
\details{
should be used only for length(input)! >> k
}
\examples{
#  5! = 5 x 4 x 3 x 2 x 1 = 120
# row wise permutations
Permutate(x=1:5, k=5)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\keyword{Sampling}

