% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{permutateSample}
\alias{permutateSample}
\title{Conditional replacement}
\usage{
permutateSample(input, k)
}
\arguments{
\item{input}{vector to be permutated.}

\item{k}{number of permutations.}

\item{.data}{The data object.}
}
\description{
Conditional replacement
}
\details{
should be used only for length(input)! >> k
}
\examples{
permutateSample(input=1:5, k=5)
}
\keyword{Sampling}

