% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{\%>\%}
\alias{\%<>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs

lhs \%<>\% rhs
}
\description{
See \code{\link[magrittr]{\%>\%}} for more details.

See \code{\link[magrittr]{\%<>\%}} for more details.
}
\keyword{internal}

