% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PIPING.R
\name{pipe_dot}
\alias{pipe_dot}
\title{pipe to dot
. : object
expr : expression
envir : environment for evaluation}
\usage{
pipe_dot(., expr, envir)
}
\description{
pipe to dot
. : object
expr : expression
envir : environment for evaluation
}

