% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PIPING.R
\name{pipe_first}
\alias{pipe_first}
\title{pipe to first argument
x : object
fun : the function name or call
envir : environment for evaluation}
\usage{
pipe_first(x, fun, envir)
}
\description{
pipe to first argument
x : object
fun : the function name or call
envir : environment for evaluation
}

