% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/piping.R
\name{pipe_lambda}
\alias{pipe_lambda}
\title{pipe by lambda expression
x : object
expr : lambda expression
envir : environment for evaluation
side_effect: TRUE to return x; FALSE to return value of expr}
\usage{
pipe_lambda(x, expr, envir, side_effect = TRUE)
}
\description{
pipe by lambda expression
x : object
expr : lambda expression
envir : environment for evaluation
side_effect: TRUE to return x; FALSE to return value of expr
}

