% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/piping.R
\name{pipe_op}
\alias{pipe_op}
\title{pipe function that determines the piping mechanism for the expression
x : object
expr : function name, call, or enclosed expression}
\usage{
pipe_op(x, expr)
}
\description{
pipe function that determines the piping mechanism for the expression
x : object
expr : function name, call, or enclosed expression
}

