% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotEllipses.R
\name{plotEllipses}
\alias{plotEllipses}
\title{Plots ellipses}
\usage{
plotEllipses(x, y, r, ...)
}
\arguments{
\item{x}{The longitude coordinate.}

\item{y}{The latitude coordinate.}

\item{r}{The radio.}

\item{\dots}{Extra parameters to plot, such as col, lty etc.}
}
\description{
Plots ellipses given the coordinates and radio.
}
\examples{
plot(-60:-30, -30:0)
plotEllipses(-49,-25,7)
}

