% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{Plotting}
\alias{Plotting}
\alias{PreviewTheme}
\alias{Previewplot}
\alias{align_title_left}
\alias{align_title_right}
\alias{legend_bottom}
\alias{legend_left}
\alias{legend_right}
\alias{legend_top}
\alias{move_legend}
\alias{no_axes}
\alias{no_axes_titles}
\alias{no_gridlines}
\alias{no_legend}
\alias{no_legend_title}
\alias{no_major_gridlines}
\alias{no_major_x_gridlines}
\alias{no_major_y_gridlines}
\alias{no_minor_gridlines}
\alias{no_minor_x_gridlines}
\alias{no_minor_y_gridlines}
\alias{no_ticks}
\alias{no_title}
\alias{no_x_axis}
\alias{no_x_gridlines}
\alias{no_x_line}
\alias{no_x_text}
\alias{no_x_ticks}
\alias{no_x_title}
\alias{no_y_axis}
\alias{no_y_gridlines}
\alias{no_y_line}
\alias{no_y_text}
\alias{no_y_ticks}
\alias{no_y_title}
\alias{rotate_axes_text}
\alias{rotate_x_text}
\alias{rotate_y_text}
\title{Plot Customizing}
\usage{
Previewplot()

PreviewTheme()

align_title_left()

align_title_right()

no_title()

no_y_gridlines()

no_x_gridlines()

no_major_x_gridlines()

no_major_y_gridlines()

no_major_gridlines()

no_minor_x_gridlines()

no_minor_y_gridlines()

no_minor_gridlines()

no_gridlines()

rotate_axes_text(angle)

rotate_x_text(angle)

rotate_y_text(angle)

no_axes()

no_x_axis()

no_y_axis()

no_x_line()

no_y_line()

no_x_text()

no_y_text()

no_ticks()

no_x_ticks()

no_y_ticks()

no_axes_titles()

no_x_title()

no_y_title()

move_legend(position)

legend_bottom()

legend_top()

legend_left()

legend_right()

no_legend()

no_legend_title()
}
\arguments{
\item{angle}{the angle of rotation.}

\item{position}{the location of the legend ('top', 'bottom', 'right', 'left' or 'none').}
}
\description{
Several wrapper functions to deal with ggplot2.
}
\examples{
Previewplot()
}
\author{
\Sexpr[results=rd, stage=build]{tools:::Rd_package_author("SciencesPo")}
}
\keyword{ggplot2}

