% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{politics}
\alias{politics}
\title{Politics by Personality}
\format{A 2x2 (contingency) table}
\usage{
data(politics)
}
\description{
A 2-by-2 contingency table comparing political identification and personality.
}
\author{
David Kahle, simplifying the dataset found on p.622 of Sheskin (see References)
}
\references{
Sheskin, D. J. Handbook of Parametric and Nonparametric Statistical Procedures. 4ed. Chapman and Hall/CRC Press, 2007.
}
\keyword{datasets}

