% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{Pollster08}
\alias{Pollster08}
\title{Polls for 2008 U.S. presidential election}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::Pollster08)} variables and \Sexpr{nrow(SciencesPo::Pollster08)} observations.
\itemize{
\item PollTaker Polling organization.
\item PollDates Dates the poll data were collected.
\item MidDate Midpoint of the polling period.
\item Days Number of days after August 28th (end of Democratic convention).
\item n Sample size for the poll.
\item Pop \code{A}=all, \code{LV}=likely voters, \code{RV}=registered voters.
\item McCain Percent supporting John McCain.
\item Obama Percent supporting Barak Obama.
\item Margin Obama percent minus McCain percent.
\item Charlie Indicator for polls after Charlie Gibson interview with VP candidate Sarah Palin (9/11).
\item Meltdown Indicator for polls after Lehman Brothers bankruptcy (9/15).
}
 @source
 \url{http://www.pollster.com}}
\usage{
data(Pollster08)
}
\description{
Polls for 2008 U.S. presidential election. hese data include
all presidential polls reported on the internet site pollster.com
that were taken between August 29th, when John Mc-Cain announced
that Sarah Palin would be his running mate as the Republican nominee for vice
president, and the end of September.
}
\keyword{datasets}

