% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{posixify}
\alias{posixify}
\title{Converts calendar date string to POSIX}
\usage{
posixify(x)
}
\arguments{
\item{x}{character vector in one of two calendar date formats}
}
\value{
a POSIX date
}
\description{
Converts calendar date string to POSIX
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

