% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{presidents}
\alias{presidents}
\title{The Measure of US Presidents}
\format{This data.frame contains 57 rows in the following 11 variables.}
\source{
Inside Gov. \url{http://us-presidents.insidegov.com}.
Wikipedia: \url{https://en.wikipedia.org/wiki/United_States_presidential_election,_2012}
Wikipedia:\url{https://en.wikipedia.org/wiki/Heights_of_presidents_and_presidential_candidates_of_the_United_States}.
}
\usage{
data(presidents)
}
\description{
The US presidents and their main opponents' heights (cm).

\itemize{
\item election The election year.
\item winner The winner candidate.
\item winner.height The winner candidate's height in centimeters (cm).
\item winner.vote The popular vote support for the winner.
\item winner.party The winner's party.
\item opponent The main opponent candidate.
\item opponent.height The opponent candidate's height in centimeters (cm).
\item opponent.vote Popular vote support for the main opponent candidate.
\item opponent.party The opponent's party.
\item turnout The electorate turnout in percentages.
\item winner.bmi The winner Body Mass Index (BMI) estimate \code{(BMI = weight in kg/(height in meter)**2)}
 }
}
\keyword{datasets}

