% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lrtest.r
\name{print.lrtest}
\alias{print.lrtest}
\title{Method for printing lrtest}
\usage{
\method{print}{lrtest}(x, ...)
}
\arguments{
\item{x}{a model class object.}

\item{\dots}{typically unecessary parameters.}
}
\description{
Method for printing lrtest
}

