% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SciencesPo.R
\encoding{UTF-8}
\name{print.SciencesPo}
\alias{print.SciencesPo}
\title{Print method for objects of class \code{SciencesPo}.}
\usage{
\method{print}{SciencesPo}(x, method = "pander", ...)
}
\arguments{
\item{x}{SciencesPo object generated with \code{Freq}, \code{Describe} or \code{Standard}.}

\item{method}{one of \dQuote{pander}, \dQuote{viewer} or \dQuote{browser}. If \dQuote{viewer} is used outside RStudio, Web Browser will be used instead.}

\item{\dots}{additional arguments ignored at the moment.}
}
\description{
Display \code{SciencesPo} objects in the console, in \emph{RStudio}'s viewer or in web browser.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{internal}

