% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{project}
\alias{project}
\title{Mehtod for project}
\usage{
project(object, ...)
}
\arguments{
\item{object}{a data object.}

\item{\dots}{Ignored parameters.}
}
\description{
Mehtod for project
}

