% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{prop.miss}
\alias{prop.miss}
\title{Summarize Missingness Across a data frame}
\usage{
prop.miss(.data, order = TRUE)
}
\arguments{
\item{.data}{The dataset or vector.}

\item{order}{A logical, should results be ordered by degree of missingness? Default is \code{order=TRUE}.}
}
\description{
This function calculates the number and percent missing for every variable in .data.
}
\examples{

foo <- data.frame(X1 = rnorm(100),
                  X2 = sample(c(NA, 'A', 'B', 'C'), 100, TRUE),
                  X3 = sample(c('A', NA, 'C', NA), 100, TRUE))

prop.miss(foo)

}

