% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Prune a numeric vector}
\usage{
prune(x, method = c("value", "percentile"), low = NULL, high = NULL,
  replace = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{A character string indicating the desired method of pruning
with \code{"value"} being the default.  This must be (an abbreviation of)
one of the strings \code{"value"} or \code{"percentile"}.
See Details for more information.}

\item{low}{The lower value/percentile for pruning  See Details for more information.}

\item{high}{The upper value/percentile for pruning  See Details for more information.}

\item{replace}{Either \code{NULL} ("rounds" values according to the low and high arguments),
\code{NA},
or a single value that will replace the pruned values.  The default is NULL.}
}
\value{
The output of \code{prune} is a pruned numeric vector with the same
  length as \code{x}.
}
\description{
\code{prune} returns a pruned numeric vector.
}
\details{
\code{prune} returns a pruned version of the numeric vector \code{x}.
  \code{NA} values in \code{x} are ignored during the pruning process
  but are preserved in the output.  \code{prune} will do one-sided pruning
  if only one \code{low} or \code{high} argument is provided
  (e.g. \code{prune(x, low=-1)} will prune \code{x} at a lower value of -1).

\code{prune} is designed to be readable from the function call.  For example:
\itemize{
  \item \code{prune(x, "value", low=-1, high=1)} can be read as
    "\strong{prune} \strong{x} at \strong{-1} and \strong{1}".
  \item \code{prune(x, "percentile", low=.05, high=.95)} can be read as
    "\strong{prune} \strong{x} at the \strong{5}th and \strong{95}th percentiles".
}

The arguments \code{low} and \code{high} are used based on \code{method}.
  \code{prune} offers several different options for \code{method}:

\itemize{
  \item \strong{value}: \code{low} and \code{high} are used as raw values
    (e.g. .05 is the value .05).
  \item \strong{percentile}: \code{low} and \code{high} are used as percentiles
    (e.g. .05 is 5th percentile).
}
}
\examples{
set.seed(51)
x <- rnorm(1e4)
summary(x)

# pruning at the values -1 and 1!
x_val <- prune(x, low=-1, high=1)
summary(x_val)

# pruning at 5th and 95th percentiles!
x_per <- prune(x, "percentile", low=.05, high=.95)
summary(x_per)

# pruning values above the 95th percentile!
x_hi <- prune(x, "percentile", high=.95)
summary(x_hi)

# pruning values converted to NAs!
x_NA <- prune(x, "percentile", low=.05, high=.95, replace=NA)
summary(x_NA)

}

