% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{pub_shape_pal}
\alias{pub_shape_pal}
\title{Shape scales for theme_pub (discrete)}
\usage{
pub_shape_pal(palette = "proportions")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Discrete shape scales for theme_pub.
}
\examples{
library(scales)
pub_shape_pal("proportions")(2)

}
\seealso{
Other shape pub: \code{\link{scale_shape_pub}}
}

