% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{qm}
\alias{qm}
\title{Quote strings}
\usage{
qm(...)
}
\arguments{
\item{\dots}{Any number of names separated by commas.}
}
\description{
Quote strings
}
\examples{
ssex[, qm(Date, Favor, DK)]
}
\keyword{manipulation}

