% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{r}
\alias{r}
\title{# make sure people are using a current version of R}
\format{\preformatted{List of 14
 $ platform      : chr "x86_64-apple-darwin13.4.0"
 $ arch          : chr "x86_64"
 $ os            : chr "darwin13.4.0"
 $ system        : chr "x86_64, darwin13.4.0"
 $ status        : chr ""
 $ major         : chr "3"
 $ minor         : chr "2.0"
 $ year          : chr "2015"
 $ month         : chr "04"
 $ day           : chr "16"
 $ svn rev       : chr "68180"
 $ language      : chr "R"
 $ version.string: chr "R version 3.2.0 (2015-04-16)"
 $ nickname      : chr "Full of Ingredients"
}}
\usage{
r
}
\description{
# make sure people are using a current version of R
}
\keyword{datasets}

