% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELECTORALTOOL.R
\name{rae}
\alias{rae}
\title{Rae Index of Disproportionality}
\usage{
rae(v, s, ...)

rae(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score (The Rae's index or Rae's index of disproportionality).
}
\description{
Calculates the Rae index of disproportionality.
}
\details{
The score is calculated as the sum of the absolute differences \code{|v-s|} divided by the number of parties.
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Rae, D. (1967) \emph{The Political Consequences of Electoral Laws.} London: Yale University Press.

 @examples
# 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

rae(pvotes, pseats)
}
\seealso{
\code{\link{cox.shugart}}, \code{\link{politicalDiversity}}, \code{\link{grofman}}, \code{\link{farina}}, \code{\link{lijphart}}. For more details, see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}
}

