% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbd.R
\encoding{UTF-8}
\name{rbd}
\alias{rbd}
\title{Create a Block-randomized design}
\usage{
rbd(blocksize, n, seed = NULL)
}
\arguments{
\item{blocksize}{is the number of control blocks or n per block/group.}

\item{n}{is the total number of subjects or units.}

\item{seed}{the random number generation seed.}
}
\description{
Generate block-randomized designs based on the number of units \code{n} and block-size, where the block-size is the number of experimental conditions. The number of Independent Variables and the number of levels in each IV are specified as input. The output is a the block randomized design. This function is intended for planning randomized trails.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}

blk <- rbd(blocksize = 10, n = 40, seed = 51)

blk;

crosstable(blk, block, condition)
}
\references{
Alan S Gerber, Donald P Green (2012). \emph{Field experiments: Design, analysis, and interpretation}. WW Norton.

RB Morton, KC Williams (2010). \emph{Experimental political science and the study of causality: From nature to the lab}. Cambridge University Press.
}

