% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zTree.R
\name{read.zTree}
\alias{read.zTree}
\title{Reads zTree output files}
\usage{
read.zTree(object, tables = c("globals", "subjects"))
}
\arguments{
\item{object}{a zTree file or a list of files.}

\item{tables}{the tables of intrest.}
}
\value{
A list of dataframes, one for each table
}
\description{
Extracts variables from a zTree output file.
}
\examples{
\dontrun{url <- 
zTables <- read.zTree( "131126_0009.xls" , "contracts" )
zTables <- read.zTree( c("131126_0009.xls",
"131126_0010.xls"), c("globals","subjects", "contracts" ))
}

}

