% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{regionSignificance}
\alias{regionSignificance}
\title{Computes the region significance}
\usage{
regionSignificance(y, x, z, zval)
}
\arguments{
\item{y}{the dependent variable.}

\item{x}{the independent variable.}

\item{z}{the moderator variable.}

\item{zval}{an specific value for the moderator (z).}
}
\description{
Probing Regression Interactions
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

