% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability}
\alias{reliability}
\title{Reliability Procedure}
\usage{
reliability(...)
}
\arguments{
\item{\dots}{The parameters.}
}
\description{
Simulates the SPSS procedure Reliability
}
\examples{
reliability(attitude)
reliability(attitude[,-6])
reliability(attitude$rating, attitude$complaints,
attitude$privileges, attitude$learning, attitude$raises)
}

