% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{religiosity}
\alias{religiosity}
\title{Data on religiosity of countries}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::religiosity)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::religiosity)} observations.
\itemize{
\item Country Name of country.
\item Religiosity A measure of degree of religiosity for residents of the country.
\item GDP Per capita Gross Domestic Product in the country.
\item Africa Indicator for countries in Africa.
\item EastEurope Indicator for countries in Eastern Europe.
\item MiddleEast Indicator for countries in the Middle East.
\item Asia Indicator for countries in Asia.
\item WestEurope Indicator for countries in Western Europe.
\item Americas Indicator for countries in North/South America.
}
 @source
 The Pew Research Center's Global Attitudes Project surveyed
 people around the world and asked (among many other questions) whether they agreed that "belief
 in God is necessary for morality," whether religion is very important in their lives, and
 whether the at least once per day. The variable Religiosity is the sum of the percentage
 of positive responses on these three items, measured in each of 44 countries. The dataset also
  includes the per capita GDP for each country and indicator variables that record the part of the world the country is in.
 \url{http://www.pewglobal.org}}
\usage{
data(religiosity)
}
\description{
Data from the 2007 Spring Survey conducted through the Pew Global Attitudes Project.
}
\keyword{datasets}

