% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Method for renaming}
\usage{
rename(x, ...)
}
\arguments{
\item{x}{The data object.}

\item{\dots}{The parameters to be used in rename.}
}
\description{
Method for renaming
}

