% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rename.vars.R
\name{rename.vars}
\alias{rename.vars}
\title{Rename variables in a data.frame}
\usage{
\method{rename}{vars}(.data, from = "", to = "", info = TRUE)
}
\arguments{
\item{.data}{dataframe to be modified.}

\item{from}{character vector containing the current name of each variable to be renamed.}

\item{to}{character vector containing the new name of each variable to be renamed.}

\item{info}{boolean value indicating whether to print details of the renaming.  Defaults to TRUE.}

\item{names}{character vector containing the names of variables to be removed.}
}
\description{
Rename variables in a data.frame.
}
\examples{
rename.vars(twins, "C", "D")
}

