% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{Render}
\alias{Render}
\title{Render layer on top of a ggplot}
\usage{
Render(plot = NULL)
}
\arguments{
\item{plot}{the plot to use as a starting point, either a ggplot2 or gtable.}
}
\description{
Set up a rendering layer on top of a ggplot.
}

