% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{replace.char}
\alias{replace.char}
\title{Replace Character}
\usage{
replace.char(str, char = "_", newchar = ".")
}
\arguments{
\item{str}{The target string.}

\item{char}{The replacing character.}

\item{newchar}{The new character.

x <- "replace_strings"

replace.char(x, char = "_", newchar = ".")}
}
\description{
Replaces characters in a given string.
}

