% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{reverseLevels}
\alias{reverseLevels}
\title{Reverse the levels of a factor.}
\usage{
reverseLevels(x)
}
\arguments{
\item{x}{a factor whose levels need to be reverse coded.}
}
\description{
Reverse the levels of a factor.
}
\examples{
mylevels <- c('Strongly Disagree', 'Disagree', 'Neither', 'Agree', 'Strongly Agree')

test <- factor(sample(mylevels[1:5], 10, replace=TRUE))

reverseLevels(test)

cbind(test, as.integer(test), as.integer(reverseLevels(test)))
}

