% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\name{rle2data.frame}
\alias{rle2data.frame}
\title{Converts rle object to data.frame}
\usage{
rle2data.frame(r)
}
\arguments{
\item{r}{an rle object.


just converts an rle object to a data.frame}
}
\description{
Converts rle object to data.frame
}

