% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\name{rleFrame}
\alias{rleFrame}
\title{Converts rle object to data.frame}
\usage{
rleFrame(r)
}
\arguments{
\item{r}{an rle object.


just converts an rle object to a data.frame}
}
\description{
Converts rle object to data.frame
}

