% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/roman.numeral.R
\name{roman.numeral}
\alias{roman.numeral}
\title{Converts integer into its Roman Numeral equivalent}
\usage{
roman.numeral(x)
}
\arguments{
\item{x}{The integer to be translated.}
}
\description{
Converts integer into its Roman Numeral equivalent
}
\examples{
roman.numeral(1977)
}

