% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/romanNumeral.R
\name{romanNumeral}
\alias{romanNumeral}
\title{Converts integer into its Roman Numeral equivalent}
\usage{
romanNumeral(x)
}
\arguments{
\item{x}{The integer to be translated.}
}
\description{
Converts integer into its Roman Numeral equivalent
}
\examples{
romanNumeral(1977)
}

