% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/romanNumerals.R
\name{romanNumerals}
\alias{romanNumerals}
\title{Converts integer into its Roman Numeral equivalent}
\usage{
romanNumerals(x)
}
\arguments{
\item{x}{The integer to be translated.}
}
\description{
Converts integer into its Roman Numeral equivalent
}
\examples{
romanNumerals(1977)
}

