% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELECTORALTOOL.R
\name{rose}
\alias{rose}
\title{Rose Index of Proportionality}
\usage{
rose(v, s, ...)

rose(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates the Rose index of proportionality to show the relationship of votes to seats. The Rose index is a standardized version of the Loosemore-Hanby index.
}
\details{
The score is calculated  as the difference between a party's
 percentage share of the vote and its percentage share of the total seats
 in Parliament, summed, divided by two and subtracted from 100. Theoretically it can range from 0 to 100.
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Rose, Richard, Neil Munro and Tom Mackie (1998) \emph{ Elections in Central and Eastern Europe Since 1990.} Glasgow: Centre for the Study of Public Policy, University of Strathclyde.

 @examples
# 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

rose(pvotes, pseats)
}
\seealso{
\code{\link{loosemore.hanby}}, \code{\link{rae}}, \code{\link{cox.shugart}}, \code{\link{cox.shugart}}, \code{\link{politicalDiversity}}, \code{\link{grofman}}, \code{\link{farina}},  \code{\link{lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}
}

