% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{rotate_x_text}
\alias{rotate_x_text}
\title{Rotate ggplot2 x-axis labels}
\usage{
rotate_x_text(angle)
}
\arguments{
\item{angle}{the angle of rotation.}
}
\description{
Rotate the x-axis labels/text by a given angle.
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}
\keyword{ggplot2}

