% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rounded.R
\encoding{UTF-8}
\name{rounded}
\alias{rounded}
\title{Round Numbers with no Leading Zero}
\usage{
rounded(x, digits = 2, add = TRUE, max = (digits + 3))
}
\arguments{
\item{x}{A numeric vector of values to be rounded.}

\item{digits}{An integer for the number of digits to round to.}

\item{add}{An optional dichotomous indicator for whether additional digits should be added if no numbers appear in pre-set digit level.}

\item{max}{Maximum number of digits to be shown if \code{add=TRUE}.}
}
\description{
Rounds numbers without leading zeros.
}
\examples{
rounded(seq(0, 1, by=.1))
}

