% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/round.extra.R
\encoding{UTF-8}
\name{roundExtra}
\alias{roundExtra}
\title{Round a number}
\usage{
roundExtra(x, down = TRUE)
}
\arguments{
\item{x}{Number or numeric vector to be rounded.}

\item{down}{If FALSE, values are rounded up instead of down (default).}
}
\value{
Returns a value or numeric vector of rounded valules.
}
\description{
Given a number or numeric vector, rounds each value up or down something nice for plotting.
}
\examples{
rough <- sort(runif(10, min=-3, max=3))
  data.frame(down=roundExtra(rough, down=TRUE),
  value=rough, up=roundExtra(rough))
}

