% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{rowMedians}
\alias{rowMedians}
\title{Calculate the Median per Row}
\usage{
rowMedians(x, na.rm = FALSE)
}
\arguments{
\item{x}{a data.frame or matrix.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
removed before the computation proceeds. The default is \code{FALSE}.}
}
\value{
A vector of the medians of each row of \code{x}
}
\description{
Compute the sample medians of the rows of a data.frame or matrix.
}
\examples{
data(ltaylor96)
rowMedians( ltaylor96)
m <- matrix( 1:12, nrow = 4 )
rowMedians( m )
}
\seealso{
\code{\link{colMedians}},\code{\link{median}},\code{\link{rowMeans}}.
}

