% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{rownames2col}
\alias{rownames2col}
\title{Rownames to column}
\usage{
rownames2col(data, rowname = "rowname")
}
\arguments{
\item{data}{the data frame.}

\item{rowname}{the column name.}
}
\description{
Moves rownames to column
}
\examples{
x <- data.frame(c = c(87,8,8,87,38,92))
rownames(x) <- c("B", "H","I","J", "K","L")
rownames2col(x)
}

