% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CONVENIENCE.R
\encoding{UTF-8}
\name{rowSample}
\alias{rowSample}
\title{Row Sample}
\usage{
rowSample(.data, n)
}
\arguments{
\item{.data}{the data object.}

\item{n}{the number of items to choose from.}
}
\description{
Sample rows of a data object.
}
\examples{
rowSample(iris, 20)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

