% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UTILS.R
\encoding{UTF-8}
\name{safe.chars}
\alias{safe.chars}
\title{Convert All Factor Columns to Character Columns}
\usage{
safe.chars(.data)
}
\arguments{
\item{.data}{The name of the \code{data.frame}}
}
\description{
By default, R converts character columns to factors.
Instead of re-reading the data using \code{stringsAsFactors}, the
\code{\link{safe.chars}} function will identify which columns are currently factors, and convert them all to characters.
}
\examples{
str(iris)
iris_2 = safe.chars(iris)
str(iris_2)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{read.table}}, \code{\link{destring}}.
}

