% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{safeRequire}
\alias{safeRequire}
\title{Safe require}
\usage{
safeRequire(packageName)
}
\arguments{
\item{packageName}{the name of a package.}
}
\description{
This function checks whether a package is installed; if not, it installs it. It then loads the package.
}

