% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMPLING.R
\name{sampleCircular}
\alias{sampleCircular}
\title{Circular Systematic Sampling}
\usage{
sampleCircular(N = 500, n = 30)
}
\arguments{
\item{N}{The population size.}

\item{n}{The sample size.}
}
\description{
Circular systematic sampling.
}
\examples{
sampleCircular(500, 30)
}
\keyword{Sampling}

