% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMPLING.R
\name{sampleCircularReplicated}
\alias{sampleCircularReplicated}
\title{Replicated Circular-Systematic Sampling}
\usage{
sampleCircularReplicated(N = 500, n = 30, g = 6)
}
\arguments{
\item{N}{The population size.}

\item{n}{The sample size.}

\item{g}{Number of independent sub-samples, each containing m = n/g units. Notice that m has to be a multiple of n and g.}
}
\description{
Replicated circular systematic sampling.
}
\examples{
sampleCircularReplicated(500, 30, 6)
}

