% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAMPLING.R
\name{sampleLinear}
\alias{sampleLinear}
\title{Linear Systematic Sampling}
\usage{
sampleLinear(N = 500, n = 30)
}
\arguments{
\item{N}{The population size.}

\item{n}{The sample size.}
}
\description{
Linear systematic sampling.
}
\examples{
sampleLinear(500, 30)
}
\keyword{Sampling}

