% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAMPLING.R
\name{sampleLinearReplicated}
\alias{sampleLinearReplicated}
\title{Replicated-Systematic Random Sampling}
\usage{
sampleLinearReplicated(N = 500, n = 30, g = 6)
}
\arguments{
\item{N}{The population size.}

\item{n}{The sample size.}

\item{g}{Number of independent sub-samples, each containing m = n/g units. Notice that m has to be a multiple of n and g.}
}
\description{
Replicated-Systematic Random Sampling.
}
\examples{
sampleLinearReplicated(500, 30, 6)
}

