% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleSize}
\alias{sampleSize}
\title{Simple Sample Size for Surveys}
\usage{
sampleSize(p, delta = "auto", popsize = NULL, deff = 1, alpha = 0.05)
}
\arguments{
\item{p}{The proportion.}

\item{delta}{The error size.}

\item{popsize}{An integer for the population size.}

\item{deff}{An intger for the deff.}

\item{alpha}{The level of alpha/significance.}
}
\description{
Compute sample size for surveys.
}
\examples{
# Comercial public opinion samples in Brazil:
sampleSize(p=.50, delta=.03)
sampleSize(p=.50, delta=.02)
}

