% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{scale}
\alias{scale}
\title{Scalling Method}
\usage{
scale(x, ...)
}
\arguments{
\item{x}{data vector or variable to be scaled.}

\item{\dots}{optional arguments that are passed to scale.}
}
\description{
Scalling Method
}

