% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_color_blind}
\alias{scale_color_blind}
\alias{scale_colour_blind}
\title{Scale color-blind-friendly for ggplot graphs}
\usage{
scale_color_blind(...)
}
\arguments{
\item{\dots}{parameters to be used.}
}
\description{
Discrete  color-blind-friendly scale for ggplot().
}

