% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_colour_parties}
\alias{scale_color_parties}
\alias{scale_colour_parties}
\alias{scale_fill_parties}
\title{Political Parties Color Scales}
\usage{
scale_colour_parties(palette = "BRA", ...)

scale_fill_parties(palette = "BRA", ...)

scale_color_parties(palette = "BRA", ...)
}
\arguments{
\item{palette}{Palette name.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Scale color for political parties.
}
\seealso{
\code{\link{parties_color_pal}} for references.

Other colour parties: \code{\link{parties_color_pal}}
}

