% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{scale_color_party}
\alias{scale_color_party}
\alias{scale_fill_party}
\title{Political Parties Color Scales}
\usage{
scale_color_party(palette = "BRA", ...)

scale_fill_party(palette = "BRA", ...)
}
\arguments{
\item{palette}{the palette name.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Scale color for political parties.
}
\seealso{
\code{\link{party_color_pal}} for references.

Other color party: \code{\link{party_color_pal}}
}

