% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_colour_pub}
\alias{scale_color_pub}
\alias{scale_colour_pub}
\alias{scale_fill_pub}
\title{Publication color scales.}
\usage{
scale_colour_pub(palette = "tableau10", ...)

scale_fill_pub(palette = "tableau10", ...)

scale_color_pub(palette = "tableau10", ...)
}
\arguments{
\item{palette}{Palette name.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See \code{\link{pub_color_pal}} for details.
}
\seealso{
\code{\link{pub_color_pal}} for references.
}

