% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_colour_tableau}
\alias{scale_color_tableau}
\alias{scale_colour_tableau}
\alias{scale_fill_tableau}
\title{Tableau color scales.}
\usage{
scale_colour_tableau(palette = "tableau10", ...)

scale_fill_tableau(palette = "tableau10", ...)

scale_color_tableau(palette = "tableau10", ...)
}
\arguments{
\item{palette}{Palette name.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See \code{\link{tableau_color_pal}} for details.
}
\details{

}
\seealso{
\code{\link{tableau_color_pal}} for references.
}

