% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_fill_colorblind}
\alias{scale_color_colorblind}
\alias{scale_colour_colorblind}
\alias{scale_fill_colorblind}
\title{Color blind color palette (discrete)}
\usage{
scale_fill_colorblind(...)

scale_colour_colorblind(...)

scale_color_colorblind(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Color palettes for color blind.
}
\seealso{
Other colour colorblind: \code{\link{colorblind_pal}}
}

