% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_colour_pub}
\alias{scale_colour_pub}
\title{Scale colors for ggplot graphs}
\usage{
scale_colour_pub(...)
}
\arguments{
\item{\dots}{parameters to be used.}
}
\description{
Discrete scale colors for the theme_pub().
}

