% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_fill_pub}
\alias{scale_fill_pub}
\title{Scale colors for ggplot graphs}
\usage{
scale_fill_pub(...)
}
\arguments{
\item{\dots}{parameters to be used.}
}
\description{
Scale colors for ggplot graphs
}

