% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_fill_tableau20}
\alias{scale_color_tableau20}
\alias{scale_fill_tableau20}
\title{The Tableau 20 colors for ggplot graphs}
\usage{
scale_fill_tableau20(...)

scale_color_tableau20(...)
}
\arguments{
\item{\dots}{parameters to be used.}

\item{\dots}{parameters to be used.}
}
\description{
Discrete scale colors like the Tableau 20 colors.

Discrete  color scale for ggplot2.
}

