% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{scale_colour_fte}
\alias{scale_color_fte}
\alias{scale_colour_fte}
\alias{scale_fill_fte}
\title{fivethirtyeight.com color scales}
\usage{
scale_colour_fte(...)

scale_color_fte(...)

scale_fill_fte(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Color scales using the colors in the fivethirtyeight graphics.
}
\seealso{
\code{\link{theme_538}} for examples.

Other colour fte: \code{\link{fte_color_pal}}
}

