% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{scale_manual}
\alias{scale_color_flex}
\alias{scale_fill_flex}
\alias{scale_manual}
\title{SciencesPo colors and fills for ggplot2.}
\usage{
scale_color_flex(..., theme = "tableau20")

scale_fill_flex(..., theme = "tableau20")
}
\arguments{
\item{...}{common discrete scale parameters: `name`, `breaks`, `labels`, `na.value`, `limits` and `guide`. See [discrete_scale()] for more details}

\item{theme}{one of "tableau20", "tableau10", "colorblind" etc. This should match the `theme_flex()` that is used with it.}
}
\description{
The SciencesPo scales add colors that work nicely with `theme_flex()`.
}
\details{
\describe{

\item{`scale_color_flex`}{
For use when `color` is specified as an `aes()` in a ggplot.}

\item{`scale_fill_flex`}{
For use when `fill` is specified as an `aes()` in a ggplot.}
}
}
\examples{

# Plot some variables
p <- ggplot(mtcars) +
     geom_point(aes(x = wt, y = mpg, color = factor(gear))) +
     facet_wrap(~am)

# Plot with SciencesPo theme and colors
p +
    theme_flex() +
   scale_color_flex(theme = '538')


}
\seealso{
[theme_flex()]
}

