% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot.R
\name{Scatterplot}
\alias{Scatterplot}
\title{Quartile-Frame Scatterplot}
\usage{
Scatterplot(x, y, ...)
}
\arguments{
\item{x, y}{aesthetics passed into each layer variables
# @param data data frame to use (optional).}

\item{\dots}{additional arguments.}
}
\description{
Plots a scatterplot with a custom axis that acts as a quartile plot (quartile-frame). Inspired by The Visual Display of Quantitative Information by Edward R. Tufte.
}
\examples{
with(mtcars, Scatterplot(x = wt, y = mpg,
main = "Vehicle Weight-Gas Mileage Relationship",
xlab = "Vehicle Weight",
ylab = "Miles per Gallon",
font.family = "serif") )

}

