% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\encoding{UTF-8}
\name{SciencesPo-deprecated}
\alias{SciencesPo-deprecated}
\alias{politicalDiversity}
\alias{svTransform}
\alias{untable}
\title{Deprecated functions in package \sQuote{SciencesPo}}
\usage{
svTransform(y)

politicalDiversity(x, index = "laakso/taagepera", margin = 1,
  base = exp(1))

untable(x, ...)
}
\arguments{
\item{y}{the dependent variable.}

\item{x}{A data.frame, a matrix-like, or a vector containing values for the number of votes or seats each party received.}

\item{index}{The type of index desired, one of "laakso/taagepera",  "golosov", "herfindahl", "gini", "shannon", "simpson", "invsimpson".}

\item{margin}{The margin for which the index is computed.}

\item{base}{The logarithm base used in some indices, such as the "shannon" index.}

\item{\dots}{Extra parameters.}
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{SciencesPo} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
 the replacement indicated below:
 \itemize{
     \item{svTransform: \code{\link{Normalize}}}
     \item{untable: \code{\link{Untable}}}
     \item{detail: \code{\link{Summary}}}
     \item{politicalDiversity: \code{\link{PoliticalDiversity}}}
     \item{winsorize: \code{\link{Winsorize}}}
     \item{recode: \code{\link{Recode}}}
     \item{dummy: \code{\link{Dummify}}}
     \item{gini: \code{\link{Gini}}}
     \item{clear: \code{\link{Clear}}}
     \item{destring: \code{\link{Destring}}}
     \item{outliers: \code{\link{Outlier}}}
     \item{voronoi: \code{\link{Voronoi}}}
     \item{jackknife: \code{\link{Jackknife}}}
     \item{bootstrap: \code{\link{Bootstrap}}}
     \item{normalize: \code{\link{Normalize}}}

}
}

