% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{SciencesPo.Palettes}
\alias{SciencesPo.Palettes}
\title{Palette data for the themes used by package}
\format{A \code{list}.}
\usage{
SciencesPo.Palettes
}
\description{
Data used by the palettes in the package.
}
\keyword{datasets}

