% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{scpo.wtd.corr}
\alias{scpo.wtd.corr}
\title{Computes Weighted Correlations}
\usage{
scpo.wtd.corr(x, y = NULL, weights = NULL)
}
\arguments{
\item{x}{A matrix or vector to correlate with \code{y}.}

\item{y}{A matrix or vector to correlate with \code{x}. If \code{y} is NULL, \code{x} will be used instead.}

\item{weights}{An optional vector of weights to be used to determining the weighted mean and variance for calculation of the correlations.}
}
\description{
Computes weighted correlations.
}
\examples{
 x <- sample(10,10)
 y <- sample(10,10)
 w <- sample(5,10, replace=TRUE)

scpo.wtd.corr(x, y, w)

}

