% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{scpo.wtd.mean}
\alias{scpo.wtd.mean}
\title{Computes Weighted Mean}
\usage{
scpo.wtd.mean(x, weights = NULL, normwt = "ignored", na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{A numeric vector of weights of \code{x}.}

\item{normwt}{Ignored.}

\item{na.rm}{A logical, if \code{TRUE}, missing data will be dropped.  If \code{na.rm = FALSE}, missing data will return an error.}
}
\description{
Computes a weighted mean of data.
}
\examples{
x <- sample(10,10)
w <- sample(5,10, replace=TRUE)

scpo.wtd.mean(x, w)
}

