% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{scpo.wtd.prop}
\alias{scpo.wtd.prop}
\title{Computes Weighted Proportions}
\usage{
scpo.wtd.prop(x, weights = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{A vector for which a set of proportions is desired.}

\item{weights}{A vector of weights to be used to determining the weighted proportion in each category of \code{x}.}

\item{na.rm}{A logical, if \code{TRUE}, missing data will be dropped.  If \code{na.rm = FALSE}, missing data will return an error.}
}
\description{
Computes a weighted proportions table of data in each category for any variable.
}
\examples{
x <- sample(10,10)
w <- sample(5,10, replace=TRUE)

scpo.wtd.prop(x, w)

}

