% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{scpo.wtd.stdz}
\alias{scpo.wtd.stdz}
\title{Computes Weighted Standardized Values}
\usage{
scpo.wtd.stdz(x, weights = NULL)
}
\arguments{
\item{x}{A vector for which a set of standardized values is desired.}

\item{weights}{A vector of weights to be used to determining the weighted values of \code{x}.}
}
\description{
Computes weighted standardized values of data in each category for any variable.
}
\examples{
 x <- sample(10,10)
 w <- sample(5,10, replace=TRUE)

scpo.wtd.stdz(x, w)
}

