% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{scpo.wtd.table}
\alias{scpo.wtd.table}
\title{Weighted Frequency Table}
\usage{
scpo.wtd.table(x, weights = NULL, type = c("list", "table"),
  normwt = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector, may be a character or category or factor vector.}

\item{weights}{A numeric vector of weights.}

\item{type}{The default type is "list", meaning that the function is to return a list containing two vectors: x is the sorted unique values of x and sum.of.weights is the sum of weights for that x. This is the default so that you don't have to convert the names attribute of the result that can be obtained with type="table" to a numeric variable when x was originally numeric. type="table" for wtd.table results in an object that is the same structure as those returned from table.}

\item{normwt}{specify normwt=TRUE to make weights sum to length(x) after deletion of NAs. If weights are frequency weights, then normwt should be FALSE, and if weights are normalization (aka reliability) weights, then normwt should be TRUE. In the case of the former, no check is made that weights are valid frequencies.}

\item{na.rm}{A logical, if \code{TRUE}, missing data will be dropped.  If \code{na.rm = FALSE}, missing data will return an error.}
}
\description{
Computes a weighted frequency table, only one stratification variable is supported.
}
\examples{
 x <- sample(5,25, replace=TRUE)
 w <- sample(3,25, replace=TRUE)

scpo.wtd.table(x, w)

}

