% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{setdiff.data.frame}
\alias{setdiff.data.frame}
\title{Check for differences in data.frames}
\usage{
setdiff.data.frame(A, B)
}
\arguments{
\item{A}{The original data object.}

\item{B}{The other data.frame}
}
\description{
Check for differences in data.frames
}

