% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{scale_shape_pub}
\alias{scale_shape_pub}
\title{Shape scales for theme_pub (discrete)}
\usage{
scale_shape_pub(palette = "pub", ...)
}
\arguments{
\item{palette}{the palette name, a character string.}

\item{...}{common discrete scale parameters: \code{name}, \code{breaks},
\code{labels}, \code{na.value}, \code{limits} and \code{guide}.  See
\code{\link{discrete_scale}} for more details}
}
\description{
See \code{\link{pub_shape_pal}} for details.
}
\examples{
library("ggplot2")
library("scales")

p <- ggplot(mtcars) +
     geom_point(aes(x = wt, y = mpg, shape = factor(gear))) +
     facet_wrap(~am)
p + scale_shape_pub()

}
\seealso{
Other shape pub: \code{\link{pub_shape_pal}}
}
\keyword{ggplot2}

