\name{sheston91}
\alias{sheston91}
\docType{data}
\title{The Penn World Table
}
\description{The Penn World Table used in Summers and Heston (1991).
}
\usage{data(sheston91)}
\format{
  A data frame with 3250 observations on the following 8 variables.
  \describe{
    \item{\code{year}}{Year}
    \item{\code{pop}}{ Population (thousands)}
    \item{\code{rgdppc}}{Real per capita GDP}
    \item{\code{savrat}}{a numeric vector}
    \item{\code{country}}{Country}
    \item{\code{com}}{ Communist regime}
    \item{\code{opec}}{OPEC country}
    \item{\code{name}}{Country name}
  }
}

\source{Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\references{
Summers, R. and Heston, A. (1991) The Penn World Table (Mark 5): an expanded set of international comparisons, 1950--1988. \emph{The Quarterly Journal of Economics,} \bold{106(2),} 327--368.

}
\examples{
data(sheston91)

}
\keyword{datasets}
