% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shift}
\alias{shift}
\title{Lag or Lead Observations}
\usage{
shift(x, delta = NA)
}
\arguments{
\item{x}{The variable to be lagged or leaded}

\item{delta}{An integer for units to move backward or forward as negative (-#) or positive (#)}
}
\value{
An object of the same type as \code{x}
}
\description{
Shift function allows one to either lag or lead a column variables in a data frame
}
\examples{
weather <- data.frame(
month = a <- c('J','F','M','A','M','J','J','A','S','O','N','D'),
precip = b <- c(78,62,74,78,76, 83, 91, 93, 93,78, 93, 81),
max = c <- c(-6, -4, 2, 11, 19, 24,  26,  25, 20, 13, 5, -2),
min = d <- c(-15, -13, -7, 1, 8,  13, 16, 14, 9, 3, -2, -10) )

(weather$L.precip <- shift(weather$precip, -1) ) # lag

(weather$precip.L <- shift(weather$precip, 1) ) # lead
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

