% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siegel.tukey.R
\name{siegel.tukey}
\alias{siegel.tukey}
\title{sample1 <- c(20.30, 22.53, 25.70, 13.23, 29.67, 24.46, 26.07, 19.35, 17.813,
            16.00, 15.30, 32.90)
sample2 <- c(10.56, 28.13, 19.94, 11.03, 8.093, 12.95, 21.14, 32.50, 10.90)
ks.test(sample1,sample2) # <-- Fail to reject H0 for alpha = .05
ks.test(sample1,sample2,alternative = "greater")
ks.test(sample1,sample2,alternative = "less")}
\usage{
siegel.tukey(x, y, id.col = FALSE, adjust.median = F, rnd = -1,
  alternative = "two.sided", mu = 0, paired = FALSE, exact = FALSE,
  correct = TRUE, conf.int = FALSE, conf.level = 0.95)
}
\description{
sample1 <- c(20.30, 22.53, 25.70, 13.23, 29.67, 24.46, 26.07, 19.35, 17.813,
            16.00, 15.30, 32.90)
sample2 <- c(10.56, 28.13, 19.94, 11.03, 8.093, 12.95, 21.14, 32.50, 10.90)
ks.test(sample1,sample2) # <-- Fail to reject H0 for alpha = .05
ks.test(sample1,sample2,alternative = "greater")
ks.test(sample1,sample2,alternative = "less")
}
\examples{
x=c(4,4,5,5,6,6); y=c(0,0,1,9,10,10);
siegel.tukey(x,y)
}

