% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq2.R
\name{smallAsMissing}
\alias{smallAsMissing}
\title{Recode very small levels into Missing values}
\usage{
smallAsMissing(variables, threshold = 0.05, new.level = "Missing")
}
\arguments{
\item{variables}{is a data.frame of factors}

\item{threshold}{is the proportion below which a level is recoded}

\item{new.level}{is the new label for the recoded levels}
}
\value{
A data.frame
}
\description{
The function recodes all levels with proportions below a certain threshold into "Missing".
 This is useful for a very crude first look at a lot of variables.
}

