% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/split.names.R
\encoding{UTF-8}
\name{split.names}
\alias{split.names}
\title{Splits name fields}
\usage{
\method{split}{names}(name, data = .data)
}
\arguments{
\item{name}{the name field column.}

\item{data}{the data.frame name.}
}
\value{
two columns or a list.
}
\description{
It does one thing: splits a name allocating the first and last names into two new columns or a list.
}
\details{
The way one may split names is region dependent, so this function may apply to very few contexts. See for instance \url{http://www.w3.org/International/questions/qa-personal-names}
}
\examples{
df <- data.frame( name = c("Martin Luther King",
 "Nelson Mandela", "Simon Bolivar") )
 split.names(df$name)
 df$n<- split.names(df$name)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

