% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/srh.R
\name{srh}
\alias{srh}
\title{Scheirer-Ray-Hare Test}
\usage{
srh(r, pf, sf, n = 3)
}
\arguments{
\item{r}{The rank vvector}

\item{pf}{The first factor.}

\item{sf}{The second factor.}

\item{n}{The number of elements in each subgroup.}
}
\description{
Performs the Scheirer-Ray-Hare test for balanced data.
}
\examples{
# Sokal RR, Rohlf FJ, 1995. Biometry, 3rd ed. New York: Freeman.
rat <- c(709,679,699,657,594,677,592,538,476,508,505,539)
fat <- gl(2,6,labels=c("Fresh","Rancid"))
sex <- gl(2,3,12,labels=c("man","woman"))
srh(rat,sex,fat)
}

