% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMENTS.R
\name{stdkurtosis}
\alias{stdkurtosis}
\alias{stdkurtosis.data.frame}
\alias{stdkurtosis.default}
\title{Standard Error of Kurtosis}
\usage{
stdkurtosis(x, na.rm = TRUE)

\method{stdkurtosis}{default}(x, na.rm = TRUE)

\method{stdkurtosis}{data.frame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An \R object.}

\item{na.rm}{a logical value indicating whether \code{NA}
should be stripped before the computation proceeds.}
}
\description{
Generate the standard error of the kurtosis.
}

