% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMENTS.R
\name{stdskewness}
\alias{stdskewness}
\alias{stdskewness.data.frame}
\alias{stdskewness.default}
\title{Standard Error of Skewness}
\usage{
stdskewness(x, na.rm = TRUE)

\method{stdskewness}{default}(x, na.rm = TRUE)

\method{stdskewness}{data.frame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An \R object.}

\item{na.rm}{a logical value indicating whether \code{NA}
should be stripped before the computation proceeds.}
}
\description{
Generate the standard error of the skewness.
}

