% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CONVENIENCE.R
\encoding{UTF-8}
\name{stripWhite}
\alias{stripWhite}
\title{Strip white spaces}
\usage{
stripWhite(x, delim = " ")
}
\arguments{
\item{x}{is a character vector.}

\item{delim}{is the delimiter, default is white spaces \code{" "}}
}
\description{
Strip white spaces
}

