% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sturges.R
\encoding{UTF-8}
\name{sturges}
\alias{sturges}
\title{Calculate breaks

Calculate breaks according to the Herbert Sturges' (1926) formula}
\usage{
sturges(x)
}
\arguments{
\item{x}{A vector of count values.}
}
\description{
Calculate breaks

Calculate breaks according to the Herbert Sturges' (1926) formula
}
\examples{
set.seed(51)
y <- sample(100)
sturges(y)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Sturges, H. (1926) \emph{The choice of a class-interval.} J. Amer. Statist. Assoc., 21, 65-66.
}

