% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{summary.crosstable}
\alias{summary.crosstable}
\title{Print Crosstable style}
\usage{
\method{summary}{crosstable}(object, digits = 2, latex = FALSE, ...)
}
\arguments{
\item{object}{The table object.}
}
\description{
Print Crosstable style
}

