% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sumres.R
\encoding{UTF-8}
\name{sumres}
\alias{sumres}
\title{Sumarizes the residuals from a linear model}
\usage{
sumres(x)
}
\arguments{
\item{x}{The fitted model object.}
}
\description{
Simply sumarizes the residuals from a linear model.
#'
}
\examples{
model<- glm(child~parent, data=galton)
sumres(model)
}

