% Generated by roxygen2 (4.0.2): do not edit by hand
\name{svTransform}
\alias{svTransform}
\title{Simple function to transform a dependent variable that in [0,1] rather than
(0, 1) to beta regression. Suggested by Smithson & Verkuilen (2006).}
\usage{
svTransform(y)
}
\arguments{
\item{y}{vector of the dependent variable that is in [0, 1].}
}
\description{
Simple function to transform a dependent variable that in [0,1] rather than
(0, 1) to beta regression. Suggested by Smithson & Verkuilen (2006).
}

