% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{swatson93}
\alias{swatson93}
\title{Stock's and Watson's (1993) Data.}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::swatson93)} variables and \Sexpr{nrow(SciencesPo::swatson93)} observations.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(swatson93)
}
\description{
Data set used by Stock and Watson (1993) to estimate co-integration. This dataset contains the following columns:
\itemize{
\item lnm1 Log M1.
\item lnp Log NNP price deflator.
\item lnnnp Log NNP.
\item cprate A numeric vector.
\item year Commercial paper rate.
 }
}
\references{
Stock, J. H., and Watson, M. W. (1993) A simple estimator of cointegrating vectors in higher order integrated systems. \emph{Econometrica: Journal of the Econometric Society,} 783--820.
}
\keyword{datasets}

