% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tableau.R
\encoding{UTF-8}
\name{tableau}
\alias{tableau}
\title{Proportion Table for Univariate Analysis}
\usage{
tableau(var, title = NULL)
}
\arguments{
\item{var}{is the vector or the object to be computed proportions.}

\item{title}{A character vector as title.}
}
\value{
A data frame object.
}
\description{
Produces a latex-like table of the variable proportions.
}
\examples{
x <- rep(1:5, 100)

cap <- "My Table With Proportions"

tableau(x, title = cap)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Descriptive}
\keyword{Tables}

