% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{tableau_color_pal}
\alias{tableau_color_pal}
\title{Color Palettes based on Tableau (discrete)}
\usage{
tableau_color_pal(palette = "tableau10")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes used in
\href{http://www.tableausoftware.com/}{Tableau}.
}

