% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableplot.R
\name{Tableplot}
\alias{Tableplot}
\title{Tableplot: a Visualization of Datasets}
\usage{
Tableplot(.data = NULL)
}
\arguments{
\item{.data}{the data frame.}
}
\description{
Visualization of Large Datasets.
}

