% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabstat.R
\name{tabstat}
\alias{tabstat}
\alias{tabstat.default}
\title{Table With Summary Statistics}
\usage{
tabstat(.data, var, by = NULL, statistics = c("nnmiss", "mean", "sd"))

\method{tabstat}{default}(.data, var, by = NULL, statistics = c("nnmiss",
  "mean", "sd"))
}
\arguments{
\item{.data}{The \code{data.frame} names.}

\item{var}{The name of variable or column.}

\item{by}{The name of grouping variable.}

\item{statistics}{The name of desired statistics.}
}
\description{
Generates a table with summary statistics.
}
\examples{
tab = tabstat(Presidents, c("winner.height", "winner.vote", "turnout"))
# knitr::kable(tab, digits=2)
}
\keyword{Exploratory}

