% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ternaryplot.R
\name{Ternaryplot}
\alias{Ternaryplot}
\title{Ternary Plot
Build ternary plots.}
\usage{
Ternaryplot(data, ...)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link{fortify}}. If not specified,
must be suppled in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Ternary Plot
Build ternary plots.
}

