% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{testAssociation}
\alias{testAssociation}
\title{Association tests for contingency tables}
\usage{
testAssociation(x)
}
\arguments{
\item{x}{an \code{r x c} table.}
}
\value{
The association statistics tests for \code{r x c} table.
}
\description{
Calculates the Likelihood Ratio chi-Squared test, the Pearson chi-Squared test, the phi coefficient, the contingency coefficient and Cramer's V. Cramer V / Phi test using two vectors with more than 2 levels.
}
\examples{
x = sample(1:2, 30, TRUE);
 y = sample(1:3, 30, TRUE)
 testAssociation(table(x, y))
}

