% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\encoding{UTF-8}
\name{textWrap}
\alias{textWrap}
\title{Insert Line Breaks in Long Strings}
\usage{
textWrap(labels, wrap, linesep = NULL)
}
\arguments{
\item{labels}{Label(s) as character string, where a line break should be
inserted. Several strings may be passed as vector  (see 'Examples').}

\item{wrap}{Maximum amount of chars per line (i.e. line length). If code{wrap = Inf},
no word wrap will be performed (i.e. \code{labels} will be returned as is).}

\item{linesep}{By default, this argument is \code{NULL} and a regular new line
string (\code{"\\n"}) is used. For HTML-purposes, for instance, \code{linesep}
could be \code{"<br>"}.}
}
\value{
New label(s) with line breaks inserted at every \code{wrap}'s position.
}
\description{
Insert line breaks in long character strings. Useful for wrapping labels / titles for plots and tables.
}
\examples{
textWrap(c("A very long string", "And another even longer string!"), 10)

}
\keyword{internal}

