% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_538}
\alias{theme_538}
\title{Themes for ggplot graphs}
\usage{
theme_538(base_size = 13, font_family = "", legend.position = "none")
}
\arguments{
\item{base_size}{Enables to set the font size of all text elements.}

\item{font_family}{Default font family of all text elements.}

\item{legend.position}{Enables to set legend position, default is "none".}
}
\description{
Theme for plotting  with ggplot.
}

