% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_black}
\alias{theme_black}
\title{ggplot2 Theme with No Background or Gridlines.}
\usage{
theme_black(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{The size to use for text.  Various textual components are
scaled off of this value.}

\item{base_family}{The base font family.}
}
\description{
A ggplot2 theme with no background and no gridlines.
}
\seealso{
\code{\link[ggplot2]{theme}}, \code{\link{theme_pub}}, \code{\link{theme_map}}.
}

