% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_darkside}
\alias{theme_darkside}
\title{The Dark Side Theme}
\usage{
theme_darkside(font_size = 12, font_family = "serif", legend = "none")
}
\arguments{
\item{font_size}{Overall font size. Default is 12.}

\item{font_family}{Default font family.}

\item{legend}{the position of the legend if any.}
}
\value{
The theme.
# plot with small amount of remaining padding
qplot(1:10, (1:10)^2, color="green") + theme_darkside()
# Check that it is a complete theme
attr(theme_darkside(), "complete")
}
\description{
The dark side of the things.
}

