% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_flex}
\alias{theme_flex}
\title{A Flex Theme}
\usage{
theme_flex(base_family = "sans", base_size = 11,
  plot_title_family = base_family, plot_title_size = 18,
  plot_title_face = "bold", plot_title_margin = 10,
  subtitle_family = base_family, subtitle_size = 12,
  subtitle_face = "plain", subtitle_margin = 15,
  strip_text_family = base_family, strip_text_size = 12,
  strip_text_face = "plain", caption_family = base_family,
  caption_size = 9, caption_face = "italic", caption_margin = 10,
  axis_title_family = subtitle_family, axis_title_size = 9,
  axis_title_face = "plain", axis_title_just = "rt",
  plot_background = TRUE, plot_margin = margin(10, 10, 10, 10),
  grid = TRUE, axis = FALSE, axis_size = 0.15, ticks = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin}{plot title family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin}{plot caption family, face, size and margin}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with [ggplot2::margin])}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of
\code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{axis \code{TRUE}, \code{FALSE}, [\code{xy}]}

\item{ticks}{ticks if `TRUE` add ticks}
}
\value{
The theme.
}
\description{
The SciencesPo flex theme for using with ggplot2 objects.
}
\examples{
# plot with small amount of padding
qplot(1:10, (1:10)^2) +
 theme_flex(grid="Y")

qplot(1:10, (1:10)^2) +
theme_flex(axis='xy', axis_size=.75)

# Check that it is a complete theme
# attr(theme_flex(), "complete")

\dontrun{
library(ggplot2)
library(dplyr)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x="Fuel effiiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 scatterplot example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_flex() +
  scale_color_flex()

# seminal bar chart

update_geom_font_defaults()

count(mpg, class) \%>\%
  ggplot(aes(class, n)) +
  geom_col() +
  geom_text(aes(label=n), nudge_y=3) +
  labs(x="Fuel efficiency (mpg)", y="Weight (tons)",
       title="Seminal ggplot2 bar chart example",
       subtitle="A plot that is only useful for demonstration purposes",
       caption="Brought to you by the letter 'g'") +
  theme_flex(grid="Y") +
  theme(axis.text.y=element_blank()) +
  scale_color_flex()


count(mpg, class) \%>\%
mutate(n=n*2000) \%>\%
arrange(n) \%>\%
mutate(class=factor(class, levels=class)) \%>\%
ggplot(aes(class, n)) +
geom_col() +
geom_text(aes(label = scales::comma(n)), hjust=0, nudge_y=2000) +
scale_y_comma(limits=c(0,150000)) +
coord_flip() +
labs(x = "Fuel effiiency (mpg)", y = "Weight (tons)",
   title="Seminal ggplot2 column chart example with commas",
   subtitle="A plot that is only useful for demonstration purposes,
   esp since you'd never\\nreally want direct labels and axis labels",
   caption="Brought to you by the letter 'g'") +
theme_flex(grid="X")



ggplot(brpopage, aes(x=Year, y=Thousands, fill=AgeGroup)) +
geom_area() +
scale_fill_flex("colorblind") +
scale_x_continuous(expand=c(0,0)) +
scale_y_comma() +
labs(title="Age distribution of population in Brazil, 1900-2010",
   subtitle="Example data from the R SciencesPo Cookbook",
      caption="Source: SciencesPo Cookbook") +
 theme_flex(grid="XY") +
 theme(axis.text.x=element_text(hjust=c(0, 0.5, 0.5, 0.5, 1))) +
 theme(legend.position="bottom")
}

}

