% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_fte}
\alias{theme_fte}
\title{Themes for ggplot graphs}
\usage{
theme_fte(base_size = 13, base_family = "sans", horizontal = TRUE,
  dkpanel = FALSE)
}
\description{
Theme for plotting  with ggplot.
}

