% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_map}
\alias{theme_map}
\title{Theme for mapping

Theme for plotting  with ggplot}
\usage{
theme_map(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Enables to set the font size of all text elements.}

\item{base_family}{Enables to set the font family of all text elements.}
}
\description{
Theme for mapping

Theme for plotting  with ggplot
}

