% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/THEME.R
\name{theme_osa}
\alias{theme_osa}
\title{Theme for mapping  with ggplot}
\usage{
theme_osa(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{Enables to set the font size of all text elements.}

\item{base_family}{Enables to set the font family of all text elements.}
}
\description{
Theme for mapping  with ggplot
}

