% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_plus}
\alias{theme_plus}
\title{ggplot2 theme}
\usage{
theme_plus(base_size = 12, legend_position = "right")
}
\arguments{
\item{base_size}{base font size}

\item{legend_position}{position of the legend ("none", "left", "right", "bottom",
"top", or two-element numeric vector)}
}
\description{
Theme intended to make ggplot2 more readable when used
in presentation or papers. Background and major grid lines were dimed
and minor grid lines removed to focus the attention on the data.
}
\examples{
qplot(1:10, (1:10)^2) + theme_plus()

# Check that it is a complete theme
attr(theme_plus(), "complete")
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}
\seealso{
\code{\link{theme_flex}}, \code{\link{theme_pub}}

Other themes darkside: \code{\link{theme_darkside}}
}
\keyword{ggplot2}

