% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\encoding{UTF-8}
\name{theme_pub}
\alias{theme_pub}
\title{The Default Theme}
\usage{
theme_pub(legend = "bottom", base_size = 12, base_family = "",
  horizontal = FALSE, line_width = 0.5, axis_line = FALSE)
}
\arguments{
\item{legend}{enables to set legend position, default is "bottom".}

\item{base_size}{overall font size. Default is 14.}

\item{base_family}{a name for default font family.}

\item{horizontal}{logical. Horizontal axis lines?}

\item{line_width}{default line size.}

\item{axis_line}{enables to set x and y axes.}
}
\value{
The theme.
}
\description{
After loading the SciencesPo package, this theme will be
set to default for all subsequent graphs made with ggplot2.
}
\examples{
PreviewTheme() + theme_pub()

# Anscombe data
dat <- data.frame()
for(i in 1:4)
dat <- rbind(dat, data.frame(set=i, x=anscombe[,i], y=anscombe[,i+4]))

ggplot(dat, aes(x, y)) +
geom_point(size=5, color="red", fill="orange", shape=21) +
geom_smooth(method="lm", fill=NA, fullrange=TRUE) +
facet_wrap(~set, ncol=2) +
theme_pub(base_family='serif') +
theme(plot.background=element_rect(fill="#f7f7f7")) +
theme(panel.background=element_rect(fill="#f7f7f7"))

}
\seealso{
\code{\link[ggplot2]{theme}}, \code{\link{theme_538}}, \code{\link{theme_blank}}.
}

