% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\encoding{UTF-8}
\name{ThemePreview}
\alias{PreviewTheme}
\alias{ThemePreview}
\title{Preview ggplot2 themes}
\usage{
ThemePreview()

PreviewTheme()
}
\description{
Used to preview ggplot2 themes.
}
\examples{
ThemePreview()
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}

