% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\docType{data}
\name{themes_data}
\alias{themes_data}
\title{Palette data for the themes used by package}
\format{A \code{list}.}
\usage{
themes_data
}
\description{
Data used by the palettes in the package.
}
\keyword{datasets}

