% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{time.data.frame}
\alias{time.data.frame}
\title{A data.frame with time series parameters}
\usage{
\method{time}{data.frame}(timevar, format, x = NULL, tz = "GMT")
}
\arguments{
\item{timevar}{the time variable}

\item{format}{the time format}

\item{x}{the data variable to be converted to a time series}

\item{tz}{Time zone code, default is \dQuote{GMT}}
}
\description{
Return a \code{data.frame} with time parameters and time variable
}
\details{
A data frame containing time variable parameters and measuring variable
}
\examples{
data(ssex)
myts <- time.data.frame(timevar = ssex$Date, format = "\%Y", x= ssex$Favor)
view(myts)

data(us2012)
Obama.ts <- time.data.frame(us2012[,3], "\%Y-\%m-\%d", us2012[,8])
view(Obama.ts)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\link{tsCollapse}
}
\keyword{Data-Manipulation}

