% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/time.plot.R
\name{timeplot}
\alias{timeplot}
\title{Make the ggplot2 version of TS plots}
\usage{
timeplot(ts, ylab = "", ylim = c(-1, 1), ci = 0.95, ...)
}
\arguments{
\item{ts}{The TS object.}

\item{ylab}{The \code{y-axis} title.}

\item{ylim}{The \code{y-axis} limits.}

\item{ci}{The desired confidence interval.}

\item{\dots}{Ignored parameters passed to ggplot.}
}
\description{
The function produces TS plots using ggplot.
}
\examples{
ts.sim <- stats::arima.sim(n = 100, list(ma=0.8), innov=rnorm(100))
 timeplot(ts.sim)
}

