% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic}
\format{2201 observations, 4 variables.}
\source{
British Board of Trade Inquiry Report (1990). \emph{Report on the Loss of the `Titanic' (S.S.)"}, Gloucester, UK: Allan Sutton Publishing.
}
\usage{
data(titanic)
}
\description{
Population at Risk and Death Rates for an Unusual Episode. For each person on board the fatal maiden voyage of the ocean liner Titanic, this dataset records sex, age [adult/child], economic status [first/second/third class, or crew] and whether or not that person survived.


\itemize{
\item CLASS Class (0 = crew, 1 = first, 2 = second, 3 = third)
\item AGE Age   (1 = adult, 0 = child)
\item SEX Sex   (1 = male, 0 = female)
\item SURVIVED Survived (1 = yes, 0 = no)
 }
}
\note{
There is not complete agreement among primary sources as to the exact numbers on board, rescued, or lost. \bold{STORY BEHIND THE DATA:} The sinking of the Titanic is a famous event, and new books are still being published about it.  Many well-known facts--from the proportions of first-class passengers to the "women and children first" policy, and the fact that that policy was not entirely successful in saving the women and children in the third class--are reflected in the survival rates for various classes of passenger.  These data were originally collected by the British Board of Trade in their investigation of the sinking.
}
\references{
Dawson (1995). "The `Unusual Episode' Data Revisited" in the \emph{Journal of Statistics Education}.
}
\keyword{datasets}

