% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\encoding{UTF-8}
\name{titleCase}
\alias{titleCase}
\title{Make Proper Case}
\usage{
titleCase(x)
}
\arguments{
\item{x}{is a character vector.}
}
\description{
Simply replaces all caps strings into proper case strings.
}
\examples{
titleCase("THIS SHOULDN'T BE ALL CAPITAL LETTERS")

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

