% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transform.R
\name{Transform}
\alias{Destring}
\alias{Factorize}
\alias{Transform}
\title{Functions for transformation of variables.}
\usage{
Factorize(.data, vars)

Destring(.data, vars)
}
\arguments{
\item{.data}{data.table or data.frame object.}

\item{vars}{Character or vector of characters with variables names to transform.}
}
\value{
Data.table object with transformed variables.
}
\description{
Functions for transformation of variables.
}
\examples{
Factorize(mtcars, c("cyl", "mpg"))
}

