% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{trim}
\alias{trim}
\title{Trim white spaces}
\usage{
trim(x, delim = " ")
}
\arguments{
\item{x}{is a character vector.}

\item{delim}{is the delimiter, default is white spaces \code{" "}}
}
\description{
Simply trims spaces from the start, end, and within of a string
}

