% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{twins}
\alias{twins}
\title{Burt's twin data}
\format{This data frame contains the following columns:}
\source{
Burt, C. (1966). The genetic estimation of differences in intelligence:
 A study of monozygotic twins reared together and apart. Br. J. Psych., 57, 147-153.
}
\usage{
data(twins)
}
\description{
The given data are IQ scores from identical twins; one raised in a foster home, and the other raised by birth parents.

\itemize{
\item C Social class, C1=high, C2=medium, C3=low, a factor.
\item IQb biological.
\item IQf foster.
 }
}
\references{
Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}

