% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOMENTS.R
\name{ucss}
\alias{ucss}
\alias{ucss.data.frame}
\alias{ucss.default}
\title{Uncorrected Sum of Squares}
\usage{
ucss(x, na.rm = TRUE)

\method{ucss}{default}(x, na.rm = TRUE)

\method{ucss}{data.frame}(x, na.rm = TRUE)
}
\arguments{
\item{x}{An \R object.}

\item{na.rm}{A logical value indicating whether \code{NA}
should be stripped before the computation proceeds.}
}
\description{
Generate the uncorrected sum of squares.
}

