% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{unclass.data.frame}
\alias{unclass.data.frame}
\title{Unclass data frame}
\usage{
unclass.data.frame(vars, data = .data)
}
\arguments{
\item{vars}{The variable(s).}

\item{data}{The data object.}
}
\description{
Unclass data frame
}

