% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unfactor.R
\name{unfactor}
\alias{unfactor}
\title{Unfactor columns}
\usage{
unfactor(.data)
}
\arguments{
\item{.data}{The data.frame containing factor columns.}
}
\description{
Try to unfactorize columns of a data.frame object.
}
\examples{
unfactor(iris)
}

