% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{unions}
\alias{unions}
\title{Poll attitudes towards British trade unions}
\format{A \code{data.frame} object with \Sexpr[eval=TRUE,echo=FALSE,results=text]{ncol(SciencesPo::unions)} variables and \Sexpr{nrow(SciencesPo::unions)} observations.
\itemize{
\item Date Month of the poll \code{Aug-77} to \code{Sep-79}.
\item AgreePct Percent who agree (unions have too much power).
\item DisagreePct Percent who disagree.
\item NetSupport DisagreePct-AgreePct.
\item Months Months since August 1975.
\item Late \code{1}=after 1986 or \code{0}=before 1986.
\item Unemployment Unemployment rate.
}
 @source
 \url{http://www.ipsos-mori.com/researchpublications/researcharchive/poll.aspx?oItemID=94}}
\usage{
data(unions)
}
\description{
The British polling company Ipsos MORI conducted several
opinion polls in the UK between 1975 and 1995 in which they asked whether
people agree or disagree with the statement "Trade unions have too much
power in Britain today".
}
\keyword{datasets}

