% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unique_obs.R
\name{uniqueObs}
\alias{uniqueObs}
\title{Number of unique elements in a vector.}
\usage{
uniqueObs(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{passed to \code{\link{unique}}}
}
\description{
A wrapper around \code{length(unique(x))}
}
\examples{
x<-sample(1000, rep=TRUE)
 uniqueObs(x)
}
\keyword{Describe}

