% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{Units}
\alias{Units}
\title{Measurement System Units}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::Units)} variables and \Sexpr{nrow(SciencesPo::Units)} observations.}
\usage{
data(Units)
}
\description{
A dataset with measurement system units.

\itemize{
\item from A character defining the original unit.
\item to A character defining the target unit.
\item factor The factor to be applied in conversion.
\item description Some details about the measure.
}
}
\keyword{datasets}

