% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{unnest}
\alias{unnest}
\title{Unnest a Nested List}
\usage{
unnest(x)
}
\arguments{
\item{x}{A nested list}
}
\value{
A list, with no nesting, hopefully
}
\description{
Unnest nested lists made easy.
}
\examples{
# Unnest the list
# a nested list
mylist <- list(); inerlist <- list()
for(i in 1:5) {
  for(j in 1:5) {
   mylist[[j]] <- i*j
 }
 inerlist[[i]] <- mylist
}
unnest(inerlist)[[1]]
unnest(inerlist)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

