% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale.R
\encoding{UTF-8}
\name{unscale}
\alias{unscale}
\title{Unscale Data}
\usage{
unscale(x)
}
\arguments{
\item{x}{a scaled object}
}
\value{
Unscaled parameters
}
\description{
Unscale objects by dropping \code{scaled:scale} and \code{scaled:center} attributes from parameters.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Models}

