% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/untable.R
\encoding{UTF-8}
\name{untable}
\alias{untable}
\title{Untable}
\usage{
untable(x, ...)
}
\arguments{
\item{x}{The table object as a data.frame, table, or, matrix.}

\item{\dots}{Extra parameters.}
}
\description{
Method for untable, convert from summarized data to long.
}
\examples{
if (interactive()) {
gss <- data.frame(
expand.grid(sex=c("female", "male"),
party=c("dem", "indep", "rep")),
count=c(279,165,73,47,225,191))

print(gss)

# Thwn expand it:
GSS <- untable(gss, freq="count")
head(GSS)
}
}

