% Generated by roxygen2 (4.0.2): do not edit by hand
\name{use}
\alias{use}
\title{Attach exclusively various file formats}
\usage{
use(file, dataFrame = .data, clear = TRUE, spss.missing = TRUE,
  tolower = TRUE)
}
\arguments{
\item{file}{}

\item{dataFrame}{}

\item{clear}{}

\item{spss.missing}{, default is \code{spss.missing = TRUE}.}

\item{tolower}{, default is \code{tolower = TRUE}.}
}
\description{
This works rigoroulsy as the \pkg{epicalc}'s \code{use} function, though limited for the file formats it can read. Fundamentally, it replaces the command attach of R and save an object with extension \code{.data}, which becomes the default dataset. All other \code{data.frames} will be detached, by the time of using the \code{use} function, unless the argument \code{clear=FALSE} is specified.
}
\examples{
data(ssex)
use(ssex)
}

