% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{user.prompt}
\alias{user.prompt}
\title{Prompt for User Action}
\usage{
user.prompt(msg = NULL)
}
\arguments{
\item{msg}{a character-string, specifying a message to be displayed}
}
\value{
This function is used for its side effects
}
\description{
Prompt user to hit enter
}
\note{
This function is primarily used by SciencesPo scripts
}

