% Generated by roxygen2 (4.0.2): do not edit by hand
\name{userPrompt}
\alias{userPrompt}
\title{Prompt for User Action}
\usage{
userPrompt(msg = NULL)
}
\arguments{
\item{msg}{a character-string, specifying a message to be displayed}
}
\value{
This function is used for its side effects
}
\description{
Prompt user to hit enter
}
\note{
This function is primarily used by SciencesPo scripts
}

