% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{USPresidents}
\alias{USPresidents}
\title{US Presidents' height}
\format{This data.frame contains the following columns:}
\source{
Wikipedia: \url{https://en.wikipedia.org/wiki/United_States_presidential_election,_2012}
Wikipedia:\url{https://en.wikipedia.org/wiki/Heights_of_presidents_and_presidential_candidates_of_the_United_States}.
}
\usage{
data(USPresidents)
}
\description{
Heights of the US presidents and their main opponents.

\itemize{
\item election The election year.
\item winner The winner candidate.
\item winner.height The winner candidate's height in centimeters (cm).
\item winner.vote The popular vote support for the winner.
\item winner.party The winner's party.
\item opponent The main opponent candidate.
\item opponent.height The opponent candidate's height in centimeters (cm).
\item opponent.vote The popular vote support for the main opponent candidate.
\item opponent.party The opponent's party.
\item turnout The percentage of the electorate turnout.
 }
}
\keyword{datasets}

