% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label.R
\name{values}
\alias{values}
\title{List the Values of a Variable}
\usage{
values(x, data = .data, ...)
}
\arguments{
\item{x}{Variable for which to construct the histogram and density plots.}

\item{data}{Data frame that contains the variable of interest, default is \code{.data}.}

\item{\dots}{Other parameter values for as defined processed by \code{\link{print}}, including \code{digits}.}
}
\description{
List the values of a variable from the global environment or a data frame.
}
\examples{
.data <- data.frame(AgeGroup=factor(
 1:9,
 labels= c("Younger than 16", "16-24",
           "25-34", "35-44", "45-54",
           "55-64", "65-74",
           "Older than 74", "(blank)"),
 ordered=TRUE),
 No=c(34, 2079, 2585, 1593,
      1274, 802, 291, 78, 13),
 Yes=c(18, 1970, 4035, 2328,
       1707, 924, 386, 68, 4))

values(Yes)
}

