% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var.order.R
\encoding{UTF-8}
\name{var.order}
\alias{var.order}
\title{Move variable order in a data.frame}
\usage{
var.order(data, var, where)
}
\arguments{
\item{data}{The data object}

\item{var}{The variable to be moved.}

\item{where}{The new location for the variable (var).}
}
\description{
Move variable order in a data.frame
}
\examples{
var.order(ssex, "DK", "Oppose")
}

