% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{view}
\alias{view}
\title{View random n elements of an object.}
\usage{
view(obj = .data, n = 10, random = TRUE, tail = FALSE,
  print.console = TRUE, ...)
}
\arguments{
\item{obj}{A matrix or data.frame object}

\item{n}{The number of rows to be shown}

\item{random}{If TRUE, observartions are sampled.}

\item{tail}{If TRUE, will show n lines from tail.}

\item{print.console}{If TRUE, dada will be printed on console.}

\item{\dots}{Some additional parameters.}
}
\description{
Provide a sly view of the data by randomly drawn observations, instead of showing only the first \code{head()} or the last \code{tail()} rows of an object, though it is also possible to constrain for the first/last elements.
}
\examples{
use(titanic)
view(titanic)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Tables}

