% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vif}
\alias{vif}
\title{Variance Inflation Factor}
\usage{
vif(model, ...)
}
\arguments{
\item{model}{a model object}

\item{\dots}{Options to pass to}
}
\value{
A numeric value indicating the variance inflation in the model

#'
}
\description{
Extracts Variance Inflation Factor from a model of class \dQuote{lm}
}
\examples{
data(mtcars)
m1 <- lm(mpg ~ qsec + hp, data=mtcars)
vif(m1)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Models}

