% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vorder.R
\encoding{UTF-8}
\name{vorder}
\alias{vorder}
\title{Move variable order in a data.frame}
\usage{
vorder(data, var, where)
}
\arguments{
\item{data}{The data object}

\item{var}{The variable to be moved.}

\item{where}{The new location for the variable (var).}
}
\description{
Move variable order in a data.frame
}
\examples{
vorder(ssex, "DK", "Oppose")
}

