% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{vote}
\alias{vote}
\title{Voting correlations}
\format{A \code{12x12 matrix} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::vote)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::vote)} observations.}
\source{
Torben Iversen and David Soskice (2006). Electoral institutions and
the politics of coalitions: Why some democracies redistribute more than
others. \emph{American Political Science Review}, 100, 165-81. Table A2.
}
\usage{
data(vote)
}
\description{
A dataset with voting correlations of traits, where each
trait is measured for 17 developed countries (Europe, US,
Japan, Australia, New Zealand).
}
\references{
Using Graphs Instead of Tables.
\url{http://www.tables2graphs.com/doku.php?id=03_descriptive_statistics}
}
\keyword{datasets}

