% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTS.R
\name{wald.test}
\alias{wald.test}
\title{Wald Test for Model Coefficients}
\usage{
wald.test(sigma, b, terms = NULL, L = NULL, H0 = NULL, df = NULL,
  verbose = FALSE)
}
\description{
Computes a Wald chi-squared test for 1 or more coefficients, given their variance-covariance matrix.
}
\details{
Sigma	A var-cov matrix, usually extracted from one of the fitting functions (e.g., lm, glm, ...).

b A vector of coefficients with var-cov matrix Sigma. These coefficients are usually extracted from one of the fitting functions available in R (e.g., lm, glm,...).

Terms	An optional integer vector specifying which coefficients should be jointly tested, using a Wald chi-squared or F test. Its elements correspond to the columns or rows of the var-cov matrix given in Sigma. Default is NULL.
L	An optional matrix conformable to b, such as its product with b i.e., L %*% b gives the linear combinations of the coefficients to be tested. Default is NULL.
H0	A numeric vector giving the null hypothesis for the test. It must be as long as Terms or must have the same number of columns as L. Default to 0 for all the coefficients to be tested.
df	A numeric vector giving the degrees of freedom to be used in an F test, i.e. the degrees of freedom of the residuals of the model from which b and Sigma were fitted. Default to NULL, for no F test. See the section Details for more information.
verbose	A logical scalar controlling the amount of output information. The default is FALSE, providing minimum output.
x Object of class "wald.test"
digits	
Number of decimal places for displaying test results. Default to 2.
... Additional arguments to print.
}

