% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{weather}
\alias{weather}
\title{NY Weather 1995-2014}
\format{A data frame with 7290 observations on the following 4 variables.}
\source{
Archived by the National Climatic Data Center (NCDC) and downloaded at: \url{http://academic.udayton.edu/kissock/http/Weather/gsod95-current/NYNEWYOR.txt}
}
\usage{
data(weather)
}
\description{
Data represents average daily temperatures.

\itemize{
 \item month the month recorded (starts on January, 1995)
  character vector for date.
 \item day the day of the month.
 \item year the year.
 \item temp the average daily temperatures ( \dQuote{24 hour} and \eqn{(min + max)/2}).
 }
}
\note{
In the original files, missing data is recorded as \dQuote{-99}, so must be replaced.
}
\references{
Edward Tufte \emph{New York Weather Chart}, January 4, 2004, p. A-15.
}
\keyword{datasets}

