% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weighted.variance.R
\name{weighted.variance}
\alias{weighted.variance}
\title{Weighted Variance}
\usage{
weighted.variance(x, w)
}
\arguments{
\item{x}{the varaible.}

\item{w}{the variance.}
}
\description{
Weighted Variance Formula
}
\keyword{Stats}

