% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{WeightedCorrelation}
\alias{WeightedCorrelation}
\title{Compute Weighted Correlations}
\usage{
WeightedCorrelation(x, y = NULL, weights = NULL)
}
\arguments{
\item{x}{a matrix or vector to correlate with \code{y}.}

\item{y}{a matrix or vector to correlate with \code{x}. If \code{y} is NULL, \code{x} will be used instead.}

\item{weights}{an optional vector of weights to be used to determining the weighted mean and variance for calculation of the correlations.}
}
\description{
Compute the weighted correlation.
}
\examples{
 x <- sample(10,10)
 y <- sample(10,10)
 w <- sample(5,10, replace=TRUE)

WeightedCorrelation(x, y, w)

}

