% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{weightedGeometricMean}
\alias{weightedGeometricMean}
\title{Calculate a weighted geometric mean}
\usage{
weightedGeometricMean(x, w, ...)
}
\arguments{
\item{x}{An R object.}

\item{...}{further arguments passed to \code{\link{prod}}}
}
\description{
Calculate a weighted geometric mean
}
\seealso{
\code{\link{weightedHarmonicMean}}, \code{\link{weighted.mean}}, \code{\link{geometricMean}}, \code{\link{harmonicMean}} and \code{\link{mean}}
}

