% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{weightedHarmonicMean}
\alias{weightedHarmonicMean}
\title{Calculate a weighted harmonic mean}
\usage{
weightedHarmonicMean(x, w, ...)
}
\arguments{
\item{x}{An invertable R object.}

\item{...}{further arguments passed to \code{\link{sum}}}
}
\description{
Calculate a weighted harmonic mean
}
\seealso{
\code{\link{weightedGeometricMean}}, \code{\link{weighted.mean}}, \code{\link{harmonicMean}} \code{\link{geometricMean}} and \code{\link{mean}}
}

