% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{WeightedMean}
\alias{WeightedMean}
\title{Computes Weighted Mean}
\usage{
WeightedMean(x, weights = NULL, normwt = "ignored", na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{weights}{a numeric vector of weights of \code{x}.}

\item{normwt}{ignored at the moment.}

\item{na.rm}{a logical, if \code{TRUE}, missing data will be dropped.  If \code{na.rm = FALSE}, missing data will return an error.}
}
\description{
Compute the weighted mean of data.
}
\examples{
x <- sample(10,10)
w <- sample(5,10, replace=TRUE)

WeightedMean(x, w)
}

