% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{WeightedStdz}
\alias{WeightedStdz}
\title{Computes Weighted Standardized Values}
\usage{
WeightedStdz(x, weights = NULL)
}
\arguments{
\item{x}{a vector for which a set of standardized values is desired.}

\item{weights}{a vector of weights to be used to determining weighted values of \code{x}.}
}
\description{
Computes weighted standardized values of data.
}
\examples{
 x <- sample(10,10)
 w <- sample(5,10, replace=TRUE)

WeightedStdz(x, w)
}

