% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\name{weightedVariance}
\alias{weightedVariance}
\title{Weighted Variance}
\usage{
weightedVariance(x, w)
}
\arguments{
\item{x}{the varaible}

\item{w}{the variance}
}
\description{
Weighted Variance Formula
}

