% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{which.evens}
\alias{evens}
\alias{odds}
\alias{which.evens}
\title{Return the odd and even values from a vector}
\usage{
which.evens(x)
}
\arguments{
\item{x}{Integer vector}
}
\value{
Returns an integer vector consisting of only the odd/even elements.
}
\description{
Takes an integer vector and returns every odd or even element
}
\examples{

x <- as.integer(c(1,2,3,4,5,6,7,8,9))
which.evens(x)
 which.odds(x)

}

