% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{wrap}
\alias{wrap}
\title{Wrap all related variables}
\usage{
wrap(data = .data)
}
\arguments{
\item{data}{is the .data object}
}
\description{
Try to  wrap all related variables into the existing .data
}

