% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write.delim}
\alias{write.delim}
\title{Writes a delimited text file}
\usage{
write.delim(.data, file, quote = FALSE, row.names = FALSE, sep = "\\t",
  ...)
}
\arguments{
\item{.data}{the data frame to be written}

\item{file}{the name of the file}

\item{quote}{If \code{TRUE}, any character or factor columns will be surrounded by double quotes. Default is \code{quote=FALSE}.}

\item{row.names}{if \code{TRUE}, the row names will be write to file. Default is \code{ row.names=FALSE}.}

\item{sep}{the field separator string.}

\item{\dots}{other uncommon arguments to write.table (ex: fileEncoding).}
}
\description{
Writes a delimited text file
}

