% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wtd.var.R
\name{wtd.var}
\alias{wtd.var}
\title{Weighted Variance}
\usage{
wtd.var(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{the variable.}

\item{w}{the variance.}

\item{na.rm}{A logical if NA should be disregarded.}
}
\description{
Weighted Variance Formula
}
\examples{
wt=c(1.23, 2.12, 1.23, 0.32, 1.53, 0.59, 0.94, 0.94, 0.84, 0.73)
x = c(5, 5, 4, 4, 3, 4, 3, 2, 2, 1)
wtd.var(x, wt)
}
\keyword{Stats}

