% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{xfill}
\alias{xfill}
\title{Fills in missing values within clusters.}
\usage{
xfill(x, by = NULL)
}
\arguments{
\item{x}{the variable to be filled.}

\item{by}{the cluster variable.}
}
\description{
\code{xfill} 'fills in' static variables, replacing missing values in a cluster with the unique non-missing value within that cluster.
}
\examples{
df = data.frame(year = c(2000, 2000, 2000, 2001, 2001, 2002, 2002, 2003), value= c("yes", NA, NA, NA, "yes", "no", NA, NA) )

#xfill(value, by=year)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

