% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/z.scores.R
\encoding{UTF-8}
\name{z.scores}
\alias{z.scores}
\title{Compute z-scores}
\usage{
z.scores(x, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\description{
Compute z-scores
}
\examples{
x <- sample(10)
z.scores(x)
}

