% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{zscore}
\alias{zscore}
\title{Compute z-scores}
\usage{
zscore(x, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\description{
Compute z-scores
}
\examples{
x <- sample(10)
zscore(x)
}

