% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{adj.residuals}
\alias{adj.residuals}
\title{Adjusted Residuals}
\usage{
adj.residuals(object, ...)
}
\arguments{
\item{object}{a model object of type \code{glm} or \code{lm}.}

\item{\dots}{Extra}
}
\description{
Adjusted Residuals
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

