% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{alpha}
\alias{alpha}
\title{A dataset that contains four test items}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::alpha)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::alpha)} observations.}
\usage{
data(alpha)
}
\description{
A dataset with four test items used in SPSS to to compute Cronbach's alpha.

\itemize{
\item q1:q4 Numeric items of a scale.
}
}
\keyword{datasets}

