% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{anonymize}
\alias{anonymize}
\alias{anonymize.default}
\title{To Make Anonymous a Data Frame}
\usage{
anonymize(.data, col.names = "V", row.names = "")

\method{anonymize}{default}(.data, col.names = "V", row.names = "")
}
\arguments{
\item{.data}{A vector or a data frame.}

\item{col.names}{A string for column names.}

\item{row.names}{A string for rown names, default is empty.}
}
\value{
An object of the same type as \code{.data}.
}
\description{
Replaces factor and character variables by a
combination of letters and numbers. Numeric columns are also transformed, see details.
}
\examples{
dt <- data.frame(
Z = sample(LETTERS,10),
X = sample(1:10),
Y = sample(c("yes", "no"), 10, replace = TRUE)
)
dt;

anonymize(dt)

}

