% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anonymize}
\alias{anonymize}
\title{Make Data Anonymous}
\usage{
anonymize(x, keep.names = TRUE)
}
\arguments{
\item{x}{A vector or a data frame}

\item{keep.names}{A logical argument. If \code{FALSE}, variable names will be replaced by Xs}
}
\value{
An object of the same type as \code{x}
}
\description{
This function replaces factor and character variables by a combination of letters and numbers, and numeric columns are also transformed.
}
\details{
By making difficult to recognize the original data while keeping the same data structure, this function is  quite useful for sharing data on help lists.
}
\examples{
# setup data
data(ssex)
anonymize(ssex)
anonymize(ssex, keep.names=FALSE)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Tables}

