% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atkinson.R
\name{Atkinson}
\alias{Atkinson}
\alias{Atkinson.default}
\title{Atkinson Index of Inequality}
\usage{
Atkinson(x, n = rep(1, length(x)), base = NULL, na.rm = FALSE, ...)

\method{Atkinson}{default}(x, n = rep(1, length(x)), base = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{a vector of data values of non-negative elements.}

\item{n}{a vector of frequencies of the same length as \code{x}.}

\item{base}{a parameter of the inequality measure (if \code{NULL}, the default parameter (0.5) of the respective measure is used).}

\item{na.rm}{logical. Should missing values be removed? Defaults is set to \code{FALSE}.}

\item{\dots}{additional arguements (currently ignored)}
}
\description{
Calculates the Atkinson index A. This inequality measure is especially good at determining which end of the distribution is contributing most to the observed inequality.
}
\examples{
if (interactive()) {
# generate a vector (of incomes)
x <- c(778, 815, 857, 888, 925, 930, 965, 990, 1012)

# compute Atkinson coefficient with base=0.5
Atkinson(x, base=0.5)
}
}
\references{
Cowell, F. A. (2000) Measurement of Inequality in Atkinson, A. B. / Bourguignon, F. (Eds): \emph{Handbook of Income Distribution}. Amsterdam.

Cowell, F. A. (1995) \emph{Measuring Inequality} Harvester Wheatshef: Prentice Hall.
}
\seealso{
\code{\link{Herfindahl}}, \code{\link{Rosenbluth}}, \code{\link{Gini}}. For more details see the \dQuote{Indices} vignette.
}

