% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{bhodrick93}
\alias{bhodrick93}
\title{Bekaert's and Hodrick's (1993) Data}
\format{A data frame with 778 observations on the following 11 variables. See Hayashi (2000) for details.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(bhodrick93)
}
\description{
Data set used by Bekaert and Hodrick (1993) on biases in the measurement of foreign exchange risk premiums.

\itemize{
 \item date A character vector for date.
 \item jyspot Price of USD in JY, spot.
 \item jyfwd Price of USD in JY, 30-day forward.
 \item jys30 Price of USD in JY, spot market at 30-day forward deliver/date.
 \item dmspot Price of USD in DM, spot.
 \item dmfwd Price of USD in DM, 30-day forward
 \item dms30 Price of USD in DM, spot market at 30-day forward deliver/date.
 \item bpspot Price of USD in BP, spot.
 \item bpfwd Price of USD in BP, 30-day forward.
 \item bps30 Price of USD in BP, spot market at 30-day forward deliver/date.
\item quote A numeric vector.
}
}
\references{
Bekaert, G., and Hodrick, R. J. (1993) On biases in the measurement of foreign exchange risk premiums. \emph{Journal of International Money and Finance,} \bold{12(2),} 115-138.
}
\keyword{datasets}

