% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DISTRIBUTIONS.R
\encoding{UTF-8}
\name{binomcdf}
\alias{binomcdf}
\title{Binomial cumulative distribution function}
\usage{
binomcdf(n, p, x)
}
\arguments{
\item{n}{the number of trials.}

\item{p}{a vector of probabilities.}

\item{x}{the number of success.}
}
\description{
Computes a binomial cdf at each of the values in \code{x} using the corresponding number of trials in \code{n} and probability of success for each trial in \code{p}.
}
\examples{
trials = 10
prob = c(.2,.25,.3,.35)
success = 4
binompdf(n = trials, p = prob, x = success)
}
\keyword{Distributions}

