% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{bootstrap}
\alias{bootstrap}
\title{Method for Bootstrap

This method is intended to be provides statistical models that support  bootstrapping.}
\usage{
bootstrap(x, ...)
}
\arguments{
\item{x}{a a vector or a fitted model object that will be used to produce bootstrapped parameters. Model objects are from the class \dQuote{glm} or \dQuote{lm}.}

\item{...}{unspecified parameters}
}
\value{
a list with the \dQuote{alpha} and \dQuote{beta} slots set. Note that \dQuote{alpha} corresponds to ancillary parameters and \dQuote{beta} corresponds to systematic components of the model.
}
\description{
Method for Bootstrap

This method is intended to be provides statistical models that support  bootstrapping.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

