\name{cgreene76}
\alias{cgreene76}
\docType{data}
\title{Christensen's and Greene's (1976) Data
}
\description{Data set used by Christensen and Greene (1976) on economies of scale in US electric power generation.
}
\usage{data(cgreene76)}
\format{
  A data frame with 99 observations on the following 8 variables. See Hayashi (2000) for more details.
  \describe{
    \item{\code{firmid}}{Observation id}
    \item{\code{costs}}{Costs in 1970, MM USD}
    \item{\code{output}}{Output, million KwH}
    \item{\code{plabor}}{Price of labor}
    \item{\code{pkap}}{ Price of capital}
    \item{\code{pfuel}}{Price of fuel}
    \item{\code{labshr}}{Labor's cost share}
    \item{\code{kapshr}}{Capital's cost share}
  }
}

\source{Hayashi, F. (2000) \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\references{Christensen, L. R., and Greene, W. H. (1976) Economies of scale in US electric power generation. \emph{The Journal of Political Economy,} 655--676.
}
\examples{
data(cgreene76)
}
\keyword{datasets}
