% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{clear}
\alias{clear}
\title{Clear Memory of All Objects}
\usage{
clear(obj = NULL, keep = TRUE)
}
\arguments{
\item{obj}{The object (as a string) that needs to be removed (or kept)}

\item{keep}{Should \code{obj} be kept (i.e., everything but \code{obj} removed)? Or dropped?}
}
\description{
This function is a wrapper for the command \code{rm(list=ls())}.
}
\examples{
# create objects
a=1; b=2; c=3; d=4; e=5
# remove d
clear("d", keep=FALSE)
ls()
# remove all but a and b
clear(c("a", "b"), keep=TRUE)
ls()
}
\author{
Daniel Marcelino
}

