% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{colorblind_pal}
\alias{colorblind_pal}
\title{Color blind color palette (discrete)}
\usage{
colorblind_pal()
}
\description{
Color palettes for color blind.
}
\examples{
scales::show_col(colorblind_pal()(8))
}
\seealso{
Other colour colorblind: \code{\link{scale_color_colorblind}},
  \code{\link{scale_colour_colorblind}},
  \code{\link{scale_fill_colorblind}}
}

