% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountMissing.R
\name{CountMissing}
\alias{CountMissing}
\title{Count missing data.}
\usage{
CountMissing(.data, vars = NULL, prop = TRUE, exclude.complete = TRUE)
}
\arguments{
\item{.data}{the data frame.}

\item{vars}{vector with name of variables.}

\item{prop}{logical. If this is \code{TRUE} will show share of missing data by variable.}

\item{exclude.complete}{Logical. Exclude complete variables from the output.}
}
\description{
Function to count missing data.
}
\examples{

CountMissing(marriage, prop = FALSE)
}

