% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{cramer}
\alias{cramer}
\title{Cramer's V for a contingency table}
\usage{
cramer(x, y)
}
\arguments{
\item{x}{one vector}

\item{y}{the other vector}
}
\value{
The table's Cramer's V.
}
\description{
Produce the Cramér V / Phi test using two vectors with more than 2 levels.
}
\details{
Compute Cramer's V for a resulting table with more than 2x2 fields.
}
\examples{
x =sample(1:2, 30, TRUE);
 y= sample(1:3, 30, TRUE)
 cramer(x, y)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

