% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstable.R
\name{Crosstab}
\alias{Crosstab}
\title{Cross Tabulation}
\usage{
Crosstab(x, y, digits = 3)
}
\arguments{
\item{x, y}{the variables for the cross tabulation.}

\item{digits}{number of digits for rounding proportions.}
}
\description{
\code{Crosstab} is a modified version of \code{\link[gmodels]{CrossTable}} (\pkg{gmodels}).
\code{Crosstab} will print a summary table with cell counts and column proportions (similar to STATA's
\code{tabulate}.
}

