% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Crosstab.R
\encoding{UTF-8}
\name{Crosstab}
\alias{Crosstab}
\title{Cross-tabulation}
\usage{
Crosstab(data, rowVar, colVar, freqVar = NULL, observed = TRUE,
  expected = FALSE, stdresid = FALSE, totPercent = FALSE,
  rowPercent = FALSE, colPercent = FALSE, chisq = FALSE, phi = FALSE,
  cramersv = FALSE, contingency = FALSE)
}
\arguments{
\item{data}{The data.frame.}

\item{rowVar}{The row variable.}

\item{colVar}{The column variable.}
}
\description{
Makes a contingency table and computes chi-square test of independence, phi coefficient, cramers v and contingency coefficient.
}

