% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\name{crosstab}
\alias{crosstab}
\title{Cross-tabulation}
\usage{
crosstab(..., row.vars = NULL, col.vars = NULL, type = NULL,
  style = "wide", decimals = NULL, percent = TRUE, margins = TRUE,
  subtotals = TRUE)
}
\arguments{
\item{row.vars}{the row variable(s).}

\item{col.vars}{the row variable(s).}

\item{type}{wether \emph{frequency count} \code{"f"}, \emph{row percentages} \code{"r"}, \emph{column percentages} \code{"c"}, or \emph{total percentages} \code{"t"}.}

\item{style}{a style for table, either \code{"wide"} or \code{"long"}.}

\item{decimals}{an integer for decimal places.}

\item{percent}{wether to show percentiles or not. Default is \code{TRUE}}

\item{margins}{wether to add margins or not. Default is \code{TRUE}}

\item{subtotals}{wether to show subtotals or not. Default is \code{TRUE}}

\item{\dots}{the data paremeters.}
}
\description{
\code{crosstab} produces all possible two-way tabulations of the variables specified.
}
\note{
Adapted from Dr Paul Williamson, Dept. of Geography and Planning, School of Environmental Sciences, University of Liverpool, UK, who adapted from \code{ctab()}.
}
\examples{
data(titanic)

crosstab(titanic, row.vars = "AGE", col.vars = "SEX", type = "f")
# Row percentages
crosstab(titanic, row.vars = "AGE", col.vars = "SEX", type = "r")

# Column percentages
crosstab(titanic, row.vars = "AGE", col.vars = "SEX", type = "c")

# Joint percentages (sums to 100 within final two table dimensions)
crosstab(titanic, row.vars = c("AGE","SEX"), col.vars = "SURVIVED", type = "c")

# Total percentages (sums to 100 across entire table)
crosstab(titanic, row.vars = c("AGE","SEX"), col.vars = "SURVIVED", type = "t")
# Style = 'long'
crosstab(titanic, row.vars = c("AGE","SEX"), col.vars = "SURVIVED", type = "t",style = "long", margins = FALSE)

# Style = 'wide' [default]
crosstab(titanic, row.vars = "AGE", col.vars = "SURVIVED", type = "t", style = "long", margins = FALSE)
}

