% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crosstable.R
\encoding{UTF-8}
\name{crosstable}
\alias{crosstable}
\title{Cross-tabulation}
\usage{
crosstable(..., deparse.level = 2)
}
\arguments{
\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. If 0, middle 2 rownames, if 1, 3 rownames, if 2, 4 rownames (default).}

\item{\dots}{The data paremeters.}
}
\value{
Well-formatted cross tabulation. Also can genarate latex syntax of cross tabulation.
}
\description{
\code{crosstab} produces all possible two-way tabulations of the variables specified.
}
\examples{
with(titanic, crosstable( SEX, AGE))
with(titanic, crosstable( SEX, AGE, SURVIVED))
}

