% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossTable.R
\encoding{UTF-8}
\name{crossTable}
\alias{crossTable}
\alias{crossTable.default}
\title{Cross-tabulation}
\usage{
crossTable(.data, rowVar, colVar, freqVar = NULL, observed = TRUE,
  expected = FALSE, stdresid = FALSE, totPercent = FALSE,
  rowPercent = FALSE, colPercent = FALSE, chisq = FALSE, phi = FALSE,
  cramersv = FALSE, contingency = FALSE)

\method{crossTable}{default}(.data, rowVar, colVar, freqVar = NULL,
  observed = TRUE, expected = FALSE, stdresid = FALSE,
  totPercent = FALSE, rowPercent = FALSE, colPercent = FALSE,
  chisq = FALSE, phi = FALSE, cramersv = FALSE, contingency = FALSE)
}
\arguments{
\item{.data}{The data.frame.}

\item{rowVar}{The row variable.}

\item{colVar}{The column variable.}

\item{freqVar}{The frequency variable.}

\item{observed}{If TRUE, observed values will be included.}

\item{expected}{If TRUE, expected values will be included.}

\item{stdresid}{If TRUE, standardized residual will be included.}

\item{totPercent}{If TRUE, total proportions will be included.}

\item{rowPercent}{If TRUE, row proportions will be included.}

\item{colPercent}{If TRUE, column proportions will be included.}

\item{chisq}{If TRUE, the results of chi-square test will be included.}

\item{phi}{}

\item{cramersv}{}

\item{contingency}{}
}
\description{
Makes a contingency table and computes chi-square test of independence, phi coefficient, cramers v and contingency coefficient.
}
\examples{
titanic \%>\% crossTable("SEX", "AGE")

crossTable(titanic, "SEX", "AGE",
rowPercent=TRUE,
 expected=TRUE,
 stdresid=TRUE)

# Two variables
titanic \%>\%
crossTable(c("SEX","AGE"), "SURVIVED")

#' # Agresti (2002), table 3.11, p. 106
# 1992 General Social Survey- Sex and Party affiliation
gss <- data.frame(
   expand.grid(sex=c("female", "male"),
   party=c("dem", "indep", "rep")),
   count=c(279,165,73,47,225,191))

crossTable(gss, "sex", "party", "count")
}

