% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crosstable.R
\encoding{UTF-8}
\name{crosstable}
\alias{crosstable}
\alias{crosstable.default}
\title{Cross-tabulation}
\usage{
crosstable(.data, ..., row = TRUE, col = TRUE, deparse.level = 2)

\method{crosstable}{default}(.data, ..., row = TRUE, col = TRUE,
  deparse.level = 2)
}
\arguments{
\item{.data}{The data object.}

\item{row}{\code{TRUE}.}

\item{col}{\code{TRUE}.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. If 0, middle 2 rownames, if 1, 3 rownames, if 2, 4 rownames (default).}

\item{\dots}{The data paremeters.}
}
\value{
A cross-tabulated object.
}
\description{
\code{crosstable} produces all possible two-way tabulations of the variables specified.
}
\examples{
crosstable(titanic, SEX, AGE)

titanic \%>\% crosstable( SEX, AGE, SURVIVED)
}
\keyword{Exploratory}

