% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{crosstable}
\alias{crosstable}
\alias{crosstable.default}
\title{Cross-tabulation}
\usage{
crosstable(..., digits = 2, latex = FALSE, deparse.level = 2)

\method{crosstable}{default}(..., digits = 2, latex = FALSE,
  deparse.level = 2)
}
\arguments{
\item{digits}{The number of digits required, default is 2.}

\item{latex}{If the output is to be printed as latex.}

\item{deparse.level}{Integer controlling the construction of labels in the case of non-matrix-like arguments. If 0, middle 2 rownames, if 1, 3 rownames, if 2, 4 rownames (default).}

\item{\dots}{The variables for the cross tabulation.}

\item{row}{\code{TRUE}.}

\item{column}{\code{TRUE}.}
}
\value{
A cross-tabulated object.
}
\description{
\code{crosstable} produces all possible two-way and three-way tabulations of variables.
}
\examples{
with(titanic, crosstable( SEX, AGE, SURVIVED))

#' # Agresti (2002), table 3.10, p. 106
# 1992 General Social Survey--Race and Party affiliation
gss <- data.frame(
   expand.grid(Race=c("black", "white"),
   party=c("dem", "indep", "rep")),
   count=c(103,341,15,105,11,405))

df <- gss[rep(1:nrow(gss), gss[["count"]]), ]

with(df, crosstable(Race, party))

# Tea-Tasting Experiment data
 tea <- data.frame(
   expand.grid(poured=c("Yes", "No"),
   guess=c("Yes", "No")),
   count=c(3,1,1,3))

# nicer way of recreating long tables
data = untable(tea, freq="count")

with(data, crosstable(poured, guess)) # fisher=TRUE

}
\seealso{
\code{\link[stats]{xtabs}}, \code{\link{Frequency}},
\code{\link[base]{table}}, \code{\link[base]{prop.table}}
}
\keyword{Exploratory}

