% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{CrossTabs}
\alias{CrossTabs}
\alias{CrossTabs.default}
\alias{twoway}
\title{Stata-Like Two-Way Tabulation}
\usage{
CrossTabs(x, y, digits = 2, missing.include = FALSE, ...)

\method{CrossTabs}{default}(x, y, digits = 2, missing.include = FALSE, ...)
}
\arguments{
\item{x,y}{The variables for the cross tabulation.}

\item{digits}{The number of digits for rounding proportions.}

\item{missing.include}{If TRUE, then remove any unused factor levels.}

\item{\dots}{Additional arguements (currently ignored).}
}
\description{
The function produces a cross-tabulation with cell counts and column proportions (similar to STATA's \code{tabulate varname1 varname2, col}).
}
\examples{
# Agresti (2002), table 3.10, p. 106
# 1992 General Social Survey- Race and Party affiliation
gss <- data.frame(
   expand.grid(Race=c("black", "white"),
   party=c("dem", "indep", "rep")),
   count=c(103,341,15,105,11,405))

df <- gss[rep(1:nrow(gss), gss[["count"]]), ]

with(df, CrossTabs(Race, party))

# Tea-Tasting Experiment data
 tea <- data.frame(
   expand.grid(poured=c("Yes", "No"),
   guess=c("Yes", "No")),
   count=c(3,1,1,3))

data = untable(tea, freq="count")
with(data, CrossTabs(guess, poured))
# with(data, CrossTabs(guess, poured, row=TRUE, column=TRUE, fisher=TRUE))
}
\seealso{
\code{\link[stats]{xtabs}}, \code{\link{CrossTabs}},
\code{\link[base]{table}}, \code{\link[base]{prop.table}}
}

