% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crosstabs.R
\encoding{UTF-8}
\name{crosstabs}
\alias{crosstabs}
\alias{crosstabs.default}
\title{Cross-tabulation}
\usage{
crosstabs(.data, row, col, weights = NULL, observed = TRUE,
  expected = TRUE, total.pct = FALSE, row.pct = FALSE, col.pct = FALSE,
  sresid = FALSE, chisq = FALSE, phi = FALSE, cramerV = FALSE,
  contingency = FALSE)

\method{crosstabs}{default}(.data, row, col, weights = NULL,
  observed = TRUE, expected = TRUE, total.pct = FALSE, row.pct = FALSE,
  col.pct = FALSE, sresid = FALSE, chisq = FALSE, phi = FALSE,
  cramerV = FALSE, contingency = FALSE)
}
\arguments{
\item{.data}{The data.frame.}

\item{row}{The row variable.}

\item{col}{The column variable.}

\item{weights}{The frequency variable.}

\item{observed}{If FALSE, observed values are excluded from the table.}

\item{expected}{If TRUE, expected values are included.}

\item{total.pct}{If TRUE, total proportions are included.}

\item{row.pct}{If TRUE, row proportions are included.}

\item{col.pct}{If TRUE, column proportions are included.}

\item{sresid}{If TRUE, standardized residuals are included.}

\item{chisq}{If TRUE, the result of chi-square test of independence is included.}

\item{phi}{If TRUE, the Phi coefficient is included.}

\item{cramerV}{If TRUE, Cramer's V coefficient is included.}

\item{contingency}{If TRUE, the Contingency coefficient (C) is included.}
}
\description{
Makes a contingency table and computes chi-square test of independence, Phi coefficient, Cramer's V and Contingency coefficient.
}
\examples{
titanic \%>\% crosstabs("SEX", "AGE")


# Tea-Tasting Experiment data
 tea <- data.frame(
   expand.grid(poured=c("Yes", "No"),
   guess=c("Yes", "No")),
   count=c(3,1,1,3))

tea \%>\% crosstabs("guess", "poured", "count", row.pct=TRUE, col.pct=TRUE)

# Agresti (2002), table 3.10, p. 104
# 1991 General Social Survey: hypothesis of independence between
# party identification and race.
gss <- data.frame(
   expand.grid(race=c("black", "white"),
   party=c("dem", "indep", "rep")),
   count=c(103,341,15,105,11,405))

crosstabs(gss, "race", "party", "count", chisq=TRUE)
}

