% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Describe.R
\name{Describe}
\alias{Describe}
\title{Statistical description}
\usage{
Describe(x, ...)
}
\arguments{
\item{x}{A data frame, matrix, vector, or formula.}

\item{\dots}{Additional arguments passed to \code{Describe.default}.}
}
\description{
Provides description of a vector, matrix, data.frame.
}
\examples{
\dontrun{
 Describe(turnout)
 desc <- Describe(turnout)
 desc$v1   # print description for just v1
 desc[c('v2','v3')]    # print description for two variables.
 desc[sort(names(desc))] # print in alphabetic order by column names.

# Describing part of a data frame:
 with(turnout, Describe(v1 ~ v2*v3 + v4) )
 with(turnout, Describe(~ v2 + v3) )
 with(turnout, Describe(~ v2 + v3, weights=freqs)) # weighted analysis
}


}

