% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Destring.R
\encoding{UTF-8}
\name{Destring}
\alias{Destring}
\title{Factors to numeric}
\usage{
Destring(x)
}
\arguments{
\item{x}{A factor whose levels will be converted.}
}
\description{
Converts factors to numeric like in Stata.
}
\examples{
mylevels <- c('Strongly Disagree', 'Disagree', 'Neither', 'Agree', 'Strongly Agree')

myvar <- factor(sample(mylevels[1:5], 10, replace=TRUE))

unclass(myvar) # testing order

Destring(myvar)

}
\seealso{
\code{\link{Safechars}}.
}
\keyword{Manipulation}

