% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/destring.R
\encoding{UTF-8}
\name{destring}
\alias{destring}
\title{Factors to numeric}
\usage{
destring(x)
}
\arguments{
\item{x}{A factor whose levels will be converted.}
}
\description{
Converts factors to numeric like in Stata.
}
\examples{
mylevels <- c('Strongly Disagree', 'Disagree', 'Neither', 'Agree', 'Strongly Agree')

myvar <- factor(sample(mylevels[1:5], 10, replace=TRUE))

unclass(myvar) # testing order

destring(myvar)
}
\keyword{Misc}

