% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dHondt.R
\name{dHondt}
\alias{dHondt}
\title{Method D'Hondt}
\usage{
dHondt(parties, votes, seats)
}
\arguments{
\item{parties}{A vector containig the identification of parties or candidates accordingly to the election outcome.}

\item{votes}{A vector containing the total number of formal votes received by the parties/candidates.}

\item{seats}{An integer for the number of seats to be filled (the district magnitude).}
}
\description{
The function calculate the seats in parliament given the total number of seats and the votes for each party.
}
\examples{
votes <- sample(1:10000, 5)
dHondt(letters[1:5], votes, 10 )

votes <- c(42201,38635,247736,170627,48236,117151,61379,35889,92321)
dHont(c("A","B","C","D","F","G","H","I","J"), votes, 26)
}

