% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dHondt.R
\name{dHondt}
\alias{dHondt}
\title{Method D'Hondt}
\usage{
dHondt(parties, votes, seats)
}
\arguments{
\item{parties}{A vector containig the identification of parties or candidates accordingly to the election outcome.}

\item{votes}{A vector containing the total number of formal votes received by the parties/candidates.}

\item{seats}{An integer for the number of seats to be filled (the district magnitude).}
}
\description{
The function calculate the seats in parliament given the total number of seats and the votes for each party.
}
\examples{
votes <- sample(1:10000, 5)
dHondt(letters[1:5], votes, 5 )

#Example: 2014 Brazilian election for the lower house in the state of Ceara:
votes <- c(490205, 1151547, 2449440, 48274, 54403, 173151)

# Coalitions leading by the following parties:
parties <- c("DEM","PMDB","PRB","PSB", "PSTU","PTC")

dHondt(parties, votes, seats=19)
}

