% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dummy.R
\encoding{UTF-8}
\name{Dummy}
\alias{Dummy}
\title{Generate dummy variables}
\usage{
Dummy(x, data = NULL, drop = TRUE)
}
\arguments{
\item{x}{a column position to generate dummies}

\item{data}{the data object as a data.frame}

\item{drop}{A logical value. If \code{TRUE}, unused levels will be omitted}
}
\description{
Provides an alternative to generate dummy variables
}
\details{
A matrix object
}
\examples{
df <- data.frame(y = rnorm(10), x = runif(10,0,1), sex = sample(1:2, 10, rep=TRUE))

Dummy(df$sex) ;

cbind(df, Dummy(df$sex));
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Models}

