% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\encoding{UTF-8}
\name{dummy}
\alias{dummy}
\title{Generate dummy variables}
\usage{
dummy(x, data = NULL, drop = TRUE)
}
\arguments{
\item{x}{a column position to generate dummies}

\item{data}{the data object as a data.frame}

\item{drop}{A logical value. If \code{TRUE}, unused levels will be omitted}
}
\description{
Provides an alternative to generate dummy variables
}
\details{
A matrix object
}
\examples{
df <- data.frame(y = rnorm(25), x = runif(25,0,1), sex = sample(1:2, 25, rep=TRUE))

dummy(df$sex)

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Models}

