% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPlot.R
\name{exportData}
\alias{exportData}
\title{Export data in a tabular file}
\usage{
exportData(data, filename)
}
\arguments{
\item{data}{the data object.}

\item{filename}{the file name.}
}
\description{
Export data in a tabular file.
}

