% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factorial.R
\encoding{UTF-8}
\name{factorial}
\alias{factorial}
\title{Compute n!}
\usage{
factorial(n)
}
\arguments{
\item{n}{The number to be factored out.}
}
\description{
Compute n!
}
\examples{
factorial(5); factorial(5565)
}

