% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{factorial}
\alias{factorial}
\title{Conditional replacement}
\usage{
factorial(n)
}
\arguments{
\item{.data}{The data object.

Compute n!}

\item{x}{The number}
}
\description{
Conditional replacement
}

