% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualization.R
\name{fade}
\alias{fade}
\title{Add transparency}
\usage{
fade(color, alpha = 0.5)
}
\arguments{
\item{color}{Any color or vector of colors}

\item{alpha}{Level for alpha, default is \code{0.5}}
}
\description{
Alpha function to add transparency in graphic objects
}
\examples{
# setup data
x <- seq(0, 50, 1)
supply <- x * -2 + 100
demand <- x * 2
# Point size and transparency
plot(supply, demand, pch = 19, cex = 3, col = fade("red", 0.5))
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Graphs}

