% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{scale_fill_pub}
\alias{scale_fill_pub}
\title{Publication color scales.}
\usage{
scale_fill_pub(palette = "pub12", ...)
}
\arguments{
\item{palette}{the palette name, a character string.}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
See \code{\link{pub_color_pal}} for details.
}
\seealso{
Other color publication: \code{\link{scale_color_pub}}
}
\keyword{ggplot2}

