% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\name{fillForward}
\alias{fillForward}
\title{fill NA by previous cell value (fill forward)}
\usage{
fillForward(var)
}
\arguments{
\item{var}{the column that contains NAs}
}
\description{
fillForward will carry values forward from one observation to the next, filling in missing values with the previous value.
}
\note{
This is not intended for imputing missing values; it is regarded as a bad choice for missing-value imputation. The intent is, rather, to fill in \dQuote{holes}, where a value should naturally prevail from one observation to the next.
}
\examples{
use(ssex)
peek(ssex)

fillForward(ssex$Favor)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

