% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fillin}
\alias{fillin}
\title{Rectangularize a dataframe Fill in missing records}
\usage{
fillin(x, by, fill = NA)
}
\arguments{
\item{x}{a data frame.}

\item{by}{a vector of at least 2 variables from the data frame. If missing all variables in the data frame will be used.}

\item{fill}{the value used to fill in all other variables from the data frame, default is \code{fill = TRUE}.}
}
\value{
a data object of the same class as \code{x}.
}
\description{
This function produces a complete rectangularization by adding observations with missing data so that all combinations (interactions) of the specified variables exist.
}
\examples{
data <- data.frame(sex=c("female","male","male"),
race=c("black","black","white"), y=c(.5,.4,.1), x=c(32,40,53))

data

fillin(data, by=c(sex,race))
}

