% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\name{flag}
\alias{flag}
\title{Flag duplicated observations}
\usage{
flag(x = .data, check.by = NULL)
}
\arguments{
\item{check.by}{the formula for checking row-wise for duplicates.}

\item{data}{the data object.}
}
\description{
Marks how many times an observation appears in the dataset.
}
\examples{
df <- data.frame(matrix(c(51,42,43,1,22,51, 92,28,21,1,22,9),ncol=3,byrow=TRUE))
colnames(df) <- c("A","B","C")
flag(x = df, check.by = c("A", "B") )
}

