% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mi.R
\name{flag}
\alias{flag}
\title{Flag duplicated observations}
\usage{
flag(obj = .data, check.by = NULL)
}
\arguments{
\item{obj}{The data object.}

\item{check.by}{The formula for checking row-wise for duplicates.}
}
\description{
Marks how many times an observation appears in the dataset.
}
\examples{
df <- data.frame(matrix(c(51,42,43,1,22,51,
                 92,28,21,1,22,9),ncol=3, byrow = TRUE))
colnames(df) <- c("A","B","C")
flag(df, check.by = c("A", "B") )
}

