% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/followup.R
\name{followup}
\alias{followup}
\title{Mark Followup Interviews by ID and Time}
\usage{
followup(id, time)
}
\arguments{
\item{id}{the subject's identification.}

\item{time}{the followup contact date.}
}
\description{
Mark Followup Interviews by ID and Time
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

