% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{Footnote}
\alias{Footnote}
\title{Add Footnote to ggplot2 Objects}
\usage{
Footnote(note, x = 0.85, y = 0.014, hjust = 0.5, vjust = 0.5,
  newlines = TRUE, fontfamily = "serif", fontface = "plain",
  colour = "gray85", size = 9, angle = 0, lineheight = 0.9, alpha = 1)
}
\arguments{
\item{note}{string or plotmath expression to be drawn.}

\item{x}{the x location of the label.}

\item{y}{the y location of the label.}

\item{hjust}{horizontal justification}

\item{vjust}{vertical justification}

\item{newlines}{should a new line be appended to the start and end
of the string, for spacing?}

\item{fontfamily}{the font family}

\item{fontface}{the font face ("plain", "bold", etc.)}

\item{colour}{text color}

\item{size}{point size of text}

\item{angle}{angle at which text is drawn}

\item{lineheight}{line height of text}

\item{alpha}{the alpha value of the text}
}
\description{
Add footnote to ggplot2 objects.
}

