% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatPercent.R
\encoding{UTF-8}
\name{formatPercent}
\alias{formatPercent}
\title{Converts to percentiles}
\usage{
formatPercent(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Converts a numeric vector to percentiles.
}
\examples{
vec <- seq(1:5)
formatPercent(vec)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

