% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{formatPercent}
\alias{formatPercent}
\title{Format a numeric proportion}
\usage{
formatPercent(x, digits = 1)
}
\arguments{
\item{x}{a number or a vector whose numbers will be formated.}

\item{digits}{the number of digits to be left.}
}
\description{
Takes a number and formats it as a percentage.
}
\examples{
x <- c(.15, .00556, .55, 0.246)
formatPercent(x, 0)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

