% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{freq}
\alias{freq}
\alias{freq.default}
\alias{oneway}
\title{Frequency Table}
\usage{
freq(.data, x, verbose = TRUE, ...)

\method{freq}{default}(.data, x, verbose = TRUE, ...)
}
\arguments{
\item{.data}{The data.frame.}

\item{x}{A column for which a frequency of values is desired.}

\item{verbose}{A logical value, if \code{TRUE}, extra statistics are also provided.}

\item{\dots}{Additional arguements (currently ignored)}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
data(cathedrals)

freq(cathedrals, Type)

cathedrals \%>\% freq(Height)
}
\seealso{
\code{\link{freqw}}, \code{\link{CrossTabs}}.
}

