% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency Table}
\usage{
freq(.data, x, verbose = TRUE)
}
\arguments{
\item{.data}{The data.frame.}

\item{x}{A column from which the frequency of values is desired.}

\item{verbose}{A logical value, if \code{TRUE}, extra statistics are also provided.}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
data("cathedral")

freq(cathedral, Type)

cathedral \%>\% freq(Height)
}

