% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency table}
\usage{
freq(.data, ...)
}
\arguments{
\item{.data}{The data.frame}

\item{\dots}{Further arguments.}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
data(women)
with(women, freq(weight))

freq(women, weight)

women \%>\% freq(weight)
}

