% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency table}
\usage{
freq(.data, x)
}
\arguments{
\item{.data}{The data.frame}

\item{x}{A column from which the frequency of values is desired.}
}
\description{
Simulating the FREQ procedure of SPSS.
}
\examples{
data("cathedral")

freq(cathedral, Type)

cathedral \%>\% freq(Height)
}

