% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{Gallagher}
\alias{Gallagher}
\title{Gallagher Index}
\usage{
Gallagher(v, s, ...)

Gallagher(v, s, ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates the Gallagher index (LSq index).
}
\details{
The score is calculated as: sqrt(sum((v_{i}-s{i})^2)/2).
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Gallagher, M. (1991) Proportionality, disproportionality and electoral systems. Electoral Studies 10(1):33-51.

 @examples
# 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

Gallagher(pvotes, pseats)
}
\seealso{
\code{\link{CoxShugart}}, \code{\link{PoliticalDiversity}}, \code{\link{Grofman}}, \code{\link{Farina}},  \code{\link{Lijphart}}. For more details see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}
}

