% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gini.simpson.R
\encoding{UTF-8}
\name{gini}
\alias{gini}
\title{Weighted Gini Index}
\usage{
gini(x, weights = rep(1, length = length(x)), ...)

gini(x, weights = rep(1, length = length(x)), ...)
}
\arguments{
\item{x}{A data.frame, a matrix-like, or a vector.}

\item{weights}{A vector containing weights for \code{x}.}

\item{\dots}{Additional arguements (currently ignored)}
}
\description{
Computes the unweighted and weighted Gini index of a distribution.
}
\examples{
# generate a vector (of incomes)
x <- c(778, 815, 857, 888, 925, 930, 965, 990, 1012)
# compute Gini index
gini(x)

gini(c(100,0,0,0))
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
.
}
\seealso{
\code{\link{gini.simpson}}.
}
\keyword{Concentration}
\keyword{Diversity,}

