% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/electoralTool.R
\name{gini.simpson}
\alias{gini.simpson}
\title{Gini Simpson Index}
\usage{
gini.simpson(x, na.rm = FALSE)
}
\arguments{
\item{x}{A data.frame, a matrix-like, or a vector.}

\item{na.rm}{A logical value to deal with NAs.}
}
\description{
Gini Simpson Index
}
\examples{
x <- as.table(c(69,17,7,62))
rownames(x) <- c("AB","C","D","0")
gini.simpson(x)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{politicalDiversity}}.
}

