% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{griliches76}
\alias{griliches76}
\title{Griliches's (1976) Data}
\format{A \code{data.frame} object with \Sexpr[eval=TRUE,echo=FALSE,results=text]{ncol(SciencesPo::griliches76)} variables and \Sexpr[eval=TRUE,echo=FALSE,results=text]{nrow(SciencesPo::griliches76)} observations.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(griliches76)
}
\description{
Data set used by Griliches (1976) on wages of very young men. This dataset contains the following columns:

\itemize{
\item rns residency in South.
\item rns80 a numeric vector.
\item mrt marital status = 1 if married.
\item mrt80 a numeric vector.
\item smsa reside metro area = 1 if urban.
\item smsa80 a numeric vector.
\item med mother's education, years.
\item iq iq score.
\item kww score on knowledge in world of work test.
\item year Year.
\item age a numeric vector.
\item age80 a numeric vector.
\item s completed years of schooling.
\item s80 a numeric vector.
\item expr experience, years.
\item expr80 a numeric vector.
\item tenure tenure, years.
\item tenure80 a numeric vector.
\item lw log wage.
\item lw80 a numeric vector.
}
}
\references{
Griliches, Z. (1976) Wages of very young men. \emph{The Journal of Political Economy,} \bold{84(4),} 69--85.
}
\keyword{datasets}

