% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{hash}
\alias{hash}
\title{Create useless hashes}
\usage{
hash(cols = 50, rows = 2)
}
\arguments{
\item{cols}{How many columns of hashes}

\item{rows}{How many rows of hashes}
}
\description{
A function to create useless hashes (####) for ease of commenting.
}
\examples{
hash(cols=50, rows=1)
}
\author{
Daniel Marcelino
}
\keyword{internal}

