% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{head}
\alias{head}
\alias{head.SQLiteConnection}
\title{Return the First Part of an Object}
\usage{
head(x, ...)

\method{head}{SQLiteConnection}(x, table = NULL, n = 10L, temp = FALSE,
  ...)
}
\arguments{
\item{x}{A \code{SQLiteConnection} object}

\item{...}{Additional arguments}

\item{table}{character specifying a table}

\item{n}{integer: Number of rows to output}

\item{temp}{logical should the function list the temp tables}
}
\description{
If just a database connection is selected, returns a dataframe of table names
If a table name is also supplied, the first n rows from this table are output
}

