% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inline.R
\name{inline}
\alias{inline}
\title{Inline data.frame}
\usage{
inline(str, header = TRUE, colClasses = NA, ...)
}
\arguments{
\item{str}{text representation of the data frame}

\item{header}{see \code{\link{read.table}}}

\item{colClasses}{see \code{\link{read.table}}}

\item{\dots}{see \code{\link{read.table}}}
}
\description{
Utility function to create of a \code{data.frame} inline.
}
\examples{
mydf <- inline(str = 'A, B, C, D
                      1, 2, 3, 4
                     .5, .2, .4, .5', sep=",")
}

