% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inline.R
\name{inline}
\alias{inline}
\title{Inline data.frame}
\usage{
inline(str, header = TRUE, colClasses = NA, ...)
}
\arguments{
\item{str}{text representation of the data frame}

\item{header}{see \code{\link{read.table}}}

\item{colClasses}{see \code{\link{read.table}}}

\item{\dots}{see \code{\link{read.table}}}
}
\description{
Utility function to create of a \code{data.frame} inline.
}
\examples{
mydf <- inline(str = "County, Gore, Bush, Buchanan, Nader
ALACHUA, 47365, 34124, 263, 3226
BAKER,    2392,  5610, 73, 53
BAY,     18850, 38637, 248, 828
BRADFORD, 3075, 5414, 65, 84
BREVARD, 97318, 115185, 570, 4470
BROWARD, 386561, 177323, 788, 7101", sep=",")

\dontrun{sapply(mydf[, 2 : 5], sum);}

\dontrun{cbind(stack(mydf), county = mydf$County);}
}

