% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert.row.R
\name{insert.row}
\alias{insert.row}
\title{Add New Row to a data.frame}
\usage{
insert.row(.data, row, where = 1)
}
\arguments{
\item{.data}{The existing data.frame}

\item{row}{The new row to be appended.}

\item{where}{An integer for the position to add the row, default is at the top of the data.frame.}
}
\description{
Facilitates inserting new rows in a data.frame.
}
\examples{
existingDF <- as.data.frame(matrix(seq(20),nrow=5,ncol=4))
existingDF
r <- 3
newrow <- seq(4)

insert.row(existingDF, newrow, r)
}

