% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GENERICS.R
\name{isid}
\alias{isid}
\title{Is an unique identifier}
\usage{
isid(object, ...)
}
\arguments{
\item{object}{a data object.}

\item{\dots}{Ignored parameters.}
}
\description{
Generic function for checking for a unique row identifier.
}

