% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\encoding{UTF-8}
\name{isid}
\alias{isid}
\title{Finds ID combination}
\usage{
isid(columns, data, verbose = TRUE)
}
\arguments{
\item{columns}{columns to combine}

\item{data}{the data object}

\item{verbose}{if \code{TRUE} messages maybe be displayed.}
}
\description{
Finds unique id combination
}
\details{
The original code published at \url{https://stackoverflow.com/} was berely touched.
}

