% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jackknife.R
\encoding{UTF-8}
\name{jackknife}
\alias{jackknife}
\title{Resamples Data Using the Jackknife Method}
\usage{
jackknife(x, p)
}
\arguments{
\item{x}{A vector}

\item{p}{An estimation of parameter}
}
\value{
est orignial estimation of parameter

jkest jackknife estimation of parameter

jkvar jackknife estimation of variance

jkbias jackknife estimate of biasness of parameter

jkbiascorr bias corrected parameter estimate
}
\description{
This function is used for estimating standard errors when the distribution is not know.
}
\examples{
x = runif(10, 0, 1)
mean(x)
jackknife(x,mean)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

