% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jackknife.R
\encoding{UTF-8}
\name{Jackknife}
\alias{Jackknife}
\title{Resamples Data Using the Jackknife Method}
\usage{
Jackknife(x, p)
}
\arguments{
\item{x}{A vector}

\item{p}{An function name for estimation of parameter as a string}
}
\value{
est orignial estimation of parameter

jkest jackknife estimation of parameter

jkvar jackknife estimation of variance

jkbias jackknife estimate of biasness of parameter

jkbiascorr bias corrected parameter estimate
}
\description{
This function is used for estimating standard errors when the distribution is not know.
}
\examples{
x = runif(10, 0, 1)
mean(x)
Jackknife(x,'mean')

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

