% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TESTS.R
\encoding{UTF-8}
\name{james.stein}
\alias{james.stein}
\title{James-Stein shrunken estimates}
\usage{
james.stein(y, k)
}
\arguments{
\item{y}{The response variable.}

\item{k}{The grouping factor.}
}
\description{
Computes James-Stein shrunken estimates of cell means given
a response variable (which may be binary) and a grouping indicator.
}
\references{
Efron, Bradley and Morris, Carl (1977) ``Stein's Paradox in Statistics.'' \emph{Scientific American} Vol. 236 (5): 119-127.

James, Willard and Stein, Charles (1961) ``Estimation with
Quadratic Loss.'' \emph{Proceedings of the Fourth Berkeley
Symposium on Mathematical Statistics and Probability}, Vol. 1: 361-379.
}
\keyword{Tests}

