% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kalmanFilter.R
\encoding{UTF-8}
\name{kalmanUpdate}
\alias{kalmanUpdate}
\title{Kalman Filter Update}
\usage{
kalmanUpdate(mean1, var1, mean2, var2)
}
\arguments{
\item{mean1}{Prior mean}

\item{var1}{Prior variance}

\item{mean2}{New mean}

\item{var2}{New variance}
}
\description{
Replicate the Kalman Filtering Process
}
\examples{
# Equal variances but different means
kalmanUpdate(10, 4, 12, 4)

# Different variances and means
kalmanUpdate(10, 8, 13, 2)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Kalman, R.E. (1960) A new approach to linear filtering and prediction problems. \emph{Journal of Basic Engineering} \bold{82(1):} 35-45.
}
\keyword{Models}

