% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\title{Return or Assign a Variable Label}
\usage{
label(x, value = NULL, data = .data)
}
\arguments{
\item{x}{The variable for which to obtain the corresponding variable label.}

\item{value}{If assigned, then the specified data frame is updated with this asssigned label.}

\item{data}{Data frame that contains the variable of interest. The output of the function is assigned to this data frame.}
}
\description{
Display a variable label for output, either text output at the console or graphics, such as a title on a graph.
}
\examples{
.data <- data.frame(AgeGroup=factor(
 1:9,
 labels= c("Younger than 16", "16-24",
           "25-34", "35-44", "45-54",
           "55-64", "65-74",
           "Older than 74", "(blank)"),
 ordered=TRUE),
 No=c(34, 2079, 2585, 1593,
      1274, 802, 291, 78, 13),
 Yes=c(18, 1970, 4035, 2328,
       1707, 924, 386, 68, 4))

dat <- label(AgeGroup, "Age group")
}

