% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm2eqn.R
\name{lm2eqn}
\alias{lm2eqn}
\title{Linear model to equation style}
\usage{
lm2eqn(.data, x, y, spaced = TRUE)
}
\arguments{
\item{.data}{The data.frame object.}

\item{x}{The independent variable(s).}

\item{spaced}{A logical value indicating if spaces should be added; default is TRUE.}

\item{x}{The dependent variable.}
}
\description{
Produces a text equation style to be added in plots.
}
\examples{
fit=lm(mpg~wt,data=mtcars)
summary(fit)
lm2equation("mtcars","wt","mpg")
require(ggplot2)

mtcars$name=rownames(mtcars)
mtcars$selected=ifelse(mtcars$name \%in\% c("Toyota Corolla","Merc 240D"),1,0)
selected=mtcars[mtcars$selected==1,]
mylabel=lm2equation("mtcars","wt","mpg")

ggplot(mtcars,aes(x=wt,y=mpg,colour=selected))+
 geom_text(data=selected,aes(label=name),hjust=-0.1)+
 geom_smooth(method=lm)+
 geom_point()+
 annotate(geom='text',x=4,y=30,size=7,label=mylabel,family='Times',fontface='italic')+
 scale_colour_gradientn(colours=c("black","red"))+
 theme_pub()
}

