% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{loglik}
\alias{loglik}
\title{Calculate the Log Likelihood of a Normal Distribution}
\usage{
loglik(x = data, mu, var)
}
\arguments{
\item{x}{data.}

\item{mu}{estimated mean.}

\item{var}{estimated variance.}
}
\value{
ll logliklihood of the distribution
}
\description{
Find the log likelihood of a normal distribution.
}
\examples{
x = rnorm(100, 3, 7)
loglik(x,3,7)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

