% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manipulation.R
\encoding{UTF-8}
\name{lookup}
\alias{lookup}
\title{Lookup}
\usage{
lookup(x, lookup.array)
}
\arguments{
\item{x}{the variable}

\item{lookup.array}{a n-by-2 array used for looking up.}
}
\description{
Recodes values of a vector from a lookup array.
}

