% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{ltaylor96}
\alias{ltaylor96}
\title{Lothian's and Taylor's (1996) Data Set}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::ltaylor96)} variables and \Sexpr{nrow(SciencesPo::ltaylor96)} observations.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(ltaylor96)
}
\description{
Data used by Lothian and Taylor (1996) on the real exchange rate behaviour. This dataset contains the following columns:

\itemize{
\item year Year
\item spot dollar/sterling exchange rate.
\item USwpi U.S. wholesale price index, 1914==100.
\item UKwpi U.K. wholesale price index, 1914==100.
}
}
\references{
Lothian, J. R., and Taylor, M. P. (1996) Real exchange rate behavior: the recent float from the perspective of the past two centuries. \emph{Journal of Political Economy,} 488--509.
}
\keyword{datasets}

