% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanFromRange.R
\encoding{UTF-8}
\name{MeanFromRange}
\alias{MeanFromRange}
\title{Estimates Mean and Standard Deviation from Median and Range}
\usage{
MeanFromRange(low, med, high, n)
}
\arguments{
\item{low}{The min of the data.}

\item{med}{The median of the data.}

\item{high}{The max of the data}

\item{n}{The size of the sample.}
}
\description{
When conductig a meta-analysis study, it is not always possible to
 recover from reports, the mean and standard deviation values, but rather the
 median and range of values. This function provides a method to compute the
 median/range values into mean and variance estimates.
}
\examples{
MeanFromRange(low=5, med=8, high=12, n=10)

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Hozo1, Stela P.; et al (2005) Estimating the mean and variance from the median, range, and the size of a sample. \emph{BMC Medical Research Methodology}, 5:13.
}

