% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanFromRange.R
\name{meanFromRange}
\alias{meanFromRange}
\title{Estimates Mean and Standard Deviation from Median and Range}
\usage{
meanFromRange(low, med, high, n)
}
\arguments{
\item{low}{The min of the data.}

\item{med}{The median of the data.}

\item{high}{The max of the data}

\item{n}{The size of the sample.}
}
\description{
When conductig a meta-analysis study, it is not always possible to recover from reports the mean and standard deviation values, but rather the medians and range of values. This function provides an approach to convert the median/range into a mean and a variance.
}
\examples{
meanFromRange(5,8,12,10)
}
\references{
Hozo1, Stela P.; et al (2005) Estimating the mean and variance from the median, range, and the size of a sample. \emph{BMC Medical Research Methodology}, 5:13.
}

