% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mi.R
\name{mi}
\alias{mi}
\title{Identify Columns with at least one NA value}
\usage{
mi(x)
}
\arguments{
\item{x}{a \code{data.frame}}
}
\value{
A message indicating whether any column in \code{x} has missing data.
}
\description{
Identify Columns with at least one NA value
}
\examples{
data(ssex)
mi(ssex)
}

