\name{mishkin92}
\alias{mishkin92}
\docType{data}
\title{Mishkin's (1992) Data
}
\description{Data from the Frederic S. Mishkin (1992) paper \dQuote{Is the Fisher Effect for real?}.
}
\usage{data(mishkin92)}
\format{
  A data frame with 491 observations on the following 8 variables.
  \describe{
    \item{\code{year}}{Year}
    \item{\code{mon}}{a numeric vector}
    \item{\code{inf1mo}}{a numeric vector}
    \item{\code{inf3mo}}{a numeric vector}
    \item{\code{tbill1mo}}{a numeric vector}
    \item{\code{tbill3mo}}{a numeric vector}
    \item{\code{cpiu}}{a numeric vector}
    \item{\code{quote}}{a numeric vector}
  }
}

\source{
Hayashi, F. (2000) \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}

}
\references{Mishkin, F. S. (1992) Is the Fisher effect for real?: A reexamination of the relationship between inflation and interest rates. \emph{Journal of Monetary Economics,} \bold{30(2),} 195--215.

}
\examples{
data(mishkin92)
}
\keyword{datasets}
