% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\encoding{UTF-8}
\name{Mode}
\alias{Mode}
\alias{Mode.data.frame}
\title{Calculate the Mode}
\usage{
Mode(x, na.rm = FALSE)

Mode(x, na.rm = FALSE)

Mode.data.frame(x, na.rm = TRUE)
}
\arguments{
\item{x}{An \R object.}

\item{na.rm}{A logical value indicating whether \code{NA}
should be stripped before the computation proceeds. Default
is \code{FALSE}}
}
\description{
Estimates the mode for a vector
}
\note{
This function replaces the \code{base} function of the same name, while \code{SciencesPo::mode} calculates the \dQuote{mode}, \code{base::mode} prints the \dQuote{class} of an object.
}
\examples{
myvar <-c(1,1,2,2,3,3,4,4,5, NA)
Mode(myvar)

Mode(myvar, FALSE)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

