% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/STATISTICS.R
\encoding{UTF-8}
\name{mode}
\alias{mode}
\title{Calculate the Mode}
\usage{
mode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A data vector}

\item{na.rm}{A logical value, default is \code{FALSE}}
}
\description{
Estimates the mode for a vector
}
\note{
this function replaces the \code{base} function with the same name, but is does different thing. While \code{SciencesPo::mode} intuitively calculates the \dQuote{mode}, \code{base::mode} prints the \dQuote{class} of an object.
}
\examples{
myvar <-c(1,1,2,2,3,3,4,4,5, NA)
mode(myvar)

mode(myvar, FALSE)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

