% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{name.split}
\alias{name.split}
\title{Splits name field variable}
\usage{
name.split(name, data = .data)
}
\arguments{
\item{name}{the name field column.}

\item{data}{the data.frame name.}
}
\value{
two columns or a list.
}
\description{
Splits a name field variable allocating the first and last names into two new columns or a list.
}
\examples{
df <- data.frame( name = c("Martin Luther King", "Nelson Mandela", "Simon Bolivar") )
name.split(df$name)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\link{unnest}.
}

