% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convenience.R
\encoding{UTF-8}
\name{nearest.loc}
\alias{nearest.loc}
\title{Find Location of the Nearest Value}
\usage{
nearest.loc(x, value)
}
\arguments{
\item{x}{A vector.}

\item{value}{The value that you want to find.}
}
\description{
Find the location of the nearest value to a number that you specify.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

