% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CONVINIENCE.R
\encoding{UTF-8}
\name{nearest}
\alias{nearest}
\title{Find the Nearest Value}
\usage{
nearest(x, value)
}
\arguments{
\item{x}{A vector.}

\item{value}{The value that you want to find.}
}
\description{
Find the the nearest value to a number that you specify.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

