\name{nerlove63}
\alias{nerlove63}
\docType{data}
\title{Marc Nerlove's (1963) data
}
\description{Data used by Marc Nerlove (1963) on returns of electricity supply.
}
\usage{data(nerlove63)}
\format{
  A data frame with 145 observations on the following 5 variables.
  \describe{
    \item{\code{totcost}}{costs in 1970, MM USD}
    \item{\code{output}}{output, billion KwH}
    \item{\code{plabor}}{ price of labor}
    \item{\code{pfuel}}{price of fuel}
    \item{\code{pkap}}{price of capital}
  }
}

\source{ayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}

}
\references{Nerlove, M. (1963) Returns to Scale in Electricity Supply. In \emph{Measurement in Economics-Studies in Mathematical Economics and Econometrics in Memory of Yehuda Grunfeld,} edited by Carl F. Christ. Stanford: Stanford.
}
\examples{
data(nerlove63)
}
\keyword{datasets}
