% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SciencesPo.R
\docType{data}
\name{nerlove63}
\alias{nerlove63}
\title{Marc Nerlove's (1963) data}
\format{A data frame with 145 observations on the following 5 variables.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(nerlove63)
}
\description{
Data used by Marc Nerlove (1963) on returns of electricity supply.
\itemize{
\item totcost costs in 1970, MM USD.
 \item output output, billion KwH.
\item plabor price of labor.
 \item pfuel price of fuel.
 \item pkap price of capital.
 }


#'
}
\references{
Nerlove, M. (1963) Returns to Scale in Electricity Supply. In \emph{Measurement in Economics-Studies in Mathematical Economics and Econometrics in Memory of Yehuda Grunfeld,} edited by Carl F. Christ. Stanford: Stanford.
}
\keyword{datasets}

