% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\encoding{UTF-8}
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in (Find Matching or Non-Matching Elements)}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{A vector of numeric, character, or factor values.}

\item{table}{A vector (numeric, character, factor), matching the mode of x}
}
\description{
"%nin%" is a binary operator, which returns a
 logical vector indicating if there is a match or not for its left
 operand. A true vector element indicates no match in left operand,
 false indicates a match.
}
\examples{
c('a','b','c') \%nin\% c('a','b')
}

