% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DISTRIBUTIONS.R
\encoding{UTF-8}
\name{normalcdf}
\alias{normalcdf}
\title{Normal Cumulative Distribution}
\usage{
normalcdf(lower, upper, mu = 0, sigma = 1)
}
\arguments{
\item{lower}{is the inferior extreme value.}

\item{upper}{is the superior extreme value.}

\item{mu}{is the mean \eqn{\mu}, its default value is \eqn{\mu = 0}}

\item{sigma}{is the standard deviation \eqn{\sigma}, its default value is \eqn{\sigma = 1}}
}
\description{
Calculates the normal distribution probability using \emph{lower bound} e \emph{upper bound} by the mean \eqn{\mu} and standard deviation.
}
\examples{
normalcdf(lower=-1.96,upper=1.96,mu=0,sigma=1)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\code{\link{draw.norm}}, \code{\link{normalpdf}}, \code{\link{invnormal}}.
}
\keyword{Distribution}

