% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{normalize}
\alias{normalize}
\title{Unity-based normalization}
\usage{
normalize(var, range)
}
\arguments{
\item{var}{is a vector to be normalized}

\item{range}{is a valid range, default is \code{c(0,1)}.}
}
\value{
Normalized values in an object of the same class as \code{var}.
}
\description{
Normalizes as feature scaling, \code{min - max}, or unity-based normalization. Typically used to bring all values into the range [0,1]. However, this can be generalized to restrict the range of values in the dataset between any arbitrary points  \code{a}  and  \code{b}, using: \deqn{X' = a + \frac{(x - x_{min})(b - a)}{(x_{max} - x_{min})} }.
}
\examples{
x <- sample(10)
normalize(x, range=c(0,1))
normalize(x)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\code{\link{scale}}, \code{\link{unscale}}
}
\keyword{Rescaling}

