% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalize.R
\encoding{UTF-8}
\name{normalize}
\alias{normalize}
\alias{normalize.character}
\alias{normalize.default}
\alias{normalize.factor}
\alias{normalize.numeric}
\title{Unity-based normalization}
\usage{
normalize(x, range, domain, ...)

\method{normalize}{factor}(x, range, domain = range(1:nlevels(x)), ...)

\method{normalize}{numeric}(x, center = TRUE, scale = TRUE, range = c(0,
  1), domain = range(x, na.rm = TRUE), ...)

\method{normalize}{default}(x, range = c(0, 1), domain, ...)

\method{normalize}{character}(x, range = c(0, 1), domain, ...)
}
\arguments{
\item{x}{is a vector to be normalized.}

\item{range}{is a numeric vector of length 2 as \code{0:1} for min and max values, default is \code{c(0,1)}.}

\item{domain}{a numeric vector of length 2.}

\item{center}{Enables values be centered, default is \code{TRUE}.}

\item{scale}{Enables values be scaled, default is \code{TRUE}.}

\item{\dots}{further arguments passed to or used by other methods.}
}
\value{
Normalized values in an object of the same class as \code{x}.
}
\description{
The function normalizes as feature scaling \code{min - max}, or unity-based normalization. Typically used to bring all values into the range [0,1], this may also be generalized to restrict the range of values to any arbitrary points \code{a}  and  \code{b}, using: \deqn{X' = a + \frac{(x - x_{min})(b - a)}{(x_{max} - x_{min})}}.
}
\note{
When center and scale are both set \code{TRUE}, the resulting values are.
}
\examples{
x <- sample(10)
normalize(x, range=(0:1))
normalize(x)

char = LETTERS
normalize(char, range=(0:1), domain=range(w, na.rm=TRUE))
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\code{\link{svTransform}}.

\code{\link{scale}}.
}
\keyword{Normalization}
\keyword{Rescaling}

