% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{odds.ratio}
\alias{odds.ratio}
\title{Odds Ratio for 2x2 table}
\usage{
odds.ratio(tab, addtocounts = 0)
}
\arguments{
\item{tab}{a 2x2 table object}

\item{addtocounts}{}
}
\description{
Odds Ratio for 2x2 table
}
\examples{
data <- data.frame(cbind(gender=c("men", "men", "women", "women"), race=c("black", "white", "white", "black"), perc=c(90.6, 90.6, 90.6,78.8)))
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

