% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\name{odds.ratio}
\alias{odds.ratio}
\title{Odds Ratio for 2 x 2 Contingency Tables}
\usage{
odds.ratio(x, pad.zeros = FALSE, conf.level = 0.95, ...)

odds.ratio(x, pad.zeros = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A 2 X 2 matrix, data frame or table of counts.}

\item{pad.zeros}{A logical. If \code{pad.zeros=TRUE}, zeros will be added to the matrix.}

\item{conf.level}{The confidence level of the interval.}

\item{\dots}{Additional arguements (currently ignored).}
}
\description{
This function calculates the odds ratio for a 2 x 2 contingency table and a confidence interval (default conf.level is 95 percent) for the each estimate. x should be a matrix, data frame or table.
}
\examples{
tea <- matrix(c(3, 1, 1, 3), nrow = 2)
rownames(tea) <- c("milk", "tea")
colnames(tea) <- c("milk", "tea")
tea
odds.ratio(tea)

mat <- matrix(c(18515, 18496, 1427, 1438), nrow = 2)
rownames(mat) <- c("Placebo", "Aspirin")
colnames(mat) <- c("No", "Yes")
mat
odds.ratio(mat)

data(titanic)
tab = table(titanic$CLASS, titanic$SURVIVED)
tab
odds.ratio(tab)

}

