% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{otherNames}
\alias{otherNames}
\title{Extracts names from a dataset other than the ones indicates}
\usage{
otherNames(data, check)
}
\arguments{
\item{data}{the input \code{data.frame}.}

\item{check}{The \code{names} you want to check.}
}
\value{
A character vector of the remaining names.
}
\description{
Extracts names from a dataset other than the ones indicates
}
\examples{
bigdf <- data.frame(a = 1:2, b = 3:4, c = 5:6)
otherNames(bigdf, "b")
}
\seealso{
\code{\link{setdiff}}
}

