% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outliers.R
\name{Outliers}
\alias{Outliers}
\alias{Outliers.default}
\title{Detect Outliers}
\usage{
Outliers(x, index = NULL, ...)

\method{Outliers}{default}(x, index = NULL, ...)
}
\arguments{
\item{x}{A numeric object}

\item{index}{A numeric value to be considered in the computations}

\item{...}{Parameters which are typically ignored}
}
\value{
Returns the minimum and maximum values, respectively preceded by their positions in the \code{vector}, \code{matrix} or \code{data.frame}.
}
\description{
Perform an exploaratory test to detect \emph{outliers}. The quantity for \emph{min} reveals the minimum deviation from the mean, the integer in \emph{closest}, highlights the position of that element. The quantity for \emph{max} is the maximum deviation from the mean, and the \code{farthest} integer is the position of that value.
}
\examples{
Outliers(x <- rnorm(20))

#data frame:
age <- sample(1:100, 1000, rep=TRUE);
Outliers(age)

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Dixon, W.J. (1950) Analysis of extreme values. \emph{Ann. Math. Stat.} \bold{21(4),} 488--506.
}
\seealso{
\link{WinsorizedMean} for diminishing the impact of outliers.
}
\keyword{Exploratory}

