% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{paired}
\alias{paired}
\title{Paired Data}
\format{A data frame with 9 observations on the following 3 variables.}
\usage{
data(paired)
}
\description{
Artificial experimental paired data

\itemize{
\item patient the patient id.
\item before_X before treatment.
\item after_Y after treatment.
 }
}
\keyword{datasets}

