% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{paired}
\alias{paired}
\title{Paired Data}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::paired)} variables and \Sexpr{nrow(SciencesPo::paired)} observations.}
\usage{
data(paired)
}
\description{
Artificial data for a paired experiment. This dataset contains the following columns:

\itemize{
\item patient the patient id.
\item before_X before treatment.
\item after_Y after treatment.
 }
}
\keyword{datasets}

