% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{parties_color_pal}
\alias{parties_color_pal}
\title{Color Palettes for Political Organizations (discrete)}
\usage{
parties_color_pal(palette = "BRA")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes for political organizations.
}
\examples{
library(scales)
show_col(parties_color_pal("BRA")(30))

# Argentine
show_col(parties_color_pal("ARG")(10))

# US
show_col(parties_color_pal("USA")(10))

# Canada
show_col(parties_color_pal("CAN")(10))
}
\seealso{
Other colour parties: \code{\link{scale_colour_parties}}
}

