% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{party_color_pal}
\alias{party_color_pal}
\title{Political Parties Color Palette (Discrete) and Scales}
\usage{
party_color_pal(palette = "BRA")
}
\arguments{
\item{palette}{the palette name, a character string.}
}
\description{
An N-color discrete palette for political parties.
}
\examples{
library(scales)
library(ggplot2)

show_col(party_color_pal("BRA")(20))

# Argentine
show_col(party_color_pal("ARG")(10))

# US
show_col(party_color_pal("USA")(4))

# Canada
show_col(party_color_pal("CAN")(10))

}
\seealso{
Other color party: \code{\link{scale_color_party}}
}

