% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palletes.R
\name{party_color_pal}
\alias{party_color_pal}
\title{Color Palettes for Political Organizations (discrete)}
\usage{
party_color_pal(palette = "BRA", label = FALSE)
}
\arguments{
\item{palette}{the palette name.}

\item{label}{if \code{TRUE}, the party label associated with the color is returned instead.}
}
\description{
Color palettes for political organizations.
}
\examples{
library(scales)
show_col(party_color_pal("BRA")(20))

# Argentine
show_col(party_color_pal("ARG")(10))

# US
show_col(party_color_pal("USA")(4))

# Canada
show_col(party_color_pal("CAN")(10))

}
\seealso{
Other color party: \code{\link{scale_color_party}}
}

