% Generated by roxygen2 (4.0.2): do not edit by hand
\name{peek}
\alias{peek}
\title{Show Randomly Drawn Observations}
\usage{
peek(x, n = 10)
}
\arguments{
\item{x}{A matrix or data.frame object}

\item{n}{The number of rows to be shown}
}
\description{
Provide a sly view of the data by randomly draw observations, instead of showing only the first \code{head()} or the last \code{tail()} rows of an object.
}
\examples{
data(titanic)
peek(titanic)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\keyword{Tables}

