% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DEPRECATED.R
\name{peek}
\alias{peek}
\title{Show Observations Randomly Drawn from the Data}
\usage{
peek(.data = FALSE, ...)
}
\arguments{
\item{.data}{The data object.}

\item{...}{further parameters.}
}
\description{
Function is deprecated, please use instead.
}

