% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CONVENIENCE.R
\encoding{UTF-8}
\name{percentile}
\alias{percentile}
\title{Converts into percentile}
\usage{
percentile(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Converts numeric vector into its percentile.
}
\examples{
vec <- seq(1:5)
percentile(vec)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

