% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{percentile}
\alias{percentile}
\title{Converts into percentiles}
\usage{
percentile(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
Converts numeric vector into its percentile.
}
\examples{
vec <- seq(1:5)
percentile(vec)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

