% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutate.R
\encoding{UTF-8}
\name{permutate}
\alias{permutate}
\title{Create k random permutations of a vector}
\usage{
permutate(input, k)
}
\arguments{
\item{input}{vector to be permutated.}

\item{k}{number of permutations.}
}
\description{
Creates a k random permutation of a vector.
}
\details{
should be used only for length(input)! >> k
}
\examples{
permutate(input=1:5, k=5)
}
\keyword{Sampling}

