% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutate.R
\encoding{UTF-8}
\name{permutate}
\alias{permutate}
\title{Create k random permutations of a vector}
\usage{
permutate(input, k)
}
\arguments{
\item{input}{A vector to be permutated.}

\item{k}{number of permutations to be conducted.}
}
\description{
Creates a k random permutation of a vector.
}
\details{
should be used only for length(input)! >> k
}
\examples{
# row wise permutations
permutate(input=1:5, k=5)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\keyword{Manipulation}
\keyword{Sampling,}

