% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TESTS.R
\name{phi}
\alias{phi}
\title{Cramer's Phi Coefficient for Tables}
\usage{
phi(x, y = NULL)
}
\arguments{
\item{x}{A vector or a matrix.}

\item{y}{A vector that is ignored if x is a matrix and required if x is a vector.}
}
\description{
Computes the Cramer's Phi coefficient for tables.
}
\details{
Phi is seldom applied for indexing a \code{2 x 2} table, because the
researcher will typically want to contrast the two proportions as
an increment or ratio, not with a correlation coefficient.
Alternatives to Phi are the Pearson's C; Tschuprow's T, and Cramer's V.
}
\examples{
if (interactive()) {
 # some data:
male <- c(33, 76, 6)
female <- c(47, 153, 25)
mat <- cbind( male, female )
colnames(mat) <- c( 'good', 'satisfactory', 'bad')

phi(mat)
}
}
\references{
Agresti, Alan (1996) \emph{Introduction to categorical data
 analysis}. NY: John Wiley and Sons
}

