% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{Presidents}
\alias{Presidents}
\title{The Measure of US Presidents}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::Presidents)} variables and \Sexpr{nrow(SciencesPo::Presidents)} observations.}
\source{
US Presidents: \url{http://www.jimwegryn.com/Names/Presidents.php}
Inside Gov. \url{http://www.us-presidents.insidegov.com}.
Wikipedia: \url{http://en.wikipedia.org/wiki/United_States_presidential_election,_2012}.
Wikipedia:\url{http://en.wikipedia.org/wiki/Heights_of_presidents_and_presidential_candidates_of_the_United_States}.
}
\usage{
data(Presidents)
}
\description{
The US presidents and their main opponents' heights (cm). This dataset contains the following columns:

\itemize{
\item election The election year.
\item winner The winner candidate.
\item winner.height The winner candidate's height in centimeters (cm).
\item winner.vote The popular vote support for the winner.
\item winner.party The winner's party.
\item opponent The main opponent candidate.
\item opponent.height The opponent candidate's height in centimeters (cm).
\item opponent.vote Popular vote support for the main opponent candidate.
\item opponent.party The opponent's party.
\item turnout The electorate turnout in percentages.
\item winner.bmi The winner Body Mass Index (BMI) estimate \code{(BMI = weight in kg/(height in meter)**2)}
 }
}
\keyword{datasets}

