% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELECTORALTOOL.R
\docType{methods}
\name{proportionality}
\alias{disproportionality}
\alias{proportionality}
\title{Indexes of Proportionality}
\usage{
proportionality(v, s, index = "gallagher", ...)

proportionality(v, s, index = "gallagher", ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{index}{The desired type of index, see details below.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates several indexes of proportionality to show the relationship of votes to seats.
}
\details{
The following measures are available:
\itemize{
\item {"cox.shugart"}{Cox-Shugart Measure of Proportionality}
\item {"inv.cox.shugart"}{The Inverse of the Cox-Shugart Index}
\item {"farina"}{Farina Index of Proportionality}
\item {"gallagher"}{Gallagher Index of Disproportionality}
\item {"inv.gallagher"}{The Inverse of the Gallagher Index}
\item {"grofman"}{Grofman Index of Proportionality}
\item {"lijphart"}{Lijphart Index of Proportionality}
\item {"loosemore.hanby"}{Loosemore-Hanby Index of Disproportionality}
\item {"rae"}{Rae Index of Disproportionality}
\item {"inv.rae"}{The Inverse of the Rae Index}
\item {"rose"}{Rose Index of Disproportionality}
\item {"inv.rose"}{The Inverse of the Rose Index}
}
}
\examples{
#' # 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

proportionality(pvotes, pseats) # default is gallagher

# 2012 Quebec provincial election:
pvotes = c(PQ=31.95, Lib=31.20, CAQ=27.05, QS=6.03, Option=1.89, Other=1.88)
pseats = c(PQ=54, Lib=50, CAQ=19, QS=2, Option=0, Other=0)

proportionality(pvotes, pseats, index="rae")

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Duncan, O. and Duncan, B. (1955) A methodological analysis of segregation indexes. \emph{American Sociological Review} 20:210-7.

Gallagher, M. (1991) Proportionality, disproportionality and electoral systems. Electoral Studies 10(1):33-51.

Loosemore, J. and Hanby, V. (1971) The theoretical limits of maximum distortion: Som analytical expressions for electoral systems. \emph{British Journal of Political Science} 1:467-77.

Koppel, M., and A. Diskin. (2009) Measuring disproportionality, volatility and malapportionment: axiomatization and solutions. Social Choice and Welfare 33, no. 2: 281-286.

Rae, D. (1967) \emph{The Political Consequences of Electoral Laws.} London: Yale University Press.

Rose, Richard, Neil Munro and Tom Mackie (1998) \emph{ Elections in Central and Eastern Europe Since 1990.} Glasgow: Centre for the Study of Public Policy, University of Strathclyde.

Taagepera, R., and B. Grofman. Mapping the indices of seats-votes disproportionality and inter-election volatility. Party Politics 9, no. 6 (2003): 659-77.
}
\seealso{
\code{\link{politicalDiversity}}, \code{\link{largestRemainders}}, \code{\link{highestAverages}}. For more details, see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}
}

