% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoralTool.R
\docType{methods}
\name{Proportionality}
\alias{Disproportionality}
\alias{Proportionality}
\alias{Proportionality.default}
\title{Indexes of Proportionality}
\usage{
Proportionality(v, s, index = "Gallagher", ...)

\method{Proportionality}{default}(v, s, index = "Gallagher", ...)
}
\arguments{
\item{v}{A numeric vector with the percentage share of votes obtained by each party.}

\item{s}{A numeric vector with the percentage share of seats obtained by each party.}

\item{index}{The desired type of index, see details below.}

\item{\dots}{Additional arguements (currently ignored)}
}
\value{
A single score.
}
\description{
Calculates several indexes of proportionality to show the relationship of votes to seats.
}
\details{
The following measures are available:
\itemize{
\item {"Cox-Shugart"}{Cox-Shugart Measure of Proportionality}
\item {"inv.Cox-Shugart"}{The Inverse of Cox-Shugart Index}
\item {"Farina"}{Farina Index of Proportionality}
\item {"Gallagher"}{Gallagher Index of Disproportionality}
\item {"inv.Gallagher"}{The Inverse of Gallagher Index}
\item {"Farina"}{Farina Index of Proportionality}
\item {"Grofman"}{Grofman Index of Proportionality}
\item {"Lijphart"}{Lijphart Index of Proportionality}
\item {"Loosemore-Hanby"}{Loosemore-Hanby Index of Disproportionality}
\item {"Rae"}{Rae Index of Disproportionality}
\item {"inv.Rae"}{The Inverse of the Rae Index}
\item {"Rose"}{Rose Index of Disproportionality}
\item {"inv.Rose"}{The Inverse of the Rose Index}
}
}
\examples{
#' # 2012 Queensland state elecion
pvotes= c(49.65, 26.66, 11.5, 7.53, 3.16, 1.47)
pseats = c(87.64, 7.87, 2.25, 0.00, 2.25, 0.00)

# Proportionality(pvotes, pseats) # default is Gallagher

# Proportionality(pvotes, pseats, index="Rae")

# Proportionality(pvotes, pseats, index="Cox-Shugart")

# 2012 Quebec provincial election:
pvotes = c(PQ=31.95, Lib=31.20, CAQ=27.05, QS=6.03, Option=1.89, Other=1.88)
pseats = c(PQ=54, Lib=50, CAQ=19, QS=2, Option=0, Other=0)

# Proportionality(pvotes, pseats, index="Rae")

}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}
\references{
Duncan, O. and Duncan, B. (1955) A methodological analysis of segregation indexes. \emph{American Sociological Review} 20:210-7.

Gallagher, M. (1991) Proportionality, disproportionality and electoral systems. Electoral Studies 10(1):33-51.

Loosemore, J. and Hanby, V. (1971) The theoretical limits of maximum distortion: Som analytical expressions for electoral systems. \emph{British Journal of Political Science} 1:467-77.

Koppel, M., and A. Diskin. (2009) Measuring disproportionality, volatility and malapportionment: axiomatization and solutions. Social Choice and Welfare 33, no. 2: 281-286.

Rae, D. (1967) \emph{The Political Consequences of Electoral Laws.} London: Yale University Press.

Rose, Richard, Neil Munro and Tom Mackie (1998) \emph{ Elections in Central and Eastern Europe Since 1990.} Glasgow: Centre for the Study of Public Policy, University of Strathclyde.

Taagepera, R., and B. Grofman. Mapping the indices of seats-votes disproportionality and inter-election volatility. Party Politics 9, no. 6 (2003): 659-77.
}
\seealso{
\code{\link{PoliticalDiversity}}, \code{\link{LargestRemainders}}, \code{\link{HighestAverages}}. For more details, see the Indices vignette: \code{vignette("Indices", package = "SciencesPo")}
}

