% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{pub_color_pal}
\alias{pub_color_pal}
\title{Color Palettes for Publication (discrete)}
\usage{
pub_color_pal(palette = "tableau10")
}
\arguments{
\item{palette}{Palette name.}
}
\description{
Color palettes based on \href{http://www.tableausoftware.com/}{Tableau}.
}
\examples{
library(scales)
show_col(pub_color_pal("tableau20")(20))
}

