% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{pub_color_pal}
\alias{pub_color_pal}
\title{Color Palettes for Publication (discrete)}
\usage{
pub_color_pal(palette = "pub12", hex = TRUE)
}
\arguments{
\item{palette}{the palette name, a character string.}

\item{hex}{logical, if \code{TRUE}, the hex code is returned.}
}
\description{
Color palettes for publication-quality graphs. See details.
}
\details{
The following palettes are available:
\itemize{
\item {"pub12"}{A 12-color colorblind safe qualitative discrete palette.}
\item{"gray5"}{5-tons of gray.}
\item {"carnival"}{A 5-color palette inspired in the Brazilian samba schools.}
\item {"tableau20"}{Based on software \href{http://www.tableausoftware.com/}{Tableau}}
\item {"tableau10"}{Based on software \href{http://www.tableausoftware.com/}{Tableau}}
 \item {"tableau10light"}{Based on software \href{http://www.tableausoftware.com/}{Tableau}}
\item {"tableau10medium"}{Based on software \href{http://www.tableausoftware.com/}{Tableau}}
 \item {"fte"}{fivethirtyeight.com color scales}
}
}
\examples{

pub_color_pal("pub12")
pub_color_pal("gray5", hex=FALSE)
pub_color_pal("carnival")
pub_color_pal("gdocs")
pub_color_pal("tableau20")
pub_color_pal("tableau10")
pub_color_pal("tableau10medium")
pub_color_pal("tableau10light")
pub_color_pal("cyclic")
pub_color_pal("bivariate1")

}

