% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{pub_shape_pal}
\alias{pub_shape_pal}
\title{Shape scales for theme_pub (discrete)}
\usage{
pub_shape_pal(palette = "pub")
}
\arguments{
\item{palette}{the palette name, a character string.}
}
\description{
Discrete shape scales for \code{theme_pub()}.
}
\examples{
library(scales)

pub_shape_pal("default")(6)
pub_shape_pal("proportions")(4)
pub_shape_pal("gender")(2)

}
\seealso{
Other shape pub: \code{\link{scale_shape_pub}}
}

