% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{quotize}
\alias{quotize}
\title{Places quotation marks}
\usage{
quotize(vec)
}
\arguments{
\item{vec}{the vector whose values will be surounded by quotes}
}
\description{
Places quotation marks
}
\examples{
x <- 1
quotize(x)
noquote(quotize(x))
a <- ("Daniel")
noquote(quotize(a))
}

