% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{rdirichlet}
\alias{rdirichlet}
\title{Dirichlet distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{number of random observations to draw.}

\item{alpha}{the Dirichlet distribution's parameters. Can be a vector (one set of parameters for all observations) or a matrix (a different set of parameters for each observation), see \dQuote{Details}.

If \code{alpha} is a matrix, a complete set of \eqn{\alpha}{alpha}-parameters must be supplied for each observation.
\code{log} returns the logarithm of the densities (therefore the log-likelihood) and \code{sum.up} returns the product or sum and thereby the likelihood or log-likelihood.
Dirichlet (log-)densities are by default computed using C-routines (\code{ddirichlet_log_vector} and \code{ddirichlet_log_matrix}), a version only using R is provided by \code{ddirichlet_R}.
Caution: Although \code{.C()} can be used to call the C routines directly, R will crash or produce wrong values, if, e.g., data types are not set properly.}
}
\value{
rdirichlet returns a matrix with random numbers according to the supplied alpha vector or matrix.
ddirichlet returns a vector of densities (if \code{sum.up = FALSE}) or the (log-)likelihood (if \code{sum.up = TRUE}) for the given data and alphas.
}
\description{
Density function and random number generation for the Dirichlet distribution
}
\examples{
X1 <- rdirichlet(100, c(5, 5, 10))
a.mat <- cbind(1:10, 5, 10:1)
a.mat
X2 <- rdirichlet(10, a.mat)
# note how the probabilities in the first an last column relate to a.mat
}

