% Generated by roxygen2 (4.0.2): do not edit by hand
\name{recode}
\alias{recode}
\title{Recode variable}
\usage{
recode(vars, ...)
}
\arguments{
\item{vars}{a varaible or a set of variables.}
}
\description{
Change values of a variable in a \code{data.frame}.
}
\examples{
data(titanic)
  use(titanic)
  # recoding CLASS "crew" to missing value:
  recode(CLASS, from = crew, to = NA)
}

