% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\encoding{UTF-8}
\name{recode}
\alias{recode}
\alias{recode.default}
\title{Recode or Replace Values With New Values}
\usage{
recode(x, from, to, warn = TRUE)

\method{recode}{default}(x, from, to, warn = TRUE)
}
\arguments{
\item{x}{The vector whose values will be recoded.}

\item{from}{a vector of the items to recode.}

\item{to}{a vector of replacement values.}

\item{warn}{A logical to print a message if any of the
old values are not actually present in \code{x}.}
}
\description{
Recodes a value or a vector of values.
}
\examples{
x <- LETTERS[1:5]
recode(x, c("B", "D"), c("Beta", "Delta"))

# On numeric vectors
x <- c(1, 4, 5, 9)
recode(x, from = c(1, 4, 5, 9), to = c(10, 40, 50, 90))

}
\keyword{Manipulation}

