% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Method for recoding}
\usage{
recode(x, ...)
}
\arguments{
\item{x}{The data object.}

\item{\dots}{The parameters to be used in recode.}
}
\description{
Method for recoding
}

