% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{recode}
\alias{recode}
\title{Recode variable}
\usage{
recode(vars, old, new, data = .data, ...)
}
\arguments{
\item{vars}{varaible(s) to be recoded}

\item{old}{the old values or arithmetic conditions.}

\item{new}{the new values for all variables listed.}

\item{data}{a \code{data.frame} object.}

\item{\dots}{typically not needed parameters.}
}
\description{
Change values of a variable in a \code{data.frame}.
label.var
}
\examples{
df = data.frame(id=1:20, x=rnorm(20, mean=2, sd=.5),
z=sample(5, 20, rep=TRUE) )
use(df)
# recoding x  to missing value:
recode(z, old = c(1,2,3,4,5), new = c(5,4,3,2,1), data=df)
}

