% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manipulation.R
\encoding{UTF-8}
\name{rename}
\alias{rename}
\alias{rename.default}
\alias{rename.pattern}
\alias{rename.var}
\title{Rename a variable}
\usage{
rename(old, new, data, ...)

\method{rename}{default}(old, new, data, ...)

\method{rename}{var}(old, new, data, ...)

\method{rename}{pattern}(old, new, data, verbose = TRUE, ...)
}
\arguments{
\item{old}{a variable or a pattern among the names of the variables inside data.}

\item{new}{new name or new pattern of the variable(s).}

\item{data}{a data.frame of which the variables will be renamed.}

\item{verbose}{whether the old and new names of the variables(s) should be printed out.}

\item{\dots}{further arguments passed to or used by other methods.}
}
\description{
Rename a variable.
}
\examples{
data(ssex)
rename("Oppose", "NO", data = ssex)
}

