% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{rescale}
\alias{rescale}
\title{Rescaling as Z-scores and Centering}
\usage{
rescale(x, range)
}
\arguments{
\item{x}{A vector to be computed}

\item{range}{A valid range}
}
\value{
Scaled values in an object the same class as \code{x}.
}
\description{
Provides a rapidly way to generate standard variables.
}
\examples{
x <- sample(10)
rescale(x, range=0:1)
rescale(x, 0:1)
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\seealso{
\code{\link{scale}}, \code{\link{unscale}}
}
\keyword{Standardization}

