% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{rowSample}
\alias{rowSample}
\title{Row Sample}
\usage{
rowSample(x = .data, n)
}
\arguments{
\item{x}{the data object.}

\item{n}{the number of items to choose from.}
}
\description{
Sample rows of a data object.
}
\examples{
rowSample(iris, 20)
}
\author{
Daniel Marcelino \email{dmarcelino@live.com}
}

