% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Examples from the SciencesPo Package}
\usage{
runExample(example)
}
\arguments{
\item{example}{the name of the shiny application.}
}
\description{
Launch a Shiny app that shows a demo.
}
\examples{
# A demo of what \\code{\\link[SciencesPo]{PoliticalDiversity}} does.
if (interactive()) {
runExample(PoliticalDiversity)
}
}

