% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunShinyApp.R
\name{RunShinyApp}
\alias{RunShinyApp}
\title{Examples from the SciencesPo Package}
\usage{
RunShinyApp(app)
}
\arguments{
\item{app}{the name of the shiny application.}
}
\description{
Launch a Shiny app that shows a demo.
}
\examples{
# A demo of what \\code{\\link[SciencesPo]{PoliticalDiversity}} does.
if (interactive()) {
RunShinyApp("PoliticalDiversity")
}

# Other examples
if (interactive()) {
 RunShinyApp("ConjugateNormal")
}

if (interactive()) {
 RunShinyApp("Regression")
}
}

