% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/safeChars.R
\name{safeChars}
\alias{safeChars}
\title{Convert All Factor Columns to Character Columns}
\usage{
safeChars(mydf)
}
\arguments{
\item{mydf}{The name of your \code{data.frame}}
}
\description{
Sometimes, we forget to use the \code{stringsAsFactors} argument when using
\code{\link{read.table}} and related functions. By default, R converts
character columns to factors. Instead of re-reading the data, the
\code{\link{safeChars}} function will identify which columns are currently
factors, and convert them all to characters.
}
\author{
Ananda Mahto
}
\seealso{
\code{\link{read.table}}
}

