% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS.R
\encoding{UTF-8}
\name{Safechars}
\alias{Safechars}
\title{Convert All Factor Columns to Character Columns of a Data Frame}
\usage{
Safechars(.data)
}
\arguments{
\item{.data}{a \code{data.frame}.}
}
\description{
By default, R converts character columns to factors.
Instead of re-reading the data using \code{stringsAsFactors}, the
\code{\link{Safechars}} function will identify which columns are currently factors, and convert them all to characters.
}
\examples{
 str(iris)
iris_2 = Safechars(iris)
str(iris_2)

}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}.
}
\seealso{
\code{\link{read.table}}, \code{\link{Destring}}.
}
\keyword{internal}

