% Generated by roxygen2 (4.0.2): do not edit by hand
\encoding{UTF-8}
\name{safeRequire}
\alias{safeRequire}
\title{Safe require}
\usage{
safeRequire(packageName)
}
\arguments{
\item{packageName}{the name of a package.}
}
\description{
This function checks whether a package is installed; if not, it installs it. It then loads the package.
}

