% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample.size.R
\name{sample.size}
\alias{sample.size}
\title{# Public opinion samples in Brazil:
sample.size(p=.50, delta=.03)
sample.size(p=.50, delta=.02)}
\usage{
sample.size(p, delta = "auto", popsize = NULL, deff = 1, alpha = 0.05)
}
\description{
# Public opinion samples in Brazil:
sample.size(p=.50, delta=.03)
sample.size(p=.50, delta=.02)
}

