\encoding{UTF-8}
\name{SciencesPo-package}
\alias{SciencesPo}
\docType{package}
\title{Tools for analyzing political behaviour data	
}
\description{SciencesPo is a facility package for the political science crowd. It provides well-founded functions for analyzing random and nonrandom data. Future updates will focus on innovative techniques published in relevant journals in the field of political science.
}

\details{
\tabular{ll}{
Package: \tab SciencesPo\cr
Type: \tab Package\cr
Version: \tab 0.05.21 \cr
Date: \tab 2014-05-21\cr
License: \tab LGPL (>= 2.0, < 3)\cr }

}
\author{Daniel Marcelino <dmarcelino@live.com>
}
\references{
Marcelino, Daniel (2010). \emph{Sobre dinheiro e eleições: um estudo dos gastos de campanha para o Congresso Nacional em 2002 e 2006.}

Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}

