\encoding{UTF-8}
\name{SciencesPo-package}
\alias{SciencesPo}
\docType{package}
\title{A tool set for political behavior data	
}
\description{SciencesPo is a facility package that includes functions and algorithms most useful for the political science crowd. Use help(package="SciencesPo") for list all functions.
}

\details{
\tabular{ll}{
Package: \tab SciencesPo\cr
Type: \tab Package\cr
Version: \tab 0.10.21 \cr
Date: \tab 2014-10-21\cr
License: \tab LGPL (>= 2.0, < 3)\cr }

}
\author{Daniel Marcelino <dmarcelino@live.com>
}
\references{
Marcelino, Daniel (2010). \emph{Sobre dinheiro e eleições: um estudo dos gastos de campanha para o Congresso Nacional em 2002 e 2006.}

Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}

