% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SciencesPo.R
\docType{package}
\encoding{UTF-8}
\name{SciencesPo}
\alias{SciencesPo}
\alias{SciencesPo-package}
\title{\Sexpr[stage=build,results=rd]{tools:::Rd_package_title("SciencesPo")}}
\description{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_description("SciencesPo")}
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}
\references{
Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}

