% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\encoding{UTF-8}
\name{SciencesPo}
\alias{SciencesPo}
\alias{SciencesPo-package}
\title{A Tool Set For Analyzing Political Behavior Data}
\description{
SciencesPo is a facility package for the political science crowd. It provides a collection of functions for ease of analyze and presentation of random and nonrandom data. Use \code{help("SciencesPo")} for list all functions.
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}
\references{
Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}

