% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SciencesPo.R
\docType{package}
\encoding{UTF-8}
\name{SciencesPo}
\alias{SciencesPo}
\alias{SciencesPo-package}
\title{A Tool Set For Analyzing Political Behavior Data}
\description{
A Tool Set For Analyzing Political Behavior Data
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
# Export all functions that do NOT begin with a dot (period)
exportPattern("^[^\\.]")
}
\references{
Marcelino, Daniel (2013). \emph{SciencesPo: A Tool Set for Analyzing Political Behaviour Data}. Available at SSRN: \url{http://dx.doi.org/10.2139/ssrn.2320547}
}

