% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VISUALIZATION.R
\encoding{UTF-8}
\name{shadenorm}
\alias{shadenorm}
\title{Shades Normal Distribuion}
\usage{
shadenorm(below = NULL, above = NULL, pcts = c(0.025, 0.975), mu = 0,
  sigma = 1, numpts = 500, color = "gray", dens = 40,
  justabove = FALSE, justbelow = FALSE, lines = FALSE, between = NULL,
  outside = NULL)
}
\arguments{
\item{below}{sets a lower endpoint.}

\item{above}{sets an upper endpoint.}

\item{pcts}{the}

\item{mu}{the mean.}

\item{sigma}{standard deviations.}

\item{numpts}{the number os points/observations to drawn upon.}

\item{color}{the color of the area.}

\item{dens}{the density of the color.}

\item{justabove}{just plots the upper tail.}

\item{justbelow}{just plots the lower tail.}

\item{lines}{to draw lines.}

\item{between}{plots between specified points.}

\item{outside}{alternative "outside" area.}
}
\value{
A plot with a normal distribution density with shaded areas
}
\description{
Produces a plot of a normal density distribution with shaded areas.
}
\examples{
shadenorm()
shadenorm(below=-1.5)
shadenorm(below=-1.5,justbelow=TRUE)
shadenorm(above=1.5, justabove=TRUE)
shadenorm(below=-1.5,above=1.5)
shadenorm(between=c(-4,0),color="black")
shadenorm(between=c(0,4),color="black")
shadenorm(between=c(-1,+1),color="darkgray")
title("P[-1 < z < 1] = 68\%")
shadenorm(between=c(-2,+2),color="darkgray")
title("P[-2 < z < 2] = 95\%")
shadenorm(between=c(-3,+3),color="darkgray")
title("P[-3 < z < 3] = 99.7\%")
shadenorm(between = c(-1.75, 0, 2, 0.5, -1))  ## Plots between specified points
shadenorm(below=50,justbelow=TRUE,color="black",mu=47.3,sigma=9.3)

## Can plot one and then another on top of it using lines = TRUE
shadenorm(mu=2, sigma=10, outside=c(-3, 12), dens=15)
shadenorm(mu=2, sigma=15, between=c(-3, 12),lines=TRUE, col="blue",dens=15)
## Example: Plotting a Hypothesis Test for the mean
## Truth:      mu.true  = 8
## Hypothesis: mu.ho    = 6
## Generate Data Under Truth
mu.true = 5 ## Alternative Mean
mu.ho   = 6
sig     = 8
N       = 250 ## Sample Size

std.err = sig/sqrt(N)
crits = qnorm(c(0.025,0.975),mean=mu.ho, sd = std.err)
shadenorm(outside = crits, mu = mu.ho, sigma = std.err,dens=15)
shadenorm(between = crits, mu = mu.true, sigma = std.err, lines=TRUE, color="green",dens=15)
}

