% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{sheston91}
\alias{sheston91}
\title{The Penn World Table}
\format{A \code{data.frame} object with \Sexpr[stage=build,results=rd]{ncol(SciencesPo::sheston91)} variables and \Sexpr[stage=build,results=rd]{nrow(SciencesPo::sheston91)} observations.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University. \url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(sheston91)
}
\description{
The Penn World Table used in Summers and Heston (1991). This dataset contains the following columns:

\itemize{
\item year Year
\item pop Population (thousands)
\item rgdppc Real per capita GDP
\item savrat a numeric vector
\item country Country
\item com Communist regime
\item opec OPEC country
\item name Country name
}
}
\references{
Summers, R. and Heston, A. (1991) The Penn World Table (Mark 5): an expanded set of international comparisons, 1950--1988. \emph{The Quarterly Journal of Economics,} \bold{106(2),} 327--368.
}
\keyword{datasets}

