% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siegel.tukey.R
\name{siegel.tukey}
\alias{siegel.tukey}
\title{Siegel-Tukey Test for equality in variability}
\usage{
siegel.tukey(x, y, id.col = FALSE, adjust.median = F, rnd = -1,
  alternative = "two.sided", mu = 0, paired = FALSE, exact = FALSE,
  correct = TRUE, conf.int = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{A numeric vector of data values. Non-finite (e.g. infinite or missing) values will be omitted.}

\item{y}{A numeric vector of data values. Non-finite (e.g. infinite or missing) values will be omitted.}

\item{id.col}{}

\item{adjust.median}{Should between-group differences in medians be leveled before performing the test? In certain cases, the Siegel-Tukey test is susceptible to median differences and may indicate significant differences in variability that, in reality, stem from differences in medians.}

\item{rnd}{}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{mu}{A number specifying an optional parameter used to form the null hypothesis.}

\item{paired}{}

\item{exact}{A logical indicating whether an exact p-value should be computed. This is passed directly to wilcox.test.}

\item{correct}{A logical indicating whether to apply continuity correction in the normal approximation for the p-value.}

\item{conf.int}{A logical indicating whether a confidence interval should be computed.}

\item{conf.level}{The confidence level of the interval.}
}
\description{
Non-parametric Siegel-Tukey test for equality in variability.
 The null hypothesis is that the variability of x is equal between two groups.
 Thus, a rejection of the null hypothesis indicates that variability differs between the two groups.
 SiegelTukeyRank returns the ranks, calculated after Siegel Tukey logic.
}
\examples{
sample1 <- c(20.30, 22.53, 25.70, 13.23, 29.67, 24.46, 26.07, 19.35, 17.813,
            16.00, 15.30, 32.90)
sample2 <- c(10.56, 28.13, 19.94, 11.03, 8.093, 12.95, 21.14, 32.50, 10.90)
ks.test(sample1,sample2) # <-- Fail to reject H0 for alpha = .05
ks.test(sample1,sample2,alternative = "greater")
ks.test(sample1,sample2,alternative = "less")
}
\references{
Siegel, S., Tukey, J. W. (1960): A nonparametric sum of ranks procedure for relative spread in unpaired samples. Journal of the American Statistical Association.
}

