% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siegel.tukey.R
\name{siegel.tukey}
\alias{siegel.tukey}
\title{Siegel-Tukey Test for equality in variability}
\usage{
siegel.tukey(x, y, id.col = FALSE, adjust.median = F, rnd = -1,
  alternative = "two.sided", mu = 0, paired = FALSE, exact = FALSE,
  correct = TRUE, conf.int = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{A numeric vector of data values. Non-finite (e.g. infinite or missing) values will be omitted.}

\item{y}{A numeric vector of data values. Non-finite (e.g. infinite or missing) values will be omitted.}

\item{id.col}{If FALSE (default), then x and y are the data vectors (columns) for group 1 and 0, respectively. If TRUE, the y is the group indicator.}

\item{adjust.median}{Should between-group differences in medians be leveled before performing the test? In certain cases, the Siegel-Tukey test is susceptible to median differences and may indicate significant differences in variability that, in reality, stem from differences in medians.}

\item{rnd}{A non-negative integer if the data should be rounded to a given decimal. The default uses the data as is.}

\item{alternative}{A character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.}

\item{mu}{A number specifying an optional parameter used to form the null hypothesis.}

\item{paired}{}

\item{exact}{A logical indicating whether an exact p-value should be computed. This is passed directly to wilcox.test.}

\item{correct}{A logical indicating whether to apply continuity correction in the normal approximation for the p-value.}

\item{conf.int}{A logical indicating whether a confidence interval should be computed.}

\item{conf.level}{The confidence level of the interval.}
}
\description{
Non-parametric Siegel-Tukey test for equality in variability.
 The null hypothesis is that the variability of x is equal between two groups.
 Thus, a rejection of the null hypothesis indicates that variability differs between the two groups.
 SiegelTukeyRank returns the ranks, calculated after Siegel Tukey logic.
}
\examples{
x <- c(33, 62, 84, 85, 88, 93, 97, 4, 16, 48, 51, 66, 98)
id <- c(0,0,0,0,0,0,0,1,1,1,1,1,1)

 siegel.tukey(x,id, TRUE)

# Other:
x=c(4,4,5,5,6,6); y=c(0,0,1,9,10,10);

siegel.tukey(x,y)

# Yet another:
x=c(4,4,5,5,6,6);
y=c(0,0,1,9,10,10)

siegel.tukey(x,y)
# should be: 1, 4, 5, 8, 9, 12, 11, 10, 7, 6, 3, 2
# adjusted ranks would be: 2.5, 2.5, 5, 8.5, 8.5, 11.5, 11.5, 8.5, 8.5, 6, 2.5, 2.5
}
\references{
Siegel, S., Tukey, J. W. (1960): A nonparametric sum of ranks procedure for relative spread in unpaired samples. Journal of the American Statistical Association.
}

