% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SciencesPo.R
\docType{data}
\name{ssex}
\alias{ssex}
\title{Same Sex Marriage Public Opinion Data}
\format{A data frame with 18 observations on the following 4 variables.}
\usage{
data(ssex)
}
\description{
Data set fielded by the PEW Research Center on same sex marriage support in US. It covers public opinion on the issue starting from 1996 up to date.

\itemize{
  \item Date The year of the measurement
  \item Oppose Percent opposing same-sex marriage
  \item Favor Percent favoring same-sex marriage
  \item DK Percent of Don't Know
}
}
\references{
PEW Research Center. \emph{Support for same-sex marriage}. \url{http://www.pewresearch.org}
}
\keyword{datasets}

