% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstable.R
\encoding{UTF-8}
\name{summary.Crosstable}
\alias{summary.Crosstable}
\title{Print Crosstable style}
\usage{
\method{summary}{Crosstable}(object, digits = 2, chisq = FALSE,
  fisher = FALSE, mcnemar = FALSE, alt = "two.sided", latex = FALSE,
  ...)
}
\arguments{
\item{object}{The table object.}

\item{digits}{number of digits after the decimal point.}

\item{chisq}{if \code{TRUE}, the results of a chi-square test will be shown below the table.}

\item{fisher}{if \code{TRUE}, the results of a Fisher Exact test will be shown below the table.}

\item{mcnemar}{if \code{TRUE}, the results of a McNemar test will be shown below the table.}

\item{alt}{the alternative hypothesis and must be one of "two.sided",
"greater" or "less". Only used in the 2 by 2 case.}

\item{latex}{if the output is to be printed in latex format.}

\item{\dots}{extra parameters.}
}
\description{
Print Crosstable style
}
\seealso{
Other Crosstables: \code{\link{Crosstable}}
}

