% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crosstable.R
\encoding{UTF-8}
\name{summary.Crosstable}
\alias{summary.Crosstable}
\title{Print Crosstable style}
\usage{
\method{summary}{Crosstable}(object, digits = 2, chisq = FALSE,
  fisher = FALSE, mcnemar = FALSE, alt = "two.sided", latex = FALSE,
  ...)
}
\arguments{
\item{object}{The table object.}

\item{\dots}{extra parameters.}
}
\description{
Print Crosstable style
}

