% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svTransform.R
\encoding{UTF-8}
\name{svTransform}
\alias{svTransform}
\title{Transform dependent variable}
\usage{
svTransform(y)
}
\arguments{
\item{y}{the dependent variable in [0, 1] interval.}
}
\description{
Simple function to transform a dependent variable that in [0,1] rather than (0, 1) to beta regression. Suggested by Smithson & Verkuilen (2006).
}
\examples{
x <- sample(10);
 y <- normalize(x, range=c(0,1));
 y;
 svTransform(y)
}
\references{
Smithson M, Verkuilen J (2006) A Better Lemon Squeezer? Maximum-Likelihood Regres- sion with Beta-Distributed Dependent Variables. \emph{Psychological Methods}, 11(1), 54-71.
}
\seealso{
\code{\link{normalize}}.
}

