\name{swatson93}
\alias{swatson93}
\docType{data}
\title{ Stock's and Watson's (1993) Data.
}
\description{Data set used by Stock and Watson (1993) to estimate co-integration. 
}
\usage{data(swatson93)}
\format{
  A data frame with 90 observations on the following 5 variables.
  \describe{
    \item{\code{lnm1}}{Log M1}
    \item{\code{lnp}}{Log NNP price deflator}
    \item{\code{lnnnp}}{Log NNP}
    \item{\code{cprate}}{A numeric vector}
    \item{\code{year}}{Commercial paper rate}
  }
}

\source{ Hayashi, F. (2000) \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.

\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}

\references{Stock, J. H., and Watson, M. W. (1993) A simple estimator of cointegrating vectors in higher order integrated systems. \emph{Econometrica: Journal of the Econometric Society,} 783--820.

}
\examples{
data(swatson93)

}
\keyword{datasets}
