% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{swatson93}
\alias{swatson93}
\title{Stock's and Watson's (1993) Data.}
\format{A data frame with 90 observations on the following 5 variables.}
\source{
Hayashi, F. (2000). \emph{Econometrics.} Princeton. New Jersey, USA: Princeton University.\url{http://fmwww.bc.edu/ec-p/data/Hayashi/}
}
\usage{
data(swatson93)
}
\description{
Data set used by Stock and Watson (1993) to estimate co-integration.
\itemize{
\item lnm1 Log M1.
\item lnp Log NNP price deflator.
\item lnnnp Log NNP.
\item cprate A numeric vector.
\item year Commercial paper rate.
 }
}
\references{
Stock, J. H., and Watson, M. W. (1993) A simple estimator of cointegrating vectors in higher order integrated systems. \emph{Econometrica: Journal of the Econometric Society,} 783--820.
}
\keyword{datasets}

