% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{tab}
\alias{tab}
\title{Conditional replacement}
\usage{
tab(..., deparse.level = 2)
}
\arguments{
\item{.data}{The data object.

# replaceIf = function(.data,...,.if=NULL) {
#   .if = substitute(.if)
#   if (!is.null(.if)) {
#     rows = eval(.if,.data)
#     d = .data[rows,]
#     d = mutate(d,...)
#     .data[rows,] = d
#   } else {
#     .data = mutate(.data,...)
#   }
#   .data
# }
# Examples
# library(dplyr)
# dat = cars[1:10,]
# replace.if(dat, dist=dist*100, .if= speed==4)
# replace.if(dat, dist=dist*100)}
}
\description{
Conditional replacement
}

