% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TABLES.R
\encoding{UTF-8}
\name{tab}
\alias{tab}
\title{Stata-Like Two-Way Tabulation

\code{tabulate} is a modified version of \code{\link{crosstable}} for printing a summary table with cell counts and column proportions (similar to STATA's
\code{tabulate} \emph{varname1} \emph{varname2}, \code{col}).}
\usage{
tab(x, y, digits = 2)
}
\arguments{
\item{x,y}{The variables for the cross tabulation.}

\item{digits}{The number of digits for rounding proportions.}
}
\description{
`tab` <- function(.data, row, col){
 argx <- deparse(substitute(row))
 argy <- deparse(substitute(col))
 return(crosstabs(.data, argx, argy,
                  row.pct=FALSE,
                  col.pct = FALSE,
                  total.pct = FALSE,
                  chisq = FALSE))
} ### end -- tab function
NULL
}
\examples{
# Agresti (2002), table 3.10, p. 106
# 1992 General Social Survey- Race and Party affiliation
gss <- data.frame(
   expand.grid(Race=c("black", "white"),
   party=c("dem", "indep", "rep")),
   count=c(103,341,15,105,11,405))

df <- gss[rep(1:nrow(gss), gss[["count"]]), ]
with(df, tab(Race, party))
}
\seealso{
\code{\link[stats]{xtabs}}, \code{\link{crosstabs}},
\code{\link[base]{table}}, \code{\link[base]{prop.table}}
}

