% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_538}
\alias{theme_538}
\title{Themes for ggplot graphs}
\usage{
theme_538(font_size = 13, font_family = "", legend = "none")
}
\arguments{
\item{font_size}{Overall font size. Default is 13.}

\item{font_family}{Default font family.}

\item{legend}{Enables to set legend position, default is "none".}
}
\value{
The theme.
}
\description{
Theme for plotting  with ggplot.
}
\examples{
qplot(1:10, (1:10)^3) + theme_538()
}

