% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_black}
\alias{theme_black}
\title{Theme with no Black Background}
\usage{
theme_black(font_size = 14, font_family = "sans")
}
\arguments{
\item{font_size}{Overall font size. Default is 14.}

\item{font_family}{Default font family.}
}
\value{
The theme.
}
\description{
A ggplot2 theme with black background.
}
\examples{
ggplot(diamonds,aes(cut,group=1)) + geom_bar()+
geom_freqpoly(stat="count",size=2) + theme_black()
}
\seealso{
\code{\link[ggplot2]{theme}}, \code{\link{theme_pub}}, \code{\link{theme_blank}}.
}

