% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_blank}
\alias{theme_blank}
\title{Create a Completely Empty Theme}
\usage{
theme_blank(font_size = 12, font_family = "")
}
\arguments{
\item{font_size}{Overall font size. Default is 12.}

\item{font_family}{Default font family.}
}
\value{
The theme.
}
\description{
The theme created by this function shows nothing but the plot panel.
}
\examples{
# plot with small amount of remaining padding
qplot(1:10, (1:10)^2) + theme_blank()
# remaining padding removed
qplot(1:10, (1:10)^2) + theme_blank() + labs(x = NULL, y = NULL)
# Check that it is a complete theme
attr(theme_blank(), "complete")
}

