% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_darkside}
\alias{theme_darkside}
\title{The Dark Side Theme}
\usage{
theme_darkside(base_size = 12, base_family = "serif", legend = "none")
}
\arguments{
\item{base_size}{overall font size. Default is 12.}

\item{base_family}{the name for default font family.}

\item{legend}{the position of the legend if any.}
}
\value{
The theme.
# plot with small amount of padding
qplot(1:10, (1:10)^2, color="green") + theme_darkside()

# Check that it is a complete theme
attr(theme_darkside(), "complete")
}
\description{
The dark side of the things.
}
\author{
\Sexpr[stage=build,results=rd]{tools:::Rd_package_author("SciencesPo")}
}
\seealso{
Other themes darkside: \code{\link{theme_plus}}
}
\keyword{ggplot2}

