% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_fte}
\alias{theme_538}
\alias{theme_fte}
\title{Themes for ggplot2 Graphs}
\usage{
theme_fte(legend = "none", legend_title = FALSE, base_size = 12,
  horizontal = TRUE, base_family = "", colors = c("#F0F0F0", "#D9D9D9",
  "#60636A", "#525252"))
}
\arguments{
\item{legend}{enables to set legend position, default is "none".}

\item{legend_title}{will the legend have a title? Defaults is \code{FALSE}.}

\item{base_size}{overall font size. Default is 13.}

\item{horizontal}{logical. Horizontal axis lines?}

\item{base_family}{a nmae for default font family.}

\item{colors}{default colors used in the plot in the following order: background, lines, text, and title.}
}
\value{
The theme.
}
\description{
Theme for plotting  with ggplot2.
}
\examples{
qplot(1:10, (1:10)^3) + theme_fte()


mycolors = c("wheat",  "#C2AF8D",  "#8F6D2F", "darkred")
qplot(1:10, (1:10)^3) +
 theme_fte(colors=mycolors)

# Check that it is a complete theme
attr(theme_fte(), "complete")
}

