% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_pub}
\alias{labs_pub}
\alias{theme_pub}
\title{SciencesPo Publication Theme}
\usage{
theme_pub(base_size = 14, base_family = "")

labs_pub(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
\itemize{
 \item \strong{theme_pub()}: Create a publication ready theme
 \item \strong{labs_pub()}: Format only plot labels to a publication ready style
}
}
\examples{
p <- ggplot(mtcars, aes(x = wt, y = mpg)) +
   geom_point(aes(color = gear))
p

# Use theme_pub()
p + theme_pub()

# Format labels
p + labs_pub()

}

