% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_pub}
\alias{theme_pub}
\title{Create the default pub theme}
\usage{
theme_pub(font_size = 14, font_family = "", line_size = 0.5)
}
\arguments{
\item{font_size}{Overall font size. Default is 14.}

\item{font_family}{Default font family.}

\item{line_size}{Default line size.}
}
\value{
The theme.
}
\description{
After loading the SciencesPo package, this theme will be
set to default for all graphs made with ggplot2.
}
\examples{
qplot(1:10, (1:10)^2) + theme_pub(font_size = 15)
}

