% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_pub}
\alias{theme_pub}
\title{Themes for ggplot graphs}
\usage{
theme_pub(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{Enables to set the font size of all text elements.}

\item{base_family}{Enables to set the font family of all text elements.}
}
\description{
Theme for mapping  with ggplot
}

