% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_pub}
\alias{theme_pub}
\title{Create the default pub theme}
\usage{
theme_pub(font_family = "", font_size = 14, line_size = 0.5,
  legend.position = "bottom", axis.line.x = element_line(),
  axis.line.y = element_line())
}
\arguments{
\item{font_family}{Default font family.}

\item{font_size}{Overall font size. Default is 14.}

\item{line_size}{Default line size.}

\item{legend.position}{Enables to set legend position, default is "bottom".}

\item{axis.line.x}{Enables to set x axis line.}

\item{axis.line.y}{Enables to set y axis line.}
}
\value{
The theme.
}
\description{
After loading the SciencesPo package, this theme will be
set to default for all graphs made with ggplot2.
}
\examples{
ggplot(diamonds,aes(cut,group=1)) + geom_bar()+
geom_freqpoly(stat="count",size=2) + theme_pub(font_size = 16, line_size=1)
}

