% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{units}
\alias{units}
\title{Measurement System Units}
\format{A \code{data.frame} object with \Sexpr{ncol(SciencesPo::units)} variables and \Sexpr{nrow(SciencesPo::units)} observations.}
\usage{
data(units)
}
\description{
An internal dataset with measurement system units to be used by a converter function.

\itemize{
\item from A character defining the original unit.
\item to A character defining the target unit.
\item factor The factor to be applied in conversion.
\item description Some details about the measure.
}
}
\keyword{datasets}

