% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label.R
\name{units}
\alias{units}
\alias{units.default}
\alias{units<-.default}
\title{Append attribute to an object}
\usage{
units(x, ...)

\method{units}{default}(x) <- value

\method{units}{default}(x, none = "", ...)
}
\arguments{
\item{x}{A character string.}

\item{value}{the units of the object, or "".}

\item{none}{value to which to set result if no appropriate attribute is found.}

\item{\dots}{Ignored for parameters.}
}
\description{
Sets or retrieves the \code{"units"} attribute of an object.
}

