% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/label.R
\name{units}
\alias{units}
\alias{units.default}
\alias{units<-.default}
\title{Append attributes to label}
\usage{
units(x, ...)

\method{units}{default}(x) <- value

\method{units}{default}(x, none = "", ...)
}
\arguments{
\item{x}{A character string.}

\item{\dots}{Ignored for parameters.}
}
\description{
Append attributes to label
}

