% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/voronoi.R
\name{voronoi}
\alias{voronoi}
\title{Voronoi diagram}
\usage{
voronoi(p, n = 100, dim = 1000, plot = TRUE)
}
\arguments{
\item{p}{An integer for the size of the}

\item{n}{An integer for the size of}

\item{dim}{The dimension of the image.}

\item{plot}{Logical, if \code{TRUE}, the plot is returned, else, the \code{data.frame} is returned.}
}
\description{
Computes the voronoi diagram.
}
\details{
\url{https://en.wikipedia.org/wiki/Voronoi_diagram}
}
\examples{
\dontrun{ voronoi(p=2, n=20, dim=1000) }
}
\author{
Daniel Marcelino, \email{dmarcelino@live.com}
}

