% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/watch.R
\name{watch}
\alias{watch}
\title{Watch For Changes in a Data Object}
\usage{
watch(x)
}
\arguments{
\item{x}{a variable or a data object.}
}
\value{
a message.
}
\description{
Watch For Changes in a Data Object
}
\examples{
a <- 1
watch("a")
a <- 2
}

