% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{WeightedVariance}
\alias{WeightedVariance}
\title{Weighted Variance}
\usage{
WeightedVariance(x, weights = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{weights}{a numeric vector of weights for \code{x}.}

\item{na.rm}{a logical if NA should be disregarded.}
}
\description{
Compute the weighted variance.
}
\examples{
wt=c(1.23, 2.12, 1.23, 0.32, 1.53, 0.59, 0.94, 0.94, 0.84, 0.73)
x = c(5, 5, 4, 4, 3, 4, 3, 2, 2, 1)

WeightedVariance(x, wt)

}
\keyword{Exploratory}

