% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\encoding{UTF-8}
\name{words}
\alias{words}
\title{Word frequencies from Mosteller and Wallace}
\format{This data frame uses the word as row labels and contains the following columns}
\source{
Mosteller, F. and Wallace, D. (1964). Inference and Disputed Authorship: The Federalist. Reading, MA: Addison-Wesley.
}
\usage{
data(words)
}
\description{
The data give the frequencies of words in works from four different sources: the political writings of eighteenth century American political figures Alexander Hamilton, James Madison, and John Jay, and the book Ulysses by twentieth century Irish writer James Joyce.

\itemize{
\item Hamilton Hamilton frequency.
\item HamiltonRank Hamilton rank.
\item Madison Madison frequency.
\item MadisonRank Madison rank.
\item Jay Jay frequency.
\item JayRank Jay rank.
\item Ulysses Word frequency in Ulysses.
\item UlyssesRank Word rank in Ulysses.
 }
}
\references{
Weisberg, S. (2014). Applied Linear Regression, 4th edition. Hoboken NJ: Wiley.
}
\keyword{datasets}

