% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MANIPULATION.R
\encoding{UTF-8}
\name{zscores}
\alias{zscores}
\title{Compute z-scores}
\usage{
zscores(x, na.rm = getOption("na.rm", FALSE))
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\description{
Compute z-scores
}
\examples{
x <- sample(10)
zscores(x)
}

